% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_create_job}
\alias{elastictranscoder_create_job}
\title{When you create a job, Elastic Transcoder returns JSON data that
includes the values that you specified plus information about the job
that is created}
\usage{
elastictranscoder_create_job(PipelineId, Input, Inputs, Output, Outputs,
  OutputKeyPrefix, Playlists, UserMetadata)
}
\arguments{
\item{PipelineId}{[required] The \code{Id} of the pipeline that you want Elastic Transcoder to use for
transcoding. The pipeline determines several settings, including the
Amazon S3 bucket from which Elastic Transcoder gets the files to
transcode and the bucket into which Elastic Transcoder puts the
transcoded files.}

\item{Input}{A section of the request body that provides information about the file
that is being transcoded.}

\item{Inputs}{A section of the request body that provides information about the files
that are being transcoded.}

\item{Output}{A section of the request body that provides information about the
transcoded (target) file. We strongly recommend that you use the
\code{Outputs} syntax instead of the \code{Output} syntax.}

\item{Outputs}{A section of the request body that provides information about the
transcoded (target) files. We recommend that you use the \code{Outputs}
syntax instead of the \code{Output} syntax.}

\item{OutputKeyPrefix}{The value, if any, that you want Elastic Transcoder to prepend to the
names of all files that this job creates, including output files,
thumbnails, and playlists.}

\item{Playlists}{If you specify a preset in \code{PresetId} for which the value of \code{Container}
is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information
about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.}

\item{UserMetadata}{User-defined metadata that you want to associate with an Elastic
Transcoder job. You specify metadata in \code{key/value} pairs, and you can
add up to 10 \code{key/value} pairs per job. Elastic Transcoder does not
guarantee that \code{key/value} pairs are returned in the same order in which
you specify them.}
}
\description{
When you create a job, Elastic Transcoder returns JSON data that
includes the values that you specified plus information about the job
that is created.
}
\details{
If you have specified more than one output for your jobs (for example,
one output for the Kindle Fire and another output for the Apple iPhone
4s), you currently must use the Elastic Transcoder API to list the jobs
(as opposed to the AWS Console).
}
\section{Request syntax}{
\preformatted{svc$create_job(
  PipelineId = "string",
  Input = list(
    Key = "string",
    FrameRate = "string",
    Resolution = "string",
    AspectRatio = "string",
    Interlaced = "string",
    Container = "string",
    Encryption = list(
      Mode = "string",
      Key = "string",
      KeyMd5 = "string",
      InitializationVector = "string"
    ),
    TimeSpan = list(
      StartTime = "string",
      Duration = "string"
    ),
    InputCaptions = list(
      MergePolicy = "string",
      CaptionSources = list(
        list(
          Key = "string",
          Language = "string",
          TimeOffset = "string",
          Label = "string",
          Encryption = list(
            Mode = "string",
            Key = "string",
            KeyMd5 = "string",
            InitializationVector = "string"
          )
        )
      )
    ),
    DetectedProperties = list(
      Width = 123,
      Height = 123,
      FrameRate = "string",
      FileSize = 123,
      DurationMillis = 123
    )
  ),
  Inputs = list(
    list(
      Key = "string",
      FrameRate = "string",
      Resolution = "string",
      AspectRatio = "string",
      Interlaced = "string",
      Container = "string",
      Encryption = list(
        Mode = "string",
        Key = "string",
        KeyMd5 = "string",
        InitializationVector = "string"
      ),
      TimeSpan = list(
        StartTime = "string",
        Duration = "string"
      ),
      InputCaptions = list(
        MergePolicy = "string",
        CaptionSources = list(
          list(
            Key = "string",
            Language = "string",
            TimeOffset = "string",
            Label = "string",
            Encryption = list(
              Mode = "string",
              Key = "string",
              KeyMd5 = "string",
              InitializationVector = "string"
            )
          )
        )
      ),
      DetectedProperties = list(
        Width = 123,
        Height = 123,
        FrameRate = "string",
        FileSize = 123,
        DurationMillis = 123
      )
    )
  ),
  Output = list(
    Key = "string",
    ThumbnailPattern = "string",
    ThumbnailEncryption = list(
      Mode = "string",
      Key = "string",
      KeyMd5 = "string",
      InitializationVector = "string"
    ),
    Rotate = "string",
    PresetId = "string",
    SegmentDuration = "string",
    Watermarks = list(
      list(
        PresetWatermarkId = "string",
        InputKey = "string",
        Encryption = list(
          Mode = "string",
          Key = "string",
          KeyMd5 = "string",
          InitializationVector = "string"
        )
      )
    ),
    AlbumArt = list(
      MergePolicy = "string",
      Artwork = list(
        list(
          InputKey = "string",
          MaxWidth = "string",
          MaxHeight = "string",
          SizingPolicy = "string",
          PaddingPolicy = "string",
          AlbumArtFormat = "string",
          Encryption = list(
            Mode = "string",
            Key = "string",
            KeyMd5 = "string",
            InitializationVector = "string"
          )
        )
      )
    ),
    Composition = list(
      list(
        TimeSpan = list(
          StartTime = "string",
          Duration = "string"
        )
      )
    ),
    Captions = list(
      MergePolicy = "string",
      CaptionSources = list(
        list(
          Key = "string",
          Language = "string",
          TimeOffset = "string",
          Label = "string",
          Encryption = list(
            Mode = "string",
            Key = "string",
            KeyMd5 = "string",
            InitializationVector = "string"
          )
        )
      ),
      CaptionFormats = list(
        list(
          Format = "string",
          Pattern = "string",
          Encryption = list(
            Mode = "string",
            Key = "string",
            KeyMd5 = "string",
            InitializationVector = "string"
          )
        )
      )
    ),
    Encryption = list(
      Mode = "string",
      Key = "string",
      KeyMd5 = "string",
      InitializationVector = "string"
    )
  ),
  Outputs = list(
    list(
      Key = "string",
      ThumbnailPattern = "string",
      ThumbnailEncryption = list(
        Mode = "string",
        Key = "string",
        KeyMd5 = "string",
        InitializationVector = "string"
      ),
      Rotate = "string",
      PresetId = "string",
      SegmentDuration = "string",
      Watermarks = list(
        list(
          PresetWatermarkId = "string",
          InputKey = "string",
          Encryption = list(
            Mode = "string",
            Key = "string",
            KeyMd5 = "string",
            InitializationVector = "string"
          )
        )
      ),
      AlbumArt = list(
        MergePolicy = "string",
        Artwork = list(
          list(
            InputKey = "string",
            MaxWidth = "string",
            MaxHeight = "string",
            SizingPolicy = "string",
            PaddingPolicy = "string",
            AlbumArtFormat = "string",
            Encryption = list(
              Mode = "string",
              Key = "string",
              KeyMd5 = "string",
              InitializationVector = "string"
            )
          )
        )
      ),
      Composition = list(
        list(
          TimeSpan = list(
            StartTime = "string",
            Duration = "string"
          )
        )
      ),
      Captions = list(
        MergePolicy = "string",
        CaptionSources = list(
          list(
            Key = "string",
            Language = "string",
            TimeOffset = "string",
            Label = "string",
            Encryption = list(
              Mode = "string",
              Key = "string",
              KeyMd5 = "string",
              InitializationVector = "string"
            )
          )
        ),
        CaptionFormats = list(
          list(
            Format = "string",
            Pattern = "string",
            Encryption = list(
              Mode = "string",
              Key = "string",
              KeyMd5 = "string",
              InitializationVector = "string"
            )
          )
        )
      ),
      Encryption = list(
        Mode = "string",
        Key = "string",
        KeyMd5 = "string",
        InitializationVector = "string"
      )
    )
  ),
  OutputKeyPrefix = "string",
  Playlists = list(
    list(
      Name = "string",
      Format = "string",
      OutputKeys = list(
        "string"
      ),
      HlsContentProtection = list(
        Method = "string",
        Key = "string",
        KeyMd5 = "string",
        InitializationVector = "string",
        LicenseAcquisitionUrl = "string",
        KeyStoragePolicy = "string"
      ),
      PlayReadyDrm = list(
        Format = "string",
        Key = "string",
        KeyMd5 = "string",
        KeyId = "string",
        InitializationVector = "string",
        LicenseAcquisitionUrl = "string"
      )
    )
  ),
  UserMetadata = list(
    "string"
  )
)
}
}

\keyword{internal}
