% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_create_pipeline}
\alias{elastictranscoder_create_pipeline}
\title{The CreatePipeline operation creates a pipeline with settings that you
specify}
\usage{
elastictranscoder_create_pipeline(Name, InputBucket, OutputBucket, Role,
  AwsKmsKeyArn, Notifications, ContentConfig, ThumbnailConfig)
}
\arguments{
\item{Name}{[required] The name of the pipeline. We recommend that the name be unique within
the AWS account, but uniqueness is not enforced.

Constraints: Maximum 40 characters.}

\item{InputBucket}{[required] The Amazon S3 bucket in which you saved the media files that you want to
transcode.}

\item{OutputBucket}{The Amazon S3 bucket in which you want Elastic Transcoder to save the
transcoded files. (Use this, or use ContentConfig:Bucket plus
ThumbnailConfig:Bucket.)

Specify this value when all of the following are true:
\itemize{
\item You want to save transcoded files, thumbnails (if any), and
playlists (if any) together in one bucket.
\item You do not want to specify the users or groups who have access to
the transcoded files, thumbnails, and playlists.
\item You do not want to specify the permissions that Elastic Transcoder
grants to the files.

When Elastic Transcoder saves files in \code{OutputBucket}, it grants
full control over the files only to the AWS account that owns the
role that is specified by \code{Role}.
\item You want to associate the transcoded files and thumbnails with the
Amazon S3 Standard storage class.
}

If you want to save transcoded files and playlists in one bucket and
thumbnails in another bucket, specify which users can access the
transcoded files or the permissions the users have, or change the Amazon
S3 storage class, omit \code{OutputBucket} and specify values for
\code{ContentConfig} and \code{ThumbnailConfig} instead.}

\item{Role}{[required] The IAM Amazon Resource Name (ARN) for the role that you want Elastic
Transcoder to use to create the pipeline.}

\item{AwsKmsKeyArn}{The AWS Key Management Service (AWS KMS) key that you want to use with
this pipeline.

If you use either \code{s3} or \code{s3-aws-kms} as your \code{Encryption:Mode}, you
don't need to provide a key with your job because a default key, known
as an AWS-KMS key, is created for you automatically. You need to provide
an AWS-KMS key only if you want to use a non-default AWS-KMS key, or if
you are using an \code{Encryption:Mode} of \code{aes-cbc-pkcs7}, \code{aes-ctr}, or
\code{aes-gcm}.}

\item{Notifications}{The Amazon Simple Notification Service (Amazon SNS) topic that you want
to notify to report job status.

To receive notifications, you must also subscribe to the new topic in
the Amazon SNS console.
\itemize{
\item \strong{Progressing}: The topic ARN for the Amazon Simple Notification
Service (Amazon SNS) topic that you want to notify when Elastic
Transcoder has started to process a job in this pipeline. This is
the ARN that Amazon SNS returned when you created the topic. For
more information, see Create a Topic in the Amazon Simple
Notification Service Developer Guide.
\item \strong{Complete}: The topic ARN for the Amazon SNS topic that you want
to notify when Elastic Transcoder has finished processing a job in
this pipeline. This is the ARN that Amazon SNS returned when you
created the topic.
\item \strong{Warning}: The topic ARN for the Amazon SNS topic that you want to
notify when Elastic Transcoder encounters a warning condition while
processing a job in this pipeline. This is the ARN that Amazon SNS
returned when you created the topic.
\item \strong{Error}: The topic ARN for the Amazon SNS topic that you want to
notify when Elastic Transcoder encounters an error condition while
processing a job in this pipeline. This is the ARN that Amazon SNS
returned when you created the topic.
}}

\item{ContentConfig}{The optional \code{ContentConfig} object specifies information about the
Amazon S3 bucket in which you want Elastic Transcoder to save transcoded
files and playlists: which bucket to use, which users you want to have
access to the files, the type of access you want users to have, and the
storage class that you want to assign to the files.

If you specify values for \code{ContentConfig}, you must also specify values
for \code{ThumbnailConfig}.

If you specify values for \code{ContentConfig} and \code{ThumbnailConfig}, omit
the \code{OutputBucket} object.
\itemize{
\item \strong{Bucket}: The Amazon S3 bucket in which you want Elastic
Transcoder to save transcoded files and playlists.
\item \strong{Permissions} (Optional): The Permissions object specifies which
users you want to have access to transcoded files and the type of
access you want them to have. You can grant permissions to a maximum
of 30 users and/or predefined Amazon S3 groups.
\item \strong{Grantee Type}: Specify the type of value that appears in the
\code{Grantee} object:
\itemize{
\item \strong{Canonical}: The value in the \code{Grantee} object is either the
canonical user ID for an AWS account or an origin access
identity for an Amazon CloudFront distribution. For more
information about canonical user IDs, see Access Control List
(ACL) Overview in the Amazon Simple Storage Service Developer
Guide. For more information about using CloudFront origin access
identities to require that users use CloudFront URLs instead of
Amazon S3 URLs, see Using an Origin Access Identity to Restrict
Access to Your Amazon S3 Content.

A canonical user ID is not the same as an AWS account number.
\item \strong{Email}: The value in the \code{Grantee} object is the registered
email address of an AWS account.
\item \strong{Group}: The value in the \code{Grantee} object is one of the
following predefined Amazon S3 groups: \code{AllUsers},
\code{AuthenticatedUsers}, or \code{LogDelivery}.
}
\item \strong{Grantee}: The AWS user or group that you want to have access to
transcoded files and playlists. To identify the user or group, you
can specify the canonical user ID for an AWS account, an origin
access identity for a CloudFront distribution, the registered email
address of an AWS account, or a predefined Amazon S3 group
\item \strong{Access}: The permission that you want to give to the AWS user
that you specified in \code{Grantee}. Permissions are granted on the
files that Elastic Transcoder adds to the bucket, including
playlists and video files. Valid values include:
\itemize{
\item \code{READ}: The grantee can read the objects and metadata for
objects that Elastic Transcoder adds to the Amazon S3 bucket.
\item \code{READ_ACP}: The grantee can read the object ACL for objects that
Elastic Transcoder adds to the Amazon S3 bucket.
\item \code{WRITE_ACP}: The grantee can write the ACL for the objects that
Elastic Transcoder adds to the Amazon S3 bucket.
\item \code{FULL_CONTROL}: The grantee has \code{READ}, \code{READ_ACP}, and
\code{WRITE_ACP} permissions for the objects that Elastic Transcoder
adds to the Amazon S3 bucket.
}
\item \strong{StorageClass}: The Amazon S3 storage class, \code{Standard} or
\code{ReducedRedundancy}, that you want Elastic Transcoder to assign to
the video files and playlists that it stores in your Amazon S3
bucket.
}}

\item{ThumbnailConfig}{The \code{ThumbnailConfig} object specifies several values, including the
Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail
files, which users you want to have access to the files, the type of
access you want users to have, and the storage class that you want to
assign to the files.

If you specify values for \code{ContentConfig}, you must also specify values
for \code{ThumbnailConfig} even if you don't want to create thumbnails.

If you specify values for \code{ContentConfig} and \code{ThumbnailConfig}, omit
the \code{OutputBucket} object.
\itemize{
\item \strong{Bucket}: The Amazon S3 bucket in which you want Elastic
Transcoder to save thumbnail files.
\item \strong{Permissions} (Optional): The \code{Permissions} object specifies which
users and/or predefined Amazon S3 groups you want to have access to
thumbnail files, and the type of access you want them to have. You
can grant permissions to a maximum of 30 users and/or predefined
Amazon S3 groups.
\item \strong{GranteeType}: Specify the type of value that appears in the
Grantee object:
\itemize{
\item \strong{Canonical}: The value in the \code{Grantee} object is either the
canonical user ID for an AWS account or an origin access
identity for an Amazon CloudFront distribution.

A canonical user ID is not the same as an AWS account number.
\item \strong{Email}: The value in the \code{Grantee} object is the registered
email address of an AWS account.
\item \strong{Group}: The value in the \code{Grantee} object is one of the
following predefined Amazon S3 groups: \code{AllUsers},
\code{AuthenticatedUsers}, or \code{LogDelivery}.
}
\item \strong{Grantee}: The AWS user or group that you want to have access to
thumbnail files. To identify the user or group, you can specify the
canonical user ID for an AWS account, an origin access identity for
a CloudFront distribution, the registered email address of an AWS
account, or a predefined Amazon S3 group.
\item \strong{Access}: The permission that you want to give to the AWS user
that you specified in \code{Grantee}. Permissions are granted on the
thumbnail files that Elastic Transcoder adds to the bucket. Valid
values include:
\itemize{
\item \code{READ}: The grantee can read the thumbnails and metadata for
objects that Elastic Transcoder adds to the Amazon S3 bucket.
\item \code{READ_ACP}: The grantee can read the object ACL for thumbnails
that Elastic Transcoder adds to the Amazon S3 bucket.
\item \code{WRITE_ACP}: The grantee can write the ACL for the thumbnails
that Elastic Transcoder adds to the Amazon S3 bucket.
\item \code{FULL_CONTROL}: The grantee has \code{READ}, \code{READ_ACP}, and
\code{WRITE_ACP} permissions for the thumbnails that Elastic
Transcoder adds to the Amazon S3 bucket.
}
\item \strong{StorageClass}: The Amazon S3 storage class, \code{Standard} or
\code{ReducedRedundancy}, that you want Elastic Transcoder to assign to
the thumbnails that it stores in your Amazon S3 bucket.
}}
}
\value{
A list with the following syntax:\preformatted{list(
  Pipeline = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    Status = "string",
    InputBucket = "string",
    OutputBucket = "string",
    Role = "string",
    AwsKmsKeyArn = "string",
    Notifications = list(
      Progressing = "string",
      Completed = "string",
      Warning = "string",
      Error = "string"
    ),
    ContentConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    ),
    ThumbnailConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    )
  ),
  Warnings = list(
    list(
      Code = "string",
      Message = "string"
    )
  )
)
}
}
\description{
The CreatePipeline operation creates a pipeline with settings that you
specify.
}
\section{Request syntax}{
\preformatted{svc$create_pipeline(
  Name = "string",
  InputBucket = "string",
  OutputBucket = "string",
  Role = "string",
  AwsKmsKeyArn = "string",
  Notifications = list(
    Progressing = "string",
    Completed = "string",
    Warning = "string",
    Error = "string"
  ),
  ContentConfig = list(
    Bucket = "string",
    StorageClass = "string",
    Permissions = list(
      list(
        GranteeType = "string",
        Grantee = "string",
        Access = list(
          "string"
        )
      )
    )
  ),
  ThumbnailConfig = list(
    Bucket = "string",
    StorageClass = "string",
    Permissions = list(
      list(
        GranteeType = "string",
        Grantee = "string",
        Access = list(
          "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
