% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_delete_preset}
\alias{elastictranscoder_delete_preset}
\title{The DeletePreset operation removes a preset that you've added in an AWS
region}
\usage{
elastictranscoder_delete_preset(Id)
}
\arguments{
\item{Id}{[required] The identifier of the preset for which you want to get detailed
information.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
The DeletePreset operation removes a preset that you've added in an AWS
region.

You can't delete the default presets that are included with Elastic
Transcoder.
}
\section{Request syntax}{
\preformatted{svc$delete_preset(
  Id = "string"
)
}
}

\keyword{internal}
