% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_list_jobs_by_status}
\alias{elastictranscoder_list_jobs_by_status}
\title{The ListJobsByStatus operation gets a list of jobs that have a specified
status}
\usage{
elastictranscoder_list_jobs_by_status(Status, Ascending, PageToken)
}
\arguments{
\item{Status}{[required] To get information about all of the jobs associated with the current AWS
account that have a given status, specify the following status:
\code{Submitted}, \code{Progressing}, \code{Complete}, \code{Canceled}, or \code{Error}.}

\item{Ascending}{To list jobs in chronological order by the date and time that they were
submitted, enter \code{true}. To list jobs in reverse chronological order,
enter \code{false}.}

\item{PageToken}{When Elastic Transcoder returns more than one page of results, use
\code{pageToken} in subsequent \code{GET} requests to get each successive page of
results.}
}
\description{
The ListJobsByStatus operation gets a list of jobs that have a specified
status. The response body contains one element for each job that
satisfies the search criteria.
}
\section{Request syntax}{
\preformatted{svc$list_jobs_by_status(
  Status = "string",
  Ascending = "string",
  PageToken = "string"
)
}
}

\keyword{internal}
