% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_list_pipelines}
\alias{elastictranscoder_list_pipelines}
\title{The ListPipelines operation gets a list of the pipelines associated with
the current AWS account}
\usage{
elastictranscoder_list_pipelines(Ascending, PageToken)
}
\arguments{
\item{Ascending}{To list pipelines in chronological order by the date and time that they
were created, enter \code{true}. To list pipelines in reverse chronological
order, enter \code{false}.}

\item{PageToken}{When Elastic Transcoder returns more than one page of results, use
\code{pageToken} in subsequent \code{GET} requests to get each successive page of
results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Pipelines = list(
    list(
      Id = "string",
      Arn = "string",
      Name = "string",
      Status = "string",
      InputBucket = "string",
      OutputBucket = "string",
      Role = "string",
      AwsKmsKeyArn = "string",
      Notifications = list(
        Progressing = "string",
        Completed = "string",
        Warning = "string",
        Error = "string"
      ),
      ContentConfig = list(
        Bucket = "string",
        StorageClass = "string",
        Permissions = list(
          list(
            GranteeType = "string",
            Grantee = "string",
            Access = list(
              "string"
            )
          )
        )
      ),
      ThumbnailConfig = list(
        Bucket = "string",
        StorageClass = "string",
        Permissions = list(
          list(
            GranteeType = "string",
            Grantee = "string",
            Access = list(
              "string"
            )
          )
        )
      )
    )
  ),
  NextPageToken = "string"
)
}
}
\description{
The ListPipelines operation gets a list of the pipelines associated with
the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_pipelines(
  Ascending = "string",
  PageToken = "string"
)
}
}

\keyword{internal}
