% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_list_presets}
\alias{elastictranscoder_list_presets}
\title{The ListPresets operation gets a list of the default presets included
with Elastic Transcoder and the presets that you've added in an AWS
region}
\usage{
elastictranscoder_list_presets(Ascending, PageToken)
}
\arguments{
\item{Ascending}{To list presets in chronological order by the date and time that they
were created, enter \code{true}. To list presets in reverse chronological
order, enter \code{false}.}

\item{PageToken}{When Elastic Transcoder returns more than one page of results, use
\code{pageToken} in subsequent \code{GET} requests to get each successive page of
results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Presets = list(
    list(
      Id = "string",
      Arn = "string",
      Name = "string",
      Description = "string",
      Container = "string",
      Audio = list(
        Codec = "string",
        SampleRate = "string",
        BitRate = "string",
        Channels = "string",
        AudioPackingMode = "string",
        CodecOptions = list(
          Profile = "string",
          BitDepth = "string",
          BitOrder = "string",
          Signed = "string"
        )
      ),
      Video = list(
        Codec = "string",
        CodecOptions = list(
          "string"
        ),
        KeyframesMaxDist = "string",
        FixedGOP = "string",
        BitRate = "string",
        FrameRate = "string",
        MaxFrameRate = "string",
        Resolution = "string",
        AspectRatio = "string",
        MaxWidth = "string",
        MaxHeight = "string",
        DisplayAspectRatio = "string",
        SizingPolicy = "string",
        PaddingPolicy = "string",
        Watermarks = list(
          list(
            Id = "string",
            MaxWidth = "string",
            MaxHeight = "string",
            SizingPolicy = "string",
            HorizontalAlign = "string",
            HorizontalOffset = "string",
            VerticalAlign = "string",
            VerticalOffset = "string",
            Opacity = "string",
            Target = "string"
          )
        )
      ),
      Thumbnails = list(
        Format = "string",
        Interval = "string",
        Resolution = "string",
        AspectRatio = "string",
        MaxWidth = "string",
        MaxHeight = "string",
        SizingPolicy = "string",
        PaddingPolicy = "string"
      ),
      Type = "string"
    )
  ),
  NextPageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
The ListPresets operation gets a list of the default presets included
with Elastic Transcoder and the presets that you've added in an AWS
region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_presets(
  Ascending = "string",
  PageToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
