% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_read_pipeline}
\alias{elastictranscoder_read_pipeline}
\title{The ReadPipeline operation gets detailed information about a pipeline}
\usage{
elastictranscoder_read_pipeline(Id)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline to read.}
}
\value{
A list with the following syntax:\preformatted{list(
  Pipeline = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    Status = "string",
    InputBucket = "string",
    OutputBucket = "string",
    Role = "string",
    AwsKmsKeyArn = "string",
    Notifications = list(
      Progressing = "string",
      Completed = "string",
      Warning = "string",
      Error = "string"
    ),
    ContentConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    ),
    ThumbnailConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    )
  ),
  Warnings = list(
    list(
      Code = "string",
      Message = "string"
    )
  )
)
}
}
\description{
The ReadPipeline operation gets detailed information about a pipeline.
}
\section{Request syntax}{
\preformatted{svc$read_pipeline(
  Id = "string"
)
}
}

\keyword{internal}
