% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_test_role}
\alias{elastictranscoder_test_role}
\title{The TestRole operation tests the IAM role used to create the pipeline}
\usage{
elastictranscoder_test_role(Role, InputBucket, OutputBucket, Topics)
}
\arguments{
\item{Role}{[required] The IAM Amazon Resource Name (ARN) for the role that you want Elastic
Transcoder to test.}

\item{InputBucket}{[required] The Amazon S3 bucket that contains media files to be transcoded. The
action attempts to read from this bucket.}

\item{OutputBucket}{[required] The Amazon S3 bucket that Elastic Transcoder writes transcoded media
files to. The action attempts to read from this bucket.}

\item{Topics}{[required] The ARNs of one or more Amazon Simple Notification Service (Amazon SNS)
topics that you want the action to send a test notification to.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Success = "string",
  Messages = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
The TestRole operation tests the IAM role used to create the pipeline.

The \code{\link[=elastictranscoder_test_role]{test_role}} action lets you determine
whether the IAM role you are using has sufficient permissions to let
Elastic Transcoder perform tasks associated with the transcoding
process. The action attempts to assume the specified IAM role, checks
read access to the input and output buckets, and tries to send a test
notification to Amazon SNS topics that you specify.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$test_role(
  Role = "string",
  InputBucket = "string",
  OutputBucket = "string",
  Topics = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
