% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_update_pipeline_notifications}
\alias{elastictranscoder_update_pipeline_notifications}
\title{With the UpdatePipelineNotifications operation, you can update Amazon
Simple Notification Service (Amazon SNS) notifications for a pipeline}
\usage{
elastictranscoder_update_pipeline_notifications(Id, Notifications)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline for which you want to change notification
settings.}

\item{Notifications}{[required] The topic ARN for the Amazon Simple Notification Service (Amazon SNS)
topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in
the Amazon SNS console.
\itemize{
\item \strong{Progressing}: The topic ARN for the Amazon Simple Notification
Service (Amazon SNS) topic that you want to notify when Elastic
Transcoder has started to process jobs that are added to this
pipeline. This is the ARN that Amazon SNS returned when you created
the topic.
\item \strong{Complete}: The topic ARN for the Amazon SNS topic that you want
to notify when Elastic Transcoder has finished processing a job.
This is the ARN that Amazon SNS returned when you created the topic.
\item \strong{Warning}: The topic ARN for the Amazon SNS topic that you want to
notify when Elastic Transcoder encounters a warning condition. This
is the ARN that Amazon SNS returned when you created the topic.
\item \strong{Error}: The topic ARN for the Amazon SNS topic that you want to
notify when Elastic Transcoder encounters an error condition. This
is the ARN that Amazon SNS returned when you created the topic.
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Pipeline = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    Status = "string",
    InputBucket = "string",
    OutputBucket = "string",
    Role = "string",
    AwsKmsKeyArn = "string",
    Notifications = list(
      Progressing = "string",
      Completed = "string",
      Warning = "string",
      Error = "string"
    ),
    ContentConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    ),
    ThumbnailConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
With the UpdatePipelineNotifications operation, you can update Amazon
Simple Notification Service (Amazon SNS) notifications for a pipeline.

When you update notifications for a pipeline, Elastic Transcoder returns
the values that you specified in the request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_pipeline_notifications(
  Id = "string",
  Notifications = list(
    Progressing = "string",
    Completed = "string",
    Warning = "string",
    Error = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
