% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_update_pipeline_status}
\alias{elastictranscoder_update_pipeline_status}
\title{The UpdatePipelineStatus operation pauses or reactivates a pipeline, so
that the pipeline stops or restarts the processing of jobs}
\usage{
elastictranscoder_update_pipeline_status(Id, Status)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline to update.}

\item{Status}{[required] The desired status of the pipeline:
\itemize{
\item \code{Active}: The pipeline is processing jobs.
\item \code{Paused}: The pipeline is not currently processing jobs.
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Pipeline = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    Status = "string",
    InputBucket = "string",
    OutputBucket = "string",
    Role = "string",
    AwsKmsKeyArn = "string",
    Notifications = list(
      Progressing = "string",
      Completed = "string",
      Warning = "string",
      Error = "string"
    ),
    ContentConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    ),
    ThumbnailConfig = list(
      Bucket = "string",
      StorageClass = "string",
      Permissions = list(
        list(
          GranteeType = "string",
          Grantee = "string",
          Access = list(
            "string"
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
The UpdatePipelineStatus operation pauses or reactivates a pipeline, so
that the pipeline stops or restarts the processing of jobs.

Changing the pipeline status is useful if you want to cancel one or more
jobs. You can't cancel jobs after Elastic Transcoder has started
processing them; if you pause the pipeline to which you submitted the
jobs, you have more time to get the job IDs for the jobs that you want
to cancel, and to send a \code{\link[=elastictranscoder_cancel_job]{cancel_job}}
request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_pipeline_status(
  Id = "string",
  Status = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
