% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_create_signaling_channel}
\alias{kinesisvideo_create_signaling_channel}
\title{Creates a signaling channel}
\usage{
kinesisvideo_create_signaling_channel(ChannelName, ChannelType,
  SingleMasterConfiguration, Tags)
}
\arguments{
\item{ChannelName}{[required] A name for the signaling channel that you are creating. It must be
unique for each AWS account and AWS Region.}

\item{ChannelType}{A type of the signaling channel that you are creating. Currently,
\code{SINGLE_MASTER} is the only supported channel type.}

\item{SingleMasterConfiguration}{A structure containing the configuration for the \code{SINGLE_MASTER} channel
type.}

\item{Tags}{A set of tags (key-value pairs) that you want to associate with this
channel.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelARN = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a signaling channel.

\code{\link[=kinesisvideo_create_signaling_channel]{create_signaling_channel}} is
an asynchronous operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_signaling_channel(
  ChannelName = "string",
  ChannelType = "SINGLE_MASTER",
  SingleMasterConfiguration = list(
    MessageTtlSeconds = 123
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
