% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_create_stream}
\alias{kinesisvideo_create_stream}
\title{Creates a new Kinesis video stream}
\usage{
kinesisvideo_create_stream(DeviceName, StreamName, MediaType, KmsKeyId,
  DataRetentionInHours, Tags)
}
\arguments{
\item{DeviceName}{The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this
name.}

\item{StreamName}{[required] A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for
each account and region.}

\item{MediaType}{The media type of the stream. Consumers of the stream can use this
information when processing the stream. For more information about media
types, see \href{http://www.iana.org/assignments/media-types/media-types.xhtml}{Media Types}.
If you choose to specify the \code{MediaType}, see Naming Requirements for
guidelines.

Example valid values include \"video/h264\" and
\"video/h264,audio/aac\".

This parameter is optional; the default value is \code{null} (or empty in
JSON).}

\item{KmsKeyId}{The ID of the AWS Key Management Service (AWS KMS) key that you want
Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key
(\code{aws/kinesisvideo}) is used.

For more information, see
\href{https://docs.aws.amazon.com/kms/latest/APIReference/API_DescribeKey.html#API_DescribeKey_RequestParameters}{DescribeKey}.}

\item{DataRetentionInHours}{The number of hours that you want to retain the data in the stream.
Kinesis Video Streams retains the data in a data store that is
associated with the stream.

The default value is 0, indicating that the stream does not persist
data.

When the \code{DataRetentionInHours} value is 0, consumers can still consume
the fragments that remain in the service host buffer, which has a
retention time limit of 5 minutes and a retention memory limit of 200
MB. Fragments are removed from the buffer when either limit is reached.}

\item{Tags}{A list of tags to associate with the specified stream. Each tag is a
key-value pair (the value is optional).}
}
\description{
Creates a new Kinesis video stream.
}
\details{
When you create a new stream, Kinesis Video Streams assigns it a version
number. When you change the stream\'s metadata, Kinesis Video Streams
updates the version.

\code{CreateStream} is an asynchronous operation.

For information about how the service works, see \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/how-it-works.html}{How it Works}.

You must have permissions for the \code{KinesisVideo:CreateStream} action.
}
\section{Request syntax}{
\preformatted{svc$create_stream(
  DeviceName = "string",
  StreamName = "string",
  MediaType = "string",
  KmsKeyId = "string",
  DataRetentionInHours = 123,
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
