% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_delete_signaling_channel}
\alias{kinesisvideo_delete_signaling_channel}
\title{Deletes a specified signaling channel}
\usage{
kinesisvideo_delete_signaling_channel(ChannelARN, CurrentVersion)
}
\arguments{
\item{ChannelARN}{[required] The Amazon Resource Name (ARN) of the signaling channel that you want to
delete.}

\item{CurrentVersion}{The current version of the signaling channel that you want to delete.
You can obtain the current version by invoking the
\code{\link[=kinesisvideo_describe_signaling_channel]{describe_signaling_channel}}
or \code{\link[=kinesisvideo_list_signaling_channels]{list_signaling_channels}} API
operations.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a specified signaling channel.
\code{\link[=kinesisvideo_delete_signaling_channel]{delete_signaling_channel}} is
an asynchronous operation. If you don't specify the channel's current
version, the most recent version is deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_signaling_channel(
  ChannelARN = "string",
  CurrentVersion = "string"
)
}
}

\keyword{internal}
