% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_describe_signaling_channel}
\alias{kinesisvideo_describe_signaling_channel}
\title{Returns the most current information about the signaling channel}
\usage{
kinesisvideo_describe_signaling_channel(ChannelName, ChannelARN)
}
\arguments{
\item{ChannelName}{The name of the signaling channel that you want to describe.}

\item{ChannelARN}{The ARN of the signaling channel that you want to describe.}
}
\description{
Returns the most current information about the signaling channel. You
must specify either the name or the ARN of the channel that you want to
describe.
}
\section{Request syntax}{
\preformatted{svc$describe_signaling_channel(
  ChannelName = "string",
  ChannelARN = "string"
)
}
}

\keyword{internal}
