% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_describe_stream}
\alias{kinesisvideo_describe_stream}
\title{Returns the most current information about the specified stream}
\usage{
kinesisvideo_describe_stream(StreamName, StreamARN)
}
\arguments{
\item{StreamName}{The name of the stream.}

\item{StreamARN}{The Amazon Resource Name (ARN) of the stream.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  StreamInfo = list(
    DeviceName = "string",
    StreamName = "string",
    StreamARN = "string",
    MediaType = "string",
    KmsKeyId = "string",
    Version = "string",
    Status = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    DataRetentionInHours = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the most current information about the specified stream. You
must specify either the \code{StreamName} or the \code{StreamARN}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_stream(
  StreamName = "string",
  StreamARN = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
