% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_get_data_endpoint}
\alias{kinesisvideo_get_data_endpoint}
\title{Gets an endpoint for a specified stream for either reading or writing}
\usage{
kinesisvideo_get_data_endpoint(StreamName, StreamARN, APIName)
}
\arguments{
\item{StreamName}{The name of the stream that you want to get the endpoint for. You must
specify either this parameter or a \code{StreamARN} in the request.}

\item{StreamARN}{The Amazon Resource Name (ARN) of the stream that you want to get the
endpoint for. You must specify either this parameter or a \code{StreamName}
in the request.}

\item{APIName}{[required] The name of the API action for which to get an endpoint.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DataEndpoint = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets an endpoint for a specified stream for either reading or writing.
Use this endpoint in your application to read from the specified stream
(using the \code{GetMedia} or \code{GetMediaForFragmentList} operations) or write
to it (using the \code{PutMedia} operation).

The returned endpoint does not have the API name appended. The client
needs to add the API name to the returned endpoint.

In the request, specify the stream either by \code{StreamName} or
\code{StreamARN}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_data_endpoint(
  StreamName = "string",
  StreamARN = "string",
  APIName = "PUT_MEDIA"|"GET_MEDIA"|"LIST_FRAGMENTS"|"GET_MEDIA_FOR_FRAGMENT_LIST"|"GET_HLS_STREAMING_SESSION_URL"|"GET_DASH_STREAMING_SESSION_URL"|"GET_CLIP"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
