% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_list_streams}
\alias{kinesisvideo_list_streams}
\title{Returns an array of StreamInfo objects}
\usage{
kinesisvideo_list_streams(MaxResults, NextToken, StreamNameCondition)
}
\arguments{
\item{MaxResults}{The maximum number of streams to return in the response. The default is
10,000.}

\item{NextToken}{If you specify this parameter, when the result of a
\code{\link[=kinesisvideo_list_streams]{list_streams}} operation is truncated, the
call returns the \code{NextToken} in the response. To get another batch of
streams, provide this token in your next request.}

\item{StreamNameCondition}{Optional: Returns only streams that satisfy a specific condition.
Currently, you can specify only the prefix of a stream name as a
condition.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  StreamInfoList = list(
    list(
      DeviceName = "string",
      StreamName = "string",
      StreamARN = "string",
      MediaType = "string",
      KmsKeyId = "string",
      Version = "string",
      Status = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      DataRetentionInHours = 123
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns an array of \code{StreamInfo} objects. Each object describes a
stream. To retrieve only streams that satisfy a specific condition, you
can specify a \code{StreamNameCondition}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_streams(
  MaxResults = 123,
  NextToken = "string",
  StreamNameCondition = list(
    ComparisonOperator = "BEGINS_WITH",
    ComparisonValue = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
