% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_list_tags_for_stream}
\alias{kinesisvideo_list_tags_for_stream}
\title{Returns a list of tags associated with the specified stream}
\usage{
kinesisvideo_list_tags_for_stream(NextToken, StreamARN, StreamName)
}
\arguments{
\item{NextToken}{If you specify this parameter and the result of a
\code{\link[=kinesisvideo_list_tags_for_stream]{list_tags_for_stream}} call is
truncated, the response includes a token that you can use in the next
request to fetch the next batch of tags.}

\item{StreamARN}{The Amazon Resource Name (ARN) of the stream that you want to list tags
for.}

\item{StreamName}{The name of the stream that you want to list tags for.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Returns a list of tags associated with the specified stream.

In the request, you must specify either the \code{StreamName} or the
\code{StreamARN}.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_stream(
  NextToken = "string",
  StreamARN = "string",
  StreamName = "string"
)
}
}

\keyword{internal}
