% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_tag_resource}
\alias{kinesisvideo_tag_resource}
\title{Adds one or more tags to a signaling channel}
\usage{
kinesisvideo_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the signaling channel to which you want to add tags.}

\item{Tags}{[required] A list of tags to associate with the specified signaling channel. Each
tag is a key-value pair.}
}
\description{
Adds one or more tags to a signaling channel. A \emph{tag} is a key-value
pair (the value is optional) that you can define and assign to AWS
resources. If you specify a tag that already exists, the tag value is
replaced with the value that you specify in the request. For more
information, see \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html}{Using Cost Allocation Tags}
in the \emph{AWS Billing and Cost Management User Guide}.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
