% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_update_signaling_channel}
\alias{kinesisvideo_update_signaling_channel}
\title{Updates the existing signaling channel}
\usage{
kinesisvideo_update_signaling_channel(ChannelARN, CurrentVersion,
  SingleMasterConfiguration)
}
\arguments{
\item{ChannelARN}{[required] The Amazon Resource Name (ARN) of the signaling channel that you want to
update.}

\item{CurrentVersion}{[required] The current version of the signaling channel that you want to update.}

\item{SingleMasterConfiguration}{The structure containing the configuration for the \code{SINGLE_MASTER} type
of the signaling channel that you want to update.}
}
\value{
An empty list.
}
\description{
Updates the existing signaling channel. This is an asynchronous
operation and takes time to complete.

If the \code{MessageTtlSeconds} value is updated (either increased or
reduced), it only applies to new messages sent via this channel after
it's been updated. Existing messages are still expired as per the
previous \code{MessageTtlSeconds} value.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_signaling_channel(
  ChannelARN = "string",
  CurrentVersion = "string",
  SingleMasterConfiguration = list(
    MessageTtlSeconds = 123
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
