% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideoarchivedmedia_operations.R
\name{kinesisvideoarchivedmedia_get_clip}
\alias{kinesisvideoarchivedmedia_get_clip}
\title{Downloads an MP4 file (clip) containing the archived, on-demand media
from the specified video stream over the specified time range}
\usage{
kinesisvideoarchivedmedia_get_clip(StreamName, StreamARN,
  ClipFragmentSelector)
}
\arguments{
\item{StreamName}{The name of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.}

\item{StreamARN}{The Amazon Resource Name (ARN) of the stream for which to retrieve the
media clip.

You must specify either the StreamName or the StreamARN.}

\item{ClipFragmentSelector}{[required] The time range of the requested clip and the source of the timestamps.}
}
\value{
A list with the following syntax:\preformatted{list(
  ContentType = "string",
  Payload = raw
)
}
}
\description{
Downloads an MP4 file (clip) containing the archived, on-demand media
from the specified video stream over the specified time range.

Both the StreamName and the StreamARN parameters are optional, but you
must specify either the StreamName or the StreamARN when invoking this
API operation.

As a prerequisite to using GetCLip API, you must obtain an endpoint
using \code{GetDataEndpoint}, specifying GET_CLIP for`` the \code{APIName}
parameter.

An Amazon Kinesis video stream has the following requirements for
providing data through MP4:
\itemize{
\item The media must contain h.264 or h.265 encoded video and, optionally,
AAC or G.711 encoded audio. Specifically, the codec ID of track 1
should be \code{V_MPEG/ISO/AVC} (for h.264) or V_MPEGH/ISO/HEVC (for
H.265). Optionally, the codec ID of track 2 should be \code{A_AAC} (for
AAC) or A_MS/ACM (for G.711).
\item Data retention must be greater than 0.
\item The video track of each fragment must contain codec private data in
the Advanced Video Coding (AVC) for H.264 format and HEVC for H.265
format. For more information, see MPEG-4 specification ISO/IEC
14496-15. For information about adapting stream data to a given
format, see \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/producer-reference-nal.html}{NAL Adaptation Flags}.
\item The audio track (if present) of each fragment must contain codec
private data in the AAC format (\href{https://www.iso.org/standard/43345.html}{AAC specification ISO/IEC 13818-7}) or the \href{http://www-mmsp.ece.mcgill.ca/Documents/AudioFormats/WAVE/WAVE.html}{MS Wave format}.
}

You can monitor the amount of outgoing data by monitoring the
\code{GetClip.OutgoingBytes} Amazon CloudWatch metric. For information about
using CloudWatch to monitor Kinesis Video Streams, see \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/monitoring.html}{Monitoring Kinesis Video Streams}.
For pricing information, see \href{https://aws.amazon.com/kinesis/video-streams/pricing/}{Amazon Kinesis Video Streams Pricing} and \href{https://aws.amazon.com/pricing/}{AWS Pricing}. Charges for outgoing AWS data
apply.
}
\section{Request syntax}{
\preformatted{svc$get_clip(
  StreamName = "string",
  StreamARN = "string",
  ClipFragmentSelector = list(
    FragmentSelectorType = "PRODUCER_TIMESTAMP"|"SERVER_TIMESTAMP",
    TimestampRange = list(
      StartTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      EndTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}

\keyword{internal}
