% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideoarchivedmedia_operations.R
\name{kinesisvideoarchivedmedia_get_media_for_fragment_list}
\alias{kinesisvideoarchivedmedia_get_media_for_fragment_list}
\title{Gets media for a list of fragments (specified by fragment number) from
the archived data in an Amazon Kinesis video stream}
\usage{
kinesisvideoarchivedmedia_get_media_for_fragment_list(StreamName,
  Fragments)
}
\arguments{
\item{StreamName}{[required] The name of the stream from which to retrieve fragment media.}

\item{Fragments}{[required] A list of the numbers of fragments for which to retrieve media. You
retrieve these values with
\code{\link[=kinesisvideoarchivedmedia_list_fragments]{list_fragments}}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ContentType = "string",
  Payload = raw
)
}
}
\description{
Gets media for a list of fragments (specified by fragment number) from
the archived data in an Amazon Kinesis video stream.

You must first call the \code{GetDataEndpoint} API to get an endpoint. Then
send the
\code{\link[=kinesisvideoarchivedmedia_get_media_for_fragment_list]{get_media_for_fragment_list}}
requests to this endpoint using the \href{https://docs.aws.amazon.com/cli/latest/reference/}{--endpoint-url parameter}.

The following limits apply when using the
\code{\link[=kinesisvideoarchivedmedia_get_media_for_fragment_list]{get_media_for_fragment_list}}
API:
\itemize{
\item A client can call
\code{\link[=kinesisvideoarchivedmedia_get_media_for_fragment_list]{get_media_for_fragment_list}}
up to five times per second per stream.
\item Kinesis Video Streams sends media data at a rate of up to 25
megabytes per second (or 200 megabits per second) during a
\code{\link[=kinesisvideoarchivedmedia_get_media_for_fragment_list]{get_media_for_fragment_list}}
session.
}

If an error is thrown after invoking a Kinesis Video Streams archived
media API, in addition to the HTTP status code and the response body, it
includes the following pieces of information:
\itemize{
\item \code{x-amz-ErrorType} HTTP header – contains a more specific error type
in addition to what the HTTP status code provides.
\item \code{x-amz-RequestId} HTTP header – if you want to report an issue to
AWS, the support team can better diagnose the problem if given the
Request Id.
}

Both the HTTP status code and the ErrorType header can be utilized to
make programmatic decisions about whether errors are retry-able and
under what conditions, as well as provide information on what actions
the client programmer might need to take in order to successfully try
again.

For more information, see the \strong{Errors} section at the bottom of this
topic, as well as \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/CommonErrors.html}{Common Errors}.
}
\section{Request syntax}{
\preformatted{svc$get_media_for_fragment_list(
  StreamName = "string",
  Fragments = list(
    "string"
  )
)
}
}

\keyword{internal}
