% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideoarchivedmedia_operations.R
\name{kinesisvideoarchivedmedia_list_fragments}
\alias{kinesisvideoarchivedmedia_list_fragments}
\title{Returns a list of Fragment objects from the specified stream and
timestamp range within the archived data}
\usage{
kinesisvideoarchivedmedia_list_fragments(StreamName, MaxResults,
  NextToken, FragmentSelector)
}
\arguments{
\item{StreamName}{[required] The name of the stream from which to retrieve a fragment list.}

\item{MaxResults}{The total number of fragments to return. If the total number of
fragments available is more than the value specified in \code{max-results},
then a ListFragmentsOutput\\$NextToken is provided in the output that you
can use to resume pagination.}

\item{NextToken}{A token to specify where to start paginating. This is the
ListFragmentsOutput\\$NextToken from a previously truncated response.}

\item{FragmentSelector}{Describes the timestamp range and timestamp origin for the range of
fragments to return.}
}
\description{
Returns a list of Fragment objects from the specified stream and
timestamp range within the archived data.
}
\details{
Listing fragments is eventually consistent. This means that even if the
producer receives an acknowledgment that a fragment is persisted, the
result might not be returned immediately from a request to
\code{ListFragments}. However, results are typically available in less than
one second.

You must first call the \code{GetDataEndpoint} API to get an endpoint. Then
send the \code{ListFragments} requests to this endpoint using the
\href{https://docs.aws.amazon.com/cli/latest/reference/}{\\--endpoint-url parameter}.

If an error is thrown after invoking a Kinesis Video Streams archived
media API, in addition to the HTTP status code and the response body, it
includes the following pieces of information:
\itemize{
\item \code{x-amz-ErrorType} HTTP header -- contains a more specific error type
in addition to what the HTTP status code provides.
\item \code{x-amz-RequestId} HTTP header -- if you want to report an issue to
AWS, the support team can better diagnose the problem if given the
Request Id.
}

Both the HTTP status code and the ErrorType header can be utilized to
make programmatic decisions about whether errors are retry-able and
under what conditions, as well as provide information on what actions
the client programmer might need to take in order to successfully try
again.

For more information, see the \strong{Errors} section at the bottom of this
topic, as well as \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/CommonErrors.html}{Common Errors}.
}
\section{Request syntax}{
\preformatted{svc$list_fragments(
  StreamName = "string",
  MaxResults = 123,
  NextToken = "string",
  FragmentSelector = list(
    FragmentSelectorType = "PRODUCER_TIMESTAMP"|"SERVER_TIMESTAMP",
    TimestampRange = list(
      StartTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      EndTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}

\keyword{internal}
