% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideomedia_operations.R
\name{kinesisvideomedia_get_media}
\alias{kinesisvideomedia_get_media}
\title{Use this API to retrieve media content from a Kinesis video stream}
\usage{
kinesisvideomedia_get_media(StreamName, StreamARN, StartSelector)
}
\arguments{
\item{StreamName}{The Kinesis video stream name from where you want to get the media
content. If you don't specify the \code{streamName}, you must specify the
\code{streamARN}.}

\item{StreamARN}{The ARN of the stream from where you want to get the media content. If
you don't specify the \code{streamARN}, you must specify the \code{streamName}.}

\item{StartSelector}{[required] Identifies the starting chunk to get from the specified stream.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ContentType = "string",
  Payload = raw
)
}\if{html}{\out{</div>}}
}
\description{
Use this API to retrieve media content from a Kinesis video stream. In
the request, you identify the stream name or stream Amazon Resource Name
(ARN), and the starting chunk. Kinesis Video Streams then returns a
stream of chunks in order by fragment number.

You must first call the \code{GetDataEndpoint} API to get an endpoint. Then
send the \code{\link[=kinesisvideomedia_get_media]{get_media}} requests to this
endpoint using the \href{https://docs.aws.amazon.com/cli/latest/reference/}{--endpoint-url parameter}.

When you put media data (fragments) on a stream, Kinesis Video Streams
stores each incoming fragment and related metadata in what is called a
"chunk." For more information, see
\href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/API_dataplane_PutMedia.html}{PutMedia}.
The \code{\link[=kinesisvideomedia_get_media]{get_media}} API returns a stream of
these chunks starting from the chunk that you specify in the request.

The following limits apply when using the
\code{\link[=kinesisvideomedia_get_media]{get_media}} API:
\itemize{
\item A client can call \code{\link[=kinesisvideomedia_get_media]{get_media}} up to
five times per second per stream.
\item Kinesis Video Streams sends media data at a rate of up to 25
megabytes per second (or 200 megabits per second) during a
\code{\link[=kinesisvideomedia_get_media]{get_media}} session.
}

If an error is thrown after invoking a Kinesis Video Streams media API,
in addition to the HTTP status code and the response body, it includes
the following pieces of information:
\itemize{
\item \code{x-amz-ErrorType} HTTP header – contains a more specific error type
in addition to what the HTTP status code provides.
\item \code{x-amz-RequestId} HTTP header – if you want to report an issue to
AWS, the support team can better diagnose the problem if given the
Request Id.
}

Both the HTTP status code and the ErrorType header can be utilized to
make programmatic decisions about whether errors are retry-able and
under what conditions, as well as provide information on what actions
the client programmer might need to take in order to successfully try
again.

For more information, see the \strong{Errors} section at the bottom of this
topic, as well as \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/CommonErrors.html}{Common Errors}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_media(
  StreamName = "string",
  StreamARN = "string",
  StartSelector = list(
    StartSelectorType = "FRAGMENT_NUMBER"|"SERVER_TIMESTAMP"|"PRODUCER_TIMESTAMP"|"NOW"|"EARLIEST"|"CONTINUATION_TOKEN",
    AfterFragmentNumber = "string",
    StartTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    ContinuationToken = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
