% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_add_flow_outputs}
\alias{mediaconnect_add_flow_outputs}
\title{Adds outputs to an existing flow}
\usage{
mediaconnect_add_flow_outputs(FlowArn, Outputs)
}
\arguments{
\item{FlowArn}{[required] The flow that you want to add outputs to.}

\item{Outputs}{[required] A list of outputs that you want to add.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FlowArn = "string",
  Outputs = list(
    list(
      DataTransferSubscriberFeePercent = 123,
      Description = "string",
      Destination = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      EntitlementArn = "string",
      MediaLiveInputArn = "string",
      Name = "string",
      OutputArn = "string",
      Port = 123,
      Transport = list(
        CidrAllowList = list(
          "string"
        ),
        MaxBitrate = 123,
        MaxLatency = 123,
        Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist",
        RemoteId = "string",
        SmoothingLatency = 123,
        StreamId = "string"
      ),
      VpcInterfaceAttachment = list(
        VpcInterfaceName = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Adds outputs to an existing flow. You can create up to 50 outputs per
flow.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$add_flow_outputs(
  FlowArn = "string",
  Outputs = list(
    list(
      CidrAllowList = list(
        "string"
      ),
      Description = "string",
      Destination = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      MaxLatency = 123,
      Name = "string",
      Port = 123,
      Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist",
      RemoteId = "string",
      SmoothingLatency = 123,
      StreamId = "string",
      VpcInterfaceAttachment = list(
        VpcInterfaceName = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
