% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_add_flow_vpc_interfaces}
\alias{mediaconnect_add_flow_vpc_interfaces}
\title{Adds VPC interfaces to flow}
\usage{
mediaconnect_add_flow_vpc_interfaces(FlowArn, VpcInterfaces)
}
\arguments{
\item{FlowArn}{[required] The flow that you want to mutate.}

\item{VpcInterfaces}{[required] A list of VPC interfaces that you want to add.}
}
\description{
Adds VPC interfaces to flow
}
\section{Request syntax}{
\preformatted{svc$add_flow_vpc_interfaces(
  FlowArn = "string",
  VpcInterfaces = list(
    list(
      Name = "string",
      RoleArn = "string",
      SecurityGroupIds = list(
        "string"
      ),
      SubnetId = "string"
    )
  )
)
}
}

\keyword{internal}
