% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_create_flow}
\alias{mediaconnect_create_flow}
\title{Creates a new flow}
\usage{
mediaconnect_create_flow(AvailabilityZone, Entitlements, Name, Outputs,
  Source, SourceFailoverConfig, Sources, VpcInterfaces)
}
\arguments{
\item{AvailabilityZone}{The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS Region.}

\item{Entitlements}{The entitlements that you want to grant on a flow.}

\item{Name}{[required] The name of the flow.}

\item{Outputs}{The outputs that you want to add to this flow.}

\item{Source}{}

\item{SourceFailoverConfig}{}

\item{Sources}{}

\item{VpcInterfaces}{The VPC interfaces you want on the flow.}
}
\description{
Creates a new flow. The request must include one source. The request optionally can include outputs (up to 50) and entitlements (up to 50).
}
\section{Request syntax}{
\preformatted{svc$create_flow(
  AvailabilityZone = "string",
  Entitlements = list(
    list(
      DataTransferSubscriberFeePercent = 123,
      Description = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      Name = "string",
      Subscribers = list(
        "string"
      )
    )
  ),
  Name = "string",
  Outputs = list(
    list(
      CidrAllowList = list(
        "string"
      ),
      Description = "string",
      Destination = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      MaxLatency = 123,
      Name = "string",
      Port = 123,
      Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist",
      RemoteId = "string",
      SmoothingLatency = 123,
      StreamId = "string",
      VpcInterfaceAttachment = list(
        VpcInterfaceName = "string"
      )
    )
  ),
  Source = list(
    Decryption = list(
      Algorithm = "aes128"|"aes192"|"aes256",
      ConstantInitializationVector = "string",
      DeviceId = "string",
      KeyType = "speke"|"static-key",
      Region = "string",
      ResourceId = "string",
      RoleArn = "string",
      SecretArn = "string",
      Url = "string"
    ),
    Description = "string",
    EntitlementArn = "string",
    IngestPort = 123,
    MaxBitrate = 123,
    MaxLatency = 123,
    Name = "string",
    Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist",
    StreamId = "string",
    VpcInterfaceName = "string",
    WhitelistCidr = "string"
  ),
  SourceFailoverConfig = list(
    RecoveryWindow = 123,
    State = "ENABLED"|"DISABLED"
  ),
  Sources = list(
    list(
      Decryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      Description = "string",
      EntitlementArn = "string",
      IngestPort = 123,
      MaxBitrate = 123,
      MaxLatency = 123,
      Name = "string",
      Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist",
      StreamId = "string",
      VpcInterfaceName = "string",
      WhitelistCidr = "string"
    )
  ),
  VpcInterfaces = list(
    list(
      Name = "string",
      RoleArn = "string",
      SecurityGroupIds = list(
        "string"
      ),
      SubnetId = "string"
    )
  )
)
}
}

\keyword{internal}
