% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_describe_offering}
\alias{mediaconnect_describe_offering}
\title{Displays the details of an offering}
\usage{
mediaconnect_describe_offering(OfferingArn)
}
\arguments{
\item{OfferingArn}{[required] The Amazon Resource Name (ARN) of the offering.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Offering = list(
    CurrencyCode = "string",
    Duration = 123,
    DurationUnits = "MONTHS",
    OfferingArn = "string",
    OfferingDescription = "string",
    PricePerUnit = "string",
    PriceUnits = "HOURLY",
    ResourceSpecification = list(
      ReservedBitrate = 123,
      ResourceType = "Mbps_Outbound_Bandwidth"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Displays the details of an offering. The response includes the offering
description, duration, outbound bandwidth, price, and Amazon Resource
Name (ARN).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_offering(
  OfferingArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
