% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_describe_reservation}
\alias{mediaconnect_describe_reservation}
\title{Displays the details of a reservation}
\usage{
mediaconnect_describe_reservation(ReservationArn)
}
\arguments{
\item{ReservationArn}{[required] The Amazon Resource Name (ARN) of the reservation.}
}
\description{
Displays the details of a reservation. The response includes the reservation name, state, start date and time, and the details of the offering that make up the rest of the reservation (such as price, duration, and outbound bandwidth).
}
\section{Request syntax}{
\preformatted{svc$describe_reservation(
  ReservationArn = "string"
)
}
}

\keyword{internal}
