% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_grant_flow_entitlements}
\alias{mediaconnect_grant_flow_entitlements}
\title{Grants entitlements to an existing flow}
\usage{
mediaconnect_grant_flow_entitlements(Entitlements, FlowArn)
}
\arguments{
\item{Entitlements}{[required] The list of entitlements that you want to grant.}

\item{FlowArn}{[required] The flow that you want to grant entitlements on.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Entitlements = list(
    list(
      DataTransferSubscriberFeePercent = 123,
      Description = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      EntitlementArn = "string",
      EntitlementStatus = "ENABLED"|"DISABLED",
      Name = "string",
      Subscribers = list(
        "string"
      )
    )
  ),
  FlowArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Grants entitlements to an existing flow.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$grant_flow_entitlements(
  Entitlements = list(
    list(
      DataTransferSubscriberFeePercent = 123,
      Description = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      EntitlementStatus = "ENABLED"|"DISABLED",
      Name = "string",
      Subscribers = list(
        "string"
      )
    )
  ),
  FlowArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
