% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_list_tags_for_resource}
\alias{mediaconnect_list_tags_for_resource}
\title{List all tags on an AWS Elemental MediaConnect resource}
\usage{
mediaconnect_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the AWS Elemental
MediaConnect resource for which to list the tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    "string"
  )
)
}
}
\description{
List all tags on an AWS Elemental MediaConnect resource
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
