% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_remove_flow_source}
\alias{mediaconnect_remove_flow_source}
\title{Removes a source from an existing flow}
\usage{
mediaconnect_remove_flow_source(FlowArn, SourceArn)
}
\arguments{
\item{FlowArn}{[required] The flow that you want to remove a source from.}

\item{SourceArn}{[required] The ARN of the source that you want to remove.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FlowArn = "string",
  SourceArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Removes a source from an existing flow. This request can be made only if
there is more than one source on the flow.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$remove_flow_source(
  FlowArn = "string",
  SourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
