% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_remove_flow_vpc_interface}
\alias{mediaconnect_remove_flow_vpc_interface}
\title{Removes a VPC Interface from an existing flow}
\usage{
mediaconnect_remove_flow_vpc_interface(FlowArn, VpcInterfaceName)
}
\arguments{
\item{FlowArn}{[required] The flow that you want to remove a VPC interface from.}

\item{VpcInterfaceName}{[required] The name of the VPC interface that you want to remove.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FlowArn = "string",
  NonDeletedNetworkInterfaceIds = list(
    "string"
  ),
  VpcInterfaceName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Removes a VPC Interface from an existing flow. This request can be made
only on a VPC interface that does not have a Source or Output associated
with it. If the VPC interface is referenced by a Source or Output, you
must first delete or update the Source or Output to no longer reference
the VPC interface.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$remove_flow_vpc_interface(
  FlowArn = "string",
  VpcInterfaceName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
