% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_revoke_flow_entitlement}
\alias{mediaconnect_revoke_flow_entitlement}
\title{Revokes an entitlement from a flow}
\usage{
mediaconnect_revoke_flow_entitlement(EntitlementArn, FlowArn)
}
\arguments{
\item{EntitlementArn}{[required] The ARN of the entitlement that you want to revoke.}

\item{FlowArn}{[required] The flow that you want to revoke an entitlement from.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EntitlementArn = "string",
  FlowArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Revokes an entitlement from a flow. Once an entitlement is revoked, the
content becomes unavailable to the subscriber and the associated output
is removed.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$revoke_flow_entitlement(
  EntitlementArn = "string",
  FlowArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
