% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_start_flow}
\alias{mediaconnect_start_flow}
\title{Starts a flow}
\usage{
mediaconnect_start_flow(FlowArn)
}
\arguments{
\item{FlowArn}{[required] The ARN of the flow that you want to start.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FlowArn = "string",
  Status = "STANDBY"|"ACTIVE"|"UPDATING"|"DELETING"|"STARTING"|"STOPPING"|"ERROR"
)
}\if{html}{\out{</div>}}
}
\description{
Starts a flow.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_flow(
  FlowArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
