% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_tag_resource}
\alias{mediaconnect_tag_resource}
\title{Associates the specified tags to a resource with the specified
resourceArn}
\usage{
mediaconnect_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the AWS Elemental
MediaConnect resource to which to add tags.}

\item{Tags}{[required] A map from tag keys to values. Tag keys can have a maximum character
length of 128 characters, and tag values can have a maximum length of
256 characters.}
}
\value{
An empty list.
}
\description{
Associates the specified tags to a resource with the specified
resourceArn. If existing tags on a resource are not specified in the
request parameters, they are not changed. When a resource is deleted,
the tags associated with that resource are deleted as well.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
