% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_untag_resource}
\alias{mediaconnect_untag_resource}
\title{Deletes specified tags from a resource}
\usage{
mediaconnect_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the AWS Elemental
MediaConnect resource from which to delete tags.}

\item{TagKeys}{[required] The keys of the tags to be removed.}
}
\value{
An empty list.
}
\description{
Deletes specified tags from a resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
