% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_update_flow_output}
\alias{mediaconnect_update_flow_output}
\title{Updates an existing flow output}
\usage{
mediaconnect_update_flow_output(CidrAllowList, Description, Destination,
  Encryption, FlowArn, MaxLatency, MinLatency, OutputArn, Port, Protocol,
  RemoteId, SmoothingLatency, StreamId, VpcInterfaceAttachment)
}
\arguments{
\item{CidrAllowList}{The range of IP addresses that should be allowed to initiate output
requests to this flow. These IP addresses should be in the form of a
Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.}

\item{Description}{A description of the output. This description appears only on the AWS
Elemental MediaConnect console and will not be seen by the end user.}

\item{Destination}{The IP address where you want to send the output.}

\item{Encryption}{The type of key used for the encryption. If no keyType is provided, the
service will use the default setting (static-key).}

\item{FlowArn}{[required] The flow that is associated with the output that you want to update.}

\item{MaxLatency}{The maximum latency in milliseconds for Zixi-based streams.}

\item{MinLatency}{The minimum latency in milliseconds for SRT-based streams. In streams
that use the SRT protocol, this value that you set on your MediaConnect
source or output represents the minimal potential latency of that
connection. The latency of the stream is set to the highest number
between the sender’s minimum latency and the receiver’s minimum latency.}

\item{OutputArn}{[required] The ARN of the output that you want to update.}

\item{Port}{The port to use when content is distributed to this output.}

\item{Protocol}{The protocol to use for the output.}

\item{RemoteId}{The remote ID for the Zixi-pull stream.}

\item{SmoothingLatency}{The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC
streams.}

\item{StreamId}{The stream ID that you want to use for this transport. This parameter
applies only to Zixi-based streams.}

\item{VpcInterfaceAttachment}{The name of the VPC interface attachment to use for this output.}
}
\value{
A list with the following syntax:\preformatted{list(
  FlowArn = "string",
  Output = list(
    DataTransferSubscriberFeePercent = 123,
    Description = "string",
    Destination = "string",
    Encryption = list(
      Algorithm = "aes128"|"aes192"|"aes256",
      ConstantInitializationVector = "string",
      DeviceId = "string",
      KeyType = "speke"|"static-key"|"srt-password",
      Region = "string",
      ResourceId = "string",
      RoleArn = "string",
      SecretArn = "string",
      Url = "string"
    ),
    EntitlementArn = "string",
    ListenerAddress = "string",
    MediaLiveInputArn = "string",
    Name = "string",
    OutputArn = "string",
    Port = 123,
    Transport = list(
      CidrAllowList = list(
        "string"
      ),
      MaxBitrate = 123,
      MaxLatency = 123,
      MinLatency = 123,
      Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist"|"srt-listener",
      RemoteId = "string",
      SmoothingLatency = 123,
      StreamId = "string"
    ),
    VpcInterfaceAttachment = list(
      VpcInterfaceName = "string"
    )
  )
)
}
}
\description{
Updates an existing flow output.
}
\section{Request syntax}{
\preformatted{svc$update_flow_output(
  CidrAllowList = list(
    "string"
  ),
  Description = "string",
  Destination = "string",
  Encryption = list(
    Algorithm = "aes128"|"aes192"|"aes256",
    ConstantInitializationVector = "string",
    DeviceId = "string",
    KeyType = "speke"|"static-key"|"srt-password",
    Region = "string",
    ResourceId = "string",
    RoleArn = "string",
    SecretArn = "string",
    Url = "string"
  ),
  FlowArn = "string",
  MaxLatency = 123,
  MinLatency = 123,
  OutputArn = "string",
  Port = 123,
  Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist"|"srt-listener",
  RemoteId = "string",
  SmoothingLatency = 123,
  StreamId = "string",
  VpcInterfaceAttachment = list(
    VpcInterfaceName = "string"
  )
)
}
}

\keyword{internal}
