% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_update_flow_source}
\alias{mediaconnect_update_flow_source}
\title{Updates the source of a flow}
\usage{
mediaconnect_update_flow_source(Decryption, Description, EntitlementArn,
  FlowArn, IngestPort, MaxBitrate, MaxLatency, MinLatency, Protocol,
  SourceArn, StreamId, VpcInterfaceName, WhitelistCidr)
}
\arguments{
\item{Decryption}{The type of encryption used on the content ingested from this source.}

\item{Description}{A description for the source. This value is not used or seen outside of
the current AWS Elemental MediaConnect account.}

\item{EntitlementArn}{The ARN of the entitlement that allows you to subscribe to this flow.
The entitlement is set by the flow originator, and the ARN is generated
as part of the originator's flow.}

\item{FlowArn}{[required] The flow that is associated with the source that you want to update.}

\item{IngestPort}{The port that the flow will be listening on for incoming content.}

\item{MaxBitrate}{The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.}

\item{MaxLatency}{The maximum latency in milliseconds. This parameter applies only to
RIST-based and Zixi-based streams.}

\item{MinLatency}{The minimum latency in milliseconds for SRT-based streams. In streams
that use the SRT protocol, this value that you set on your MediaConnect
source or output represents the minimal potential latency of that
connection. The latency of the stream is set to the highest number
between the sender’s minimum latency and the receiver’s minimum latency.}

\item{Protocol}{The protocol that is used by the source.}

\item{SourceArn}{[required] The ARN of the source that you want to update.}

\item{StreamId}{The stream ID that you want to use for this transport. This parameter
applies only to Zixi-based streams.}

\item{VpcInterfaceName}{The name of the VPC Interface to configure this Source with.}

\item{WhitelistCidr}{The range of IP addresses that should be allowed to contribute content
to your source. These IP addresses should be in the form of a Classless
Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.}
}
\value{
A list with the following syntax:\preformatted{list(
  FlowArn = "string",
  Source = list(
    DataTransferSubscriberFeePercent = 123,
    Decryption = list(
      Algorithm = "aes128"|"aes192"|"aes256",
      ConstantInitializationVector = "string",
      DeviceId = "string",
      KeyType = "speke"|"static-key"|"srt-password",
      Region = "string",
      ResourceId = "string",
      RoleArn = "string",
      SecretArn = "string",
      Url = "string"
    ),
    Description = "string",
    EntitlementArn = "string",
    IngestIp = "string",
    IngestPort = 123,
    Name = "string",
    SourceArn = "string",
    Transport = list(
      CidrAllowList = list(
        "string"
      ),
      MaxBitrate = 123,
      MaxLatency = 123,
      MinLatency = 123,
      Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist"|"srt-listener",
      RemoteId = "string",
      SmoothingLatency = 123,
      StreamId = "string"
    ),
    VpcInterfaceName = "string",
    WhitelistCidr = "string"
  )
)
}
}
\description{
Updates the source of a flow.
}
\section{Request syntax}{
\preformatted{svc$update_flow_source(
  Decryption = list(
    Algorithm = "aes128"|"aes192"|"aes256",
    ConstantInitializationVector = "string",
    DeviceId = "string",
    KeyType = "speke"|"static-key"|"srt-password",
    Region = "string",
    ResourceId = "string",
    RoleArn = "string",
    SecretArn = "string",
    Url = "string"
  ),
  Description = "string",
  EntitlementArn = "string",
  FlowArn = "string",
  IngestPort = 123,
  MaxBitrate = 123,
  MaxLatency = 123,
  MinLatency = 123,
  Protocol = "zixi-push"|"rtp-fec"|"rtp"|"zixi-pull"|"rist"|"srt-listener",
  SourceArn = "string",
  StreamId = "string",
  VpcInterfaceName = "string",
  WhitelistCidr = "string"
)
}
}

\keyword{internal}
