% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_service.R
\name{mediaconvert}
\alias{mediaconvert}
\title{AWS Elemental MediaConvert}
\usage{
mediaconvert()
}
\description{
AWS Elemental MediaConvert
}
\section{Operations}{

\tabular{ll}{
\link[=mediaconvert_associate_certificate]{associate_certificate} \tab Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN) with AWS Elemental MediaConvert \cr
\link[=mediaconvert_cancel_job]{cancel_job} \tab Permanently cancel a job \cr
\link[=mediaconvert_create_job]{create_job} \tab Create a new transcoding job \cr
\link[=mediaconvert_create_job_template]{create_job_template} \tab Create a new job template \cr
\link[=mediaconvert_create_preset]{create_preset} \tab Create a new preset \cr
\link[=mediaconvert_create_queue]{create_queue} \tab Create a new transcoding queue \cr
\link[=mediaconvert_delete_job_template]{delete_job_template} \tab Permanently delete a job template you have created \cr
\link[=mediaconvert_delete_preset]{delete_preset} \tab Permanently delete a preset you have created \cr
\link[=mediaconvert_delete_queue]{delete_queue} \tab Permanently delete a queue you have created \cr
\link[=mediaconvert_describe_endpoints]{describe_endpoints} \tab Send an request with an empty body to the regional API endpoint to get your account API endpoint \cr
\link[=mediaconvert_disassociate_certificate]{disassociate_certificate} \tab Removes an association between the Amazon Resource Name (ARN) of an AWS Certificate Manager (ACM) certificate and an AWS Elemental MediaConvert resource\cr
\link[=mediaconvert_get_job]{get_job} \tab Retrieve the JSON for a specific completed transcoding job \cr
\link[=mediaconvert_get_job_template]{get_job_template} \tab Retrieve the JSON for a specific job template \cr
\link[=mediaconvert_get_preset]{get_preset} \tab Retrieve the JSON for a specific preset \cr
\link[=mediaconvert_get_queue]{get_queue} \tab Retrieve the JSON for a specific queue \cr
\link[=mediaconvert_list_job_templates]{list_job_templates} \tab Retrieve a JSON array of up to twenty of your job templates \cr
\link[=mediaconvert_list_jobs]{list_jobs} \tab Retrieve a JSON array of up to twenty of your most recently created jobs \cr
\link[=mediaconvert_list_presets]{list_presets} \tab Retrieve a JSON array of up to twenty of your presets \cr
\link[=mediaconvert_list_queues]{list_queues} \tab Retrieve a JSON array of up to twenty of your queues \cr
\link[=mediaconvert_list_tags_for_resource]{list_tags_for_resource} \tab Retrieve the tags for a MediaConvert resource \cr
\link[=mediaconvert_tag_resource]{tag_resource} \tab Add tags to a MediaConvert queue, preset, or job template \cr
\link[=mediaconvert_untag_resource]{untag_resource} \tab Remove tags from a MediaConvert queue, preset, or job template \cr
\link[=mediaconvert_update_job_template]{update_job_template} \tab Modify one of your existing job templates \cr
\link[=mediaconvert_update_preset]{update_preset} \tab Modify one of your existing presets \cr
\link[=mediaconvert_update_queue]{update_queue} \tab Modify one of your existing queues
}
}

\examples{
\donttest{svc <- mediaconvert()
svc$associate_certificate(
  Foo = 123
)}

}
