% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_delete_preset}
\alias{mediaconvert_delete_preset}
\title{Permanently delete a preset you have created}
\usage{
mediaconvert_delete_preset(Name)
}
\arguments{
\item{Name}{[required] The name of the preset to be deleted.}
}
\value{
An empty list.
}
\description{
Permanently delete a preset you have created.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_preset(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
