% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_delete_queue}
\alias{mediaconvert_delete_queue}
\title{Permanently delete a queue you have created}
\usage{
mediaconvert_delete_queue(Name)
}
\arguments{
\item{Name}{[required] The name of the queue that you want to delete.}
}
\value{
An empty list.
}
\description{
Permanently delete a queue you have created.
}
\section{Request syntax}{
\preformatted{svc$delete_queue(
  Name = "string"
)
}
}

\keyword{internal}
