% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_update_preset}
\alias{mediaconvert_update_preset}
\title{Modify one of your existing presets}
\usage{
mediaconvert_update_preset(Category, Description, Name, Settings)
}
\arguments{
\item{Category}{The new category for the preset, if you are changing it.}

\item{Description}{The new description for the preset, if you are changing it.}

\item{Name}{[required] The name of the preset you are modifying.}

\item{Settings}{Settings for preset}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Preset = list(
    Arn = "string",
    Category = "string",
    CreatedAt = as.POSIXct(
      "2015-01-01"
    ),
    Description = "string",
    LastUpdated = as.POSIXct(
      "2015-01-01"
    ),
    Name = "string",
    Settings = list(
      AudioDescriptions = list(
        list(
          AudioChannelTaggingSettings = list(
            ChannelTag = "L"|"R"|"C"|"LFE"|"LS"|"RS"|"LC"|"RC"|"CS"|"LSD"|"RSD"|"TCS"|"VHL"|"VHC"|"VHR"
          ),
          AudioNormalizationSettings = list(
            Algorithm = "ITU_BS_1770_1"|"ITU_BS_1770_2"|"ITU_BS_1770_3"|"ITU_BS_1770_4",
            AlgorithmControl = "CORRECT_AUDIO"|"MEASURE_ONLY",
            CorrectionGateLevel = 123,
            LoudnessLogging = "LOG"|"DONT_LOG",
            PeakCalculation = "TRUE_PEAK"|"NONE",
            TargetLkfs = 123.0
          ),
          AudioSourceName = "string",
          AudioType = 123,
          AudioTypeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
          CodecSettings = list(
            AacSettings = list(
              AudioDescriptionBroadcasterMix = "BROADCASTER_MIXED_AD"|"NORMAL",
              Bitrate = 123,
              CodecProfile = "LC"|"HEV1"|"HEV2",
              CodingMode = "AD_RECEIVER_MIX"|"CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_5_1",
              RateControlMode = "CBR"|"VBR",
              RawFormat = "LATM_LOAS"|"NONE",
              SampleRate = 123,
              Specification = "MPEG2"|"MPEG4",
              VbrQuality = "LOW"|"MEDIUM_LOW"|"MEDIUM_HIGH"|"HIGH"
            ),
            Ac3Settings = list(
              Bitrate = 123,
              BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"DIALOGUE"|"EMERGENCY"|"HEARING_IMPAIRED"|"MUSIC_AND_EFFECTS"|"VISUALLY_IMPAIRED"|"VOICE_OVER",
              CodingMode = "CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_3_2_LFE",
              Dialnorm = 123,
              DynamicRangeCompressionProfile = "FILM_STANDARD"|"NONE",
              LfeFilter = "ENABLED"|"DISABLED",
              MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
              SampleRate = 123
            ),
            AiffSettings = list(
              BitDepth = 123,
              Channels = 123,
              SampleRate = 123
            ),
            Codec = "AAC"|"MP2"|"MP3"|"WAV"|"AIFF"|"AC3"|"EAC3"|"EAC3_ATMOS"|"VORBIS"|"OPUS"|"PASSTHROUGH",
            Eac3AtmosSettings = list(
              Bitrate = 123,
              BitstreamMode = "COMPLETE_MAIN",
              CodingMode = "CODING_MODE_9_1_6",
              DialogueIntelligence = "ENABLED"|"DISABLED",
              DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
              DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
              LoRoCenterMixLevel = 123.0,
              LoRoSurroundMixLevel = 123.0,
              LtRtCenterMixLevel = 123.0,
              LtRtSurroundMixLevel = 123.0,
              MeteringMode = "LEQ_A"|"ITU_BS_1770_1"|"ITU_BS_1770_2"|"ITU_BS_1770_3"|"ITU_BS_1770_4",
              SampleRate = 123,
              SpeechThreshold = 123,
              StereoDownmix = "NOT_INDICATED"|"STEREO"|"SURROUND"|"DPL2",
              SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
            ),
            Eac3Settings = list(
              AttenuationControl = "ATTENUATE_3_DB"|"NONE",
              Bitrate = 123,
              BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"EMERGENCY"|"HEARING_IMPAIRED"|"VISUALLY_IMPAIRED",
              CodingMode = "CODING_MODE_1_0"|"CODING_MODE_2_0"|"CODING_MODE_3_2",
              DcFilter = "ENABLED"|"DISABLED",
              Dialnorm = 123,
              DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
              DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
              LfeControl = "LFE"|"NO_LFE",
              LfeFilter = "ENABLED"|"DISABLED",
              LoRoCenterMixLevel = 123.0,
              LoRoSurroundMixLevel = 123.0,
              LtRtCenterMixLevel = 123.0,
              LtRtSurroundMixLevel = 123.0,
              MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
              PassthroughControl = "WHEN_POSSIBLE"|"NO_PASSTHROUGH",
              PhaseControl = "SHIFT_90_DEGREES"|"NO_SHIFT",
              SampleRate = 123,
              StereoDownmix = "NOT_INDICATED"|"LO_RO"|"LT_RT"|"DPL2",
              SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED",
              SurroundMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
            ),
            Mp2Settings = list(
              Bitrate = 123,
              Channels = 123,
              SampleRate = 123
            ),
            Mp3Settings = list(
              Bitrate = 123,
              Channels = 123,
              RateControlMode = "CBR"|"VBR",
              SampleRate = 123,
              VbrQuality = 123
            ),
            OpusSettings = list(
              Bitrate = 123,
              Channels = 123,
              SampleRate = 123
            ),
            VorbisSettings = list(
              Channels = 123,
              SampleRate = 123,
              VbrQuality = 123
            ),
            WavSettings = list(
              BitDepth = 123,
              Channels = 123,
              Format = "RIFF"|"RF64",
              SampleRate = 123
            )
          ),
          CustomLanguageCode = "string",
          LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
          LanguageCodeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
          RemixSettings = list(
            ChannelMapping = list(
              OutputChannels = list(
                list(
                  InputChannels = list(
                    123
                  ),
                  InputChannelsFineTune = list(
                    123.0
                  )
                )
              )
            ),
            ChannelsIn = 123,
            ChannelsOut = 123
          ),
          StreamName = "string"
        )
      ),
      CaptionDescriptions = list(
        list(
          CustomLanguageCode = "string",
          DestinationSettings = list(
            BurninDestinationSettings = list(
              Alignment = "CENTERED"|"LEFT",
              BackgroundColor = "NONE"|"BLACK"|"WHITE",
              BackgroundOpacity = 123,
              FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
              FontOpacity = 123,
              FontResolution = 123,
              FontScript = "AUTOMATIC"|"HANS"|"HANT",
              FontSize = 123,
              OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
              OutlineSize = 123,
              ShadowColor = "NONE"|"BLACK"|"WHITE",
              ShadowOpacity = 123,
              ShadowXOffset = 123,
              ShadowYOffset = 123,
              TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
              XPosition = 123,
              YPosition = 123
            ),
            DestinationType = "BURN_IN"|"DVB_SUB"|"EMBEDDED"|"EMBEDDED_PLUS_SCTE20"|"IMSC"|"SCTE20_PLUS_EMBEDDED"|"SCC"|"SRT"|"SMI"|"TELETEXT"|"TTML"|"WEBVTT",
            DvbSubDestinationSettings = list(
              Alignment = "CENTERED"|"LEFT",
              BackgroundColor = "NONE"|"BLACK"|"WHITE",
              BackgroundOpacity = 123,
              FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
              FontOpacity = 123,
              FontResolution = 123,
              FontScript = "AUTOMATIC"|"HANS"|"HANT",
              FontSize = 123,
              OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
              OutlineSize = 123,
              ShadowColor = "NONE"|"BLACK"|"WHITE",
              ShadowOpacity = 123,
              ShadowXOffset = 123,
              ShadowYOffset = 123,
              SubtitlingType = "HEARING_IMPAIRED"|"STANDARD",
              TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
              XPosition = 123,
              YPosition = 123
            ),
            EmbeddedDestinationSettings = list(
              Destination608ChannelNumber = 123,
              Destination708ServiceNumber = 123
            ),
            ImscDestinationSettings = list(
              StylePassthrough = "ENABLED"|"DISABLED"
            ),
            SccDestinationSettings = list(
              Framerate = "FRAMERATE_23_97"|"FRAMERATE_24"|"FRAMERATE_25"|"FRAMERATE_29_97_DROPFRAME"|"FRAMERATE_29_97_NON_DROPFRAME"
            ),
            TeletextDestinationSettings = list(
              PageNumber = "string",
              PageTypes = list(
                "PAGE_TYPE_INITIAL"|"PAGE_TYPE_SUBTITLE"|"PAGE_TYPE_ADDL_INFO"|"PAGE_TYPE_PROGRAM_SCHEDULE"|"PAGE_TYPE_HEARING_IMPAIRED_SUBTITLE"
              )
            ),
            TtmlDestinationSettings = list(
              StylePassthrough = "ENABLED"|"DISABLED"
            )
          ),
          LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
          LanguageDescription = "string"
        )
      ),
      ContainerSettings = list(
        CmfcSettings = list(
          AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
          IFrameOnlyManifest = "INCLUDE"|"EXCLUDE",
          Scte35Esam = "INSERT"|"NONE",
          Scte35Source = "PASSTHROUGH"|"NONE"
        ),
        Container = "F4V"|"ISMV"|"M2TS"|"M3U8"|"CMFC"|"MOV"|"MP4"|"MPD"|"MXF"|"WEBM"|"RAW",
        F4vSettings = list(
          MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL"
        ),
        M2tsSettings = list(
          AudioBufferModel = "DVB"|"ATSC",
          AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
          AudioFramesPerPes = 123,
          AudioPids = list(
            123
          ),
          Bitrate = 123,
          BufferModel = "MULTIPLEX"|"NONE",
          DvbNitSettings = list(
            NetworkId = 123,
            NetworkName = "string",
            NitInterval = 123
          ),
          DvbSdtSettings = list(
            OutputSdt = "SDT_FOLLOW"|"SDT_FOLLOW_IF_PRESENT"|"SDT_MANUAL"|"SDT_NONE",
            SdtInterval = 123,
            ServiceName = "string",
            ServiceProviderName = "string"
          ),
          DvbSubPids = list(
            123
          ),
          DvbTdtSettings = list(
            TdtInterval = 123
          ),
          DvbTeletextPid = 123,
          EbpAudioInterval = "VIDEO_AND_FIXED_INTERVALS"|"VIDEO_INTERVAL",
          EbpPlacement = "VIDEO_AND_AUDIO_PIDS"|"VIDEO_PID",
          EsRateInPes = "INCLUDE"|"EXCLUDE",
          ForceTsVideoEbpOrder = "FORCE"|"DEFAULT",
          FragmentTime = 123.0,
          MaxPcrInterval = 123,
          MinEbpInterval = 123,
          NielsenId3 = "INSERT"|"NONE",
          NullPacketBitrate = 123.0,
          PatInterval = 123,
          PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
          PcrPid = 123,
          PmtInterval = 123,
          PmtPid = 123,
          PrivateMetadataPid = 123,
          ProgramNumber = 123,
          RateMode = "VBR"|"CBR",
          Scte35Esam = list(
            Scte35EsamPid = 123
          ),
          Scte35Pid = 123,
          Scte35Source = "PASSTHROUGH"|"NONE",
          SegmentationMarkers = "NONE"|"RAI_SEGSTART"|"RAI_ADAPT"|"PSI_SEGSTART"|"EBP"|"EBP_LEGACY",
          SegmentationStyle = "MAINTAIN_CADENCE"|"RESET_CADENCE",
          SegmentationTime = 123.0,
          TimedMetadataPid = 123,
          TransportStreamId = 123,
          VideoPid = 123
        ),
        M3u8Settings = list(
          AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
          AudioFramesPerPes = 123,
          AudioPids = list(
            123
          ),
          NielsenId3 = "INSERT"|"NONE",
          PatInterval = 123,
          PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
          PcrPid = 123,
          PmtInterval = 123,
          PmtPid = 123,
          PrivateMetadataPid = 123,
          ProgramNumber = 123,
          Scte35Pid = 123,
          Scte35Source = "PASSTHROUGH"|"NONE",
          TimedMetadata = "PASSTHROUGH"|"NONE",
          TimedMetadataPid = 123,
          TransportStreamId = 123,
          VideoPid = 123
        ),
        MovSettings = list(
          ClapAtom = "INCLUDE"|"EXCLUDE",
          CslgAtom = "INCLUDE"|"EXCLUDE",
          Mpeg2FourCCControl = "XDCAM"|"MPEG",
          PaddingControl = "OMNEON"|"NONE",
          Reference = "SELF_CONTAINED"|"EXTERNAL"
        ),
        Mp4Settings = list(
          AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
          CslgAtom = "INCLUDE"|"EXCLUDE",
          CttsVersion = 123,
          FreeSpaceBox = "INCLUDE"|"EXCLUDE",
          MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL",
          Mp4MajorBrand = "string"
        ),
        MpdSettings = list(
          AccessibilityCaptionHints = "INCLUDE"|"EXCLUDE",
          AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
          CaptionContainerType = "RAW"|"FRAGMENTED_MP4",
          Scte35Esam = "INSERT"|"NONE",
          Scte35Source = "PASSTHROUGH"|"NONE"
        ),
        MxfSettings = list(
          AfdSignaling = "NO_COPY"|"COPY_FROM_VIDEO",
          Profile = "D_10"|"XDCAM"|"OP1A"
        )
      ),
      VideoDescription = list(
        AfdSignaling = "NONE"|"AUTO"|"FIXED",
        AntiAlias = "DISABLED"|"ENABLED",
        CodecSettings = list(
          Av1Settings = list(
            AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            GopSize = 123.0,
            MaxBitrate = 123,
            NumberBFramesBetweenReferenceFrames = 123,
            QvbrSettings = list(
              QvbrQualityLevel = 123,
              QvbrQualityLevelFineTune = 123.0
            ),
            RateControlMode = "QVBR",
            Slices = 123,
            SpatialAdaptiveQuantization = "DISABLED"|"ENABLED"
          ),
          AvcIntraSettings = list(
            AvcIntraClass = "CLASS_50"|"CLASS_100"|"CLASS_200"|"CLASS_4K_2K",
            AvcIntraUhdSettings = list(
              QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS"
            ),
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
            ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
            SlowPal = "DISABLED"|"ENABLED",
            Telecine = "NONE"|"HARD"
          ),
          Codec = "AV1"|"AVC_INTRA"|"FRAME_CAPTURE"|"H_264"|"H_265"|"MPEG2"|"PRORES"|"VC3"|"VP8"|"VP9",
          FrameCaptureSettings = list(
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            MaxCaptures = 123,
            Quality = 123
          ),
          H264Settings = list(
            AdaptiveQuantization = "OFF"|"AUTO"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
            Bitrate = 123,
            CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_1_1"|"LEVEL_1_2"|"LEVEL_1_3"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_2_2"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_3_2"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_4_2"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2",
            CodecProfile = "BASELINE"|"HIGH"|"HIGH_10BIT"|"HIGH_422"|"HIGH_422_10BIT"|"MAIN",
            DynamicSubGop = "ADAPTIVE"|"STATIC",
            EntropyEncoding = "CABAC"|"CAVLC",
            FieldEncoding = "PAFF"|"FORCE_FIELD",
            FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            GopBReference = "DISABLED"|"ENABLED",
            GopClosedCadence = 123,
            GopSize = 123.0,
            GopSizeUnits = "FRAMES"|"SECONDS",
            HrdBufferInitialFillPercentage = 123,
            HrdBufferSize = 123,
            InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
            MaxBitrate = 123,
            MinIInterval = 123,
            NumberBFramesBetweenReferenceFrames = 123,
            NumberReferenceFrames = 123,
            ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            ParDenominator = 123,
            ParNumerator = 123,
            QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
            QvbrSettings = list(
              MaxAverageBitrate = 123,
              QvbrQualityLevel = 123,
              QvbrQualityLevelFineTune = 123.0
            ),
            RateControlMode = "VBR"|"CBR"|"QVBR",
            RepeatPps = "DISABLED"|"ENABLED",
            ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
            SceneChangeDetect = "DISABLED"|"ENABLED"|"TRANSITION_DETECTION",
            Slices = 123,
            SlowPal = "DISABLED"|"ENABLED",
            Softness = 123,
            SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
            Syntax = "DEFAULT"|"RP2027",
            Telecine = "NONE"|"SOFT"|"HARD",
            TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
            UnregisteredSeiTimecode = "DISABLED"|"ENABLED"
          ),
          H265Settings = list(
            AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
            AlternateTransferFunctionSei = "DISABLED"|"ENABLED",
            Bitrate = 123,
            CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2"|"LEVEL_6"|"LEVEL_6_1"|"LEVEL_6_2",
            CodecProfile = "MAIN_MAIN"|"MAIN_HIGH"|"MAIN10_MAIN"|"MAIN10_HIGH"|"MAIN_422_8BIT_MAIN"|"MAIN_422_8BIT_HIGH"|"MAIN_422_10BIT_MAIN"|"MAIN_422_10BIT_HIGH",
            DynamicSubGop = "ADAPTIVE"|"STATIC",
            FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            GopBReference = "DISABLED"|"ENABLED",
            GopClosedCadence = 123,
            GopSize = 123.0,
            GopSizeUnits = "FRAMES"|"SECONDS",
            HrdBufferInitialFillPercentage = 123,
            HrdBufferSize = 123,
            InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
            MaxBitrate = 123,
            MinIInterval = 123,
            NumberBFramesBetweenReferenceFrames = 123,
            NumberReferenceFrames = 123,
            ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            ParDenominator = 123,
            ParNumerator = 123,
            QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
            QvbrSettings = list(
              MaxAverageBitrate = 123,
              QvbrQualityLevel = 123,
              QvbrQualityLevelFineTune = 123.0
            ),
            RateControlMode = "VBR"|"CBR"|"QVBR",
            SampleAdaptiveOffsetFilterMode = "DEFAULT"|"ADAPTIVE"|"OFF",
            ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
            SceneChangeDetect = "DISABLED"|"ENABLED"|"TRANSITION_DETECTION",
            Slices = 123,
            SlowPal = "DISABLED"|"ENABLED",
            SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
            Telecine = "NONE"|"SOFT"|"HARD",
            TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
            TemporalIds = "DISABLED"|"ENABLED",
            Tiles = "DISABLED"|"ENABLED",
            UnregisteredSeiTimecode = "DISABLED"|"ENABLED",
            WriteMp4PackagingType = "HVC1"|"HEV1"
          ),
          Mpeg2Settings = list(
            AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH",
            Bitrate = 123,
            CodecLevel = "AUTO"|"LOW"|"MAIN"|"HIGH1440"|"HIGH",
            CodecProfile = "MAIN"|"PROFILE_422",
            DynamicSubGop = "ADAPTIVE"|"STATIC",
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            GopClosedCadence = 123,
            GopSize = 123.0,
            GopSizeUnits = "FRAMES"|"SECONDS",
            HrdBufferInitialFillPercentage = 123,
            HrdBufferSize = 123,
            InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
            IntraDcPrecision = "AUTO"|"INTRA_DC_PRECISION_8"|"INTRA_DC_PRECISION_9"|"INTRA_DC_PRECISION_10"|"INTRA_DC_PRECISION_11",
            MaxBitrate = 123,
            MinIInterval = 123,
            NumberBFramesBetweenReferenceFrames = 123,
            ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            ParDenominator = 123,
            ParNumerator = 123,
            QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS",
            RateControlMode = "VBR"|"CBR",
            ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
            SceneChangeDetect = "DISABLED"|"ENABLED",
            SlowPal = "DISABLED"|"ENABLED",
            Softness = 123,
            SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
            Syntax = "DEFAULT"|"D_10",
            Telecine = "NONE"|"SOFT"|"HARD",
            TemporalAdaptiveQuantization = "DISABLED"|"ENABLED"
          ),
          ProresSettings = list(
            CodecProfile = "APPLE_PRORES_422"|"APPLE_PRORES_422_HQ"|"APPLE_PRORES_422_LT"|"APPLE_PRORES_422_PROXY",
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
            ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            ParDenominator = 123,
            ParNumerator = 123,
            ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
            SlowPal = "DISABLED"|"ENABLED",
            Telecine = "NONE"|"HARD"
          ),
          Vc3Settings = list(
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            InterlaceMode = "INTERLACED"|"PROGRESSIVE",
            ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
            SlowPal = "DISABLED"|"ENABLED",
            Telecine = "NONE"|"HARD",
            Vc3Class = "CLASS_145_8BIT"|"CLASS_220_8BIT"|"CLASS_220_10BIT"
          ),
          Vp8Settings = list(
            Bitrate = 123,
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            GopSize = 123.0,
            HrdBufferSize = 123,
            MaxBitrate = 123,
            ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            ParDenominator = 123,
            ParNumerator = 123,
            QualityTuningLevel = "MULTI_PASS"|"MULTI_PASS_HQ",
            RateControlMode = "VBR"
          ),
          Vp9Settings = list(
            Bitrate = 123,
            FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
            FramerateDenominator = 123,
            FramerateNumerator = 123,
            GopSize = 123.0,
            HrdBufferSize = 123,
            MaxBitrate = 123,
            ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
            ParDenominator = 123,
            ParNumerator = 123,
            QualityTuningLevel = "MULTI_PASS"|"MULTI_PASS_HQ",
            RateControlMode = "VBR"
          )
        ),
        ColorMetadata = "IGNORE"|"INSERT",
        Crop = list(
          Height = 123,
          Width = 123,
          X = 123,
          Y = 123
        ),
        DropFrameTimecode = "DISABLED"|"ENABLED",
        FixedAfd = 123,
        Height = 123,
        Position = list(
          Height = 123,
          Width = 123,
          X = 123,
          Y = 123
        ),
        RespondToAfd = "NONE"|"RESPOND"|"PASSTHROUGH",
        ScalingBehavior = "DEFAULT"|"STRETCH_TO_OUTPUT",
        Sharpness = 123,
        TimecodeInsertion = "DISABLED"|"PIC_TIMING_SEI",
        VideoPreprocessors = list(
          ColorCorrector = list(
            Brightness = 123,
            ColorSpaceConversion = "NONE"|"FORCE_601"|"FORCE_709"|"FORCE_HDR10"|"FORCE_HLG_2020",
            Contrast = 123,
            Hdr10Metadata = list(
              BluePrimaryX = 123,
              BluePrimaryY = 123,
              GreenPrimaryX = 123,
              GreenPrimaryY = 123,
              MaxContentLightLevel = 123,
              MaxFrameAverageLightLevel = 123,
              MaxLuminance = 123,
              MinLuminance = 123,
              RedPrimaryX = 123,
              RedPrimaryY = 123,
              WhitePointX = 123,
              WhitePointY = 123
            ),
            Hue = 123,
            Saturation = 123
          ),
          Deinterlacer = list(
            Algorithm = "INTERPOLATE"|"INTERPOLATE_TICKER"|"BLEND"|"BLEND_TICKER",
            Control = "FORCE_ALL_FRAMES"|"NORMAL",
            Mode = "DEINTERLACE"|"INVERSE_TELECINE"|"ADAPTIVE"
          ),
          DolbyVision = list(
            L6Metadata = list(
              MaxCll = 123,
              MaxFall = 123
            ),
            L6Mode = "PASSTHROUGH"|"RECALCULATE"|"SPECIFY",
            Profile = "PROFILE_5"
          ),
          ImageInserter = list(
            InsertableImages = list(
              list(
                Duration = 123,
                FadeIn = 123,
                FadeOut = 123,
                Height = 123,
                ImageInserterInput = "string",
                ImageX = 123,
                ImageY = 123,
                Layer = 123,
                Opacity = 123,
                StartTime = "string",
                Width = 123
              )
            )
          ),
          NoiseReducer = list(
            Filter = "BILATERAL"|"MEAN"|"GAUSSIAN"|"LANCZOS"|"SHARPEN"|"CONSERVE"|"SPATIAL"|"TEMPORAL",
            FilterSettings = list(
              Strength = 123
            ),
            SpatialFilterSettings = list(
              PostFilterSharpenStrength = 123,
              Speed = 123,
              Strength = 123
            ),
            TemporalFilterSettings = list(
              AggressiveMode = 123,
              PostTemporalSharpening = "DISABLED"|"ENABLED"|"AUTO",
              Speed = 123,
              Strength = 123
            )
          ),
          PartnerWatermarking = list(
            NexguardFileMarkerSettings = list(
              License = "string",
              Payload = 123,
              Preset = "string",
              Strength = "LIGHTEST"|"LIGHTER"|"DEFAULT"|"STRONGER"|"STRONGEST"
            )
          ),
          TimecodeBurnin = list(
            FontSize = 123,
            Position = "TOP_CENTER"|"TOP_LEFT"|"TOP_RIGHT"|"MIDDLE_LEFT"|"MIDDLE_CENTER"|"MIDDLE_RIGHT"|"BOTTOM_LEFT"|"BOTTOM_CENTER"|"BOTTOM_RIGHT",
            Prefix = "string"
          )
        ),
        Width = 123
      )
    ),
    Type = "SYSTEM"|"CUSTOM"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Modify one of your existing presets.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_preset(
  Category = "string",
  Description = "string",
  Name = "string",
  Settings = list(
    AudioDescriptions = list(
      list(
        AudioChannelTaggingSettings = list(
          ChannelTag = "L"|"R"|"C"|"LFE"|"LS"|"RS"|"LC"|"RC"|"CS"|"LSD"|"RSD"|"TCS"|"VHL"|"VHC"|"VHR"
        ),
        AudioNormalizationSettings = list(
          Algorithm = "ITU_BS_1770_1"|"ITU_BS_1770_2"|"ITU_BS_1770_3"|"ITU_BS_1770_4",
          AlgorithmControl = "CORRECT_AUDIO"|"MEASURE_ONLY",
          CorrectionGateLevel = 123,
          LoudnessLogging = "LOG"|"DONT_LOG",
          PeakCalculation = "TRUE_PEAK"|"NONE",
          TargetLkfs = 123.0
        ),
        AudioSourceName = "string",
        AudioType = 123,
        AudioTypeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
        CodecSettings = list(
          AacSettings = list(
            AudioDescriptionBroadcasterMix = "BROADCASTER_MIXED_AD"|"NORMAL",
            Bitrate = 123,
            CodecProfile = "LC"|"HEV1"|"HEV2",
            CodingMode = "AD_RECEIVER_MIX"|"CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_5_1",
            RateControlMode = "CBR"|"VBR",
            RawFormat = "LATM_LOAS"|"NONE",
            SampleRate = 123,
            Specification = "MPEG2"|"MPEG4",
            VbrQuality = "LOW"|"MEDIUM_LOW"|"MEDIUM_HIGH"|"HIGH"
          ),
          Ac3Settings = list(
            Bitrate = 123,
            BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"DIALOGUE"|"EMERGENCY"|"HEARING_IMPAIRED"|"MUSIC_AND_EFFECTS"|"VISUALLY_IMPAIRED"|"VOICE_OVER",
            CodingMode = "CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_3_2_LFE",
            Dialnorm = 123,
            DynamicRangeCompressionProfile = "FILM_STANDARD"|"NONE",
            LfeFilter = "ENABLED"|"DISABLED",
            MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
            SampleRate = 123
          ),
          AiffSettings = list(
            BitDepth = 123,
            Channels = 123,
            SampleRate = 123
          ),
          Codec = "AAC"|"MP2"|"MP3"|"WAV"|"AIFF"|"AC3"|"EAC3"|"EAC3_ATMOS"|"VORBIS"|"OPUS"|"PASSTHROUGH",
          Eac3AtmosSettings = list(
            Bitrate = 123,
            BitstreamMode = "COMPLETE_MAIN",
            CodingMode = "CODING_MODE_9_1_6",
            DialogueIntelligence = "ENABLED"|"DISABLED",
            DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
            DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
            LoRoCenterMixLevel = 123.0,
            LoRoSurroundMixLevel = 123.0,
            LtRtCenterMixLevel = 123.0,
            LtRtSurroundMixLevel = 123.0,
            MeteringMode = "LEQ_A"|"ITU_BS_1770_1"|"ITU_BS_1770_2"|"ITU_BS_1770_3"|"ITU_BS_1770_4",
            SampleRate = 123,
            SpeechThreshold = 123,
            StereoDownmix = "NOT_INDICATED"|"STEREO"|"SURROUND"|"DPL2",
            SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
          ),
          Eac3Settings = list(
            AttenuationControl = "ATTENUATE_3_DB"|"NONE",
            Bitrate = 123,
            BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"EMERGENCY"|"HEARING_IMPAIRED"|"VISUALLY_IMPAIRED",
            CodingMode = "CODING_MODE_1_0"|"CODING_MODE_2_0"|"CODING_MODE_3_2",
            DcFilter = "ENABLED"|"DISABLED",
            Dialnorm = 123,
            DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
            DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
            LfeControl = "LFE"|"NO_LFE",
            LfeFilter = "ENABLED"|"DISABLED",
            LoRoCenterMixLevel = 123.0,
            LoRoSurroundMixLevel = 123.0,
            LtRtCenterMixLevel = 123.0,
            LtRtSurroundMixLevel = 123.0,
            MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
            PassthroughControl = "WHEN_POSSIBLE"|"NO_PASSTHROUGH",
            PhaseControl = "SHIFT_90_DEGREES"|"NO_SHIFT",
            SampleRate = 123,
            StereoDownmix = "NOT_INDICATED"|"LO_RO"|"LT_RT"|"DPL2",
            SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED",
            SurroundMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
          ),
          Mp2Settings = list(
            Bitrate = 123,
            Channels = 123,
            SampleRate = 123
          ),
          Mp3Settings = list(
            Bitrate = 123,
            Channels = 123,
            RateControlMode = "CBR"|"VBR",
            SampleRate = 123,
            VbrQuality = 123
          ),
          OpusSettings = list(
            Bitrate = 123,
            Channels = 123,
            SampleRate = 123
          ),
          VorbisSettings = list(
            Channels = 123,
            SampleRate = 123,
            VbrQuality = 123
          ),
          WavSettings = list(
            BitDepth = 123,
            Channels = 123,
            Format = "RIFF"|"RF64",
            SampleRate = 123
          )
        ),
        CustomLanguageCode = "string",
        LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
        LanguageCodeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
        RemixSettings = list(
          ChannelMapping = list(
            OutputChannels = list(
              list(
                InputChannels = list(
                  123
                ),
                InputChannelsFineTune = list(
                  123.0
                )
              )
            )
          ),
          ChannelsIn = 123,
          ChannelsOut = 123
        ),
        StreamName = "string"
      )
    ),
    CaptionDescriptions = list(
      list(
        CustomLanguageCode = "string",
        DestinationSettings = list(
          BurninDestinationSettings = list(
            Alignment = "CENTERED"|"LEFT",
            BackgroundColor = "NONE"|"BLACK"|"WHITE",
            BackgroundOpacity = 123,
            FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            FontOpacity = 123,
            FontResolution = 123,
            FontScript = "AUTOMATIC"|"HANS"|"HANT",
            FontSize = 123,
            OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            OutlineSize = 123,
            ShadowColor = "NONE"|"BLACK"|"WHITE",
            ShadowOpacity = 123,
            ShadowXOffset = 123,
            ShadowYOffset = 123,
            TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
            XPosition = 123,
            YPosition = 123
          ),
          DestinationType = "BURN_IN"|"DVB_SUB"|"EMBEDDED"|"EMBEDDED_PLUS_SCTE20"|"IMSC"|"SCTE20_PLUS_EMBEDDED"|"SCC"|"SRT"|"SMI"|"TELETEXT"|"TTML"|"WEBVTT",
          DvbSubDestinationSettings = list(
            Alignment = "CENTERED"|"LEFT",
            BackgroundColor = "NONE"|"BLACK"|"WHITE",
            BackgroundOpacity = 123,
            FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            FontOpacity = 123,
            FontResolution = 123,
            FontScript = "AUTOMATIC"|"HANS"|"HANT",
            FontSize = 123,
            OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            OutlineSize = 123,
            ShadowColor = "NONE"|"BLACK"|"WHITE",
            ShadowOpacity = 123,
            ShadowXOffset = 123,
            ShadowYOffset = 123,
            SubtitlingType = "HEARING_IMPAIRED"|"STANDARD",
            TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
            XPosition = 123,
            YPosition = 123
          ),
          EmbeddedDestinationSettings = list(
            Destination608ChannelNumber = 123,
            Destination708ServiceNumber = 123
          ),
          ImscDestinationSettings = list(
            StylePassthrough = "ENABLED"|"DISABLED"
          ),
          SccDestinationSettings = list(
            Framerate = "FRAMERATE_23_97"|"FRAMERATE_24"|"FRAMERATE_25"|"FRAMERATE_29_97_DROPFRAME"|"FRAMERATE_29_97_NON_DROPFRAME"
          ),
          TeletextDestinationSettings = list(
            PageNumber = "string",
            PageTypes = list(
              "PAGE_TYPE_INITIAL"|"PAGE_TYPE_SUBTITLE"|"PAGE_TYPE_ADDL_INFO"|"PAGE_TYPE_PROGRAM_SCHEDULE"|"PAGE_TYPE_HEARING_IMPAIRED_SUBTITLE"
            )
          ),
          TtmlDestinationSettings = list(
            StylePassthrough = "ENABLED"|"DISABLED"
          )
        ),
        LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
        LanguageDescription = "string"
      )
    ),
    ContainerSettings = list(
      CmfcSettings = list(
        AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
        IFrameOnlyManifest = "INCLUDE"|"EXCLUDE",
        Scte35Esam = "INSERT"|"NONE",
        Scte35Source = "PASSTHROUGH"|"NONE"
      ),
      Container = "F4V"|"ISMV"|"M2TS"|"M3U8"|"CMFC"|"MOV"|"MP4"|"MPD"|"MXF"|"WEBM"|"RAW",
      F4vSettings = list(
        MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL"
      ),
      M2tsSettings = list(
        AudioBufferModel = "DVB"|"ATSC",
        AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
        AudioFramesPerPes = 123,
        AudioPids = list(
          123
        ),
        Bitrate = 123,
        BufferModel = "MULTIPLEX"|"NONE",
        DvbNitSettings = list(
          NetworkId = 123,
          NetworkName = "string",
          NitInterval = 123
        ),
        DvbSdtSettings = list(
          OutputSdt = "SDT_FOLLOW"|"SDT_FOLLOW_IF_PRESENT"|"SDT_MANUAL"|"SDT_NONE",
          SdtInterval = 123,
          ServiceName = "string",
          ServiceProviderName = "string"
        ),
        DvbSubPids = list(
          123
        ),
        DvbTdtSettings = list(
          TdtInterval = 123
        ),
        DvbTeletextPid = 123,
        EbpAudioInterval = "VIDEO_AND_FIXED_INTERVALS"|"VIDEO_INTERVAL",
        EbpPlacement = "VIDEO_AND_AUDIO_PIDS"|"VIDEO_PID",
        EsRateInPes = "INCLUDE"|"EXCLUDE",
        ForceTsVideoEbpOrder = "FORCE"|"DEFAULT",
        FragmentTime = 123.0,
        MaxPcrInterval = 123,
        MinEbpInterval = 123,
        NielsenId3 = "INSERT"|"NONE",
        NullPacketBitrate = 123.0,
        PatInterval = 123,
        PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
        PcrPid = 123,
        PmtInterval = 123,
        PmtPid = 123,
        PrivateMetadataPid = 123,
        ProgramNumber = 123,
        RateMode = "VBR"|"CBR",
        Scte35Esam = list(
          Scte35EsamPid = 123
        ),
        Scte35Pid = 123,
        Scte35Source = "PASSTHROUGH"|"NONE",
        SegmentationMarkers = "NONE"|"RAI_SEGSTART"|"RAI_ADAPT"|"PSI_SEGSTART"|"EBP"|"EBP_LEGACY",
        SegmentationStyle = "MAINTAIN_CADENCE"|"RESET_CADENCE",
        SegmentationTime = 123.0,
        TimedMetadataPid = 123,
        TransportStreamId = 123,
        VideoPid = 123
      ),
      M3u8Settings = list(
        AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
        AudioFramesPerPes = 123,
        AudioPids = list(
          123
        ),
        NielsenId3 = "INSERT"|"NONE",
        PatInterval = 123,
        PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
        PcrPid = 123,
        PmtInterval = 123,
        PmtPid = 123,
        PrivateMetadataPid = 123,
        ProgramNumber = 123,
        Scte35Pid = 123,
        Scte35Source = "PASSTHROUGH"|"NONE",
        TimedMetadata = "PASSTHROUGH"|"NONE",
        TimedMetadataPid = 123,
        TransportStreamId = 123,
        VideoPid = 123
      ),
      MovSettings = list(
        ClapAtom = "INCLUDE"|"EXCLUDE",
        CslgAtom = "INCLUDE"|"EXCLUDE",
        Mpeg2FourCCControl = "XDCAM"|"MPEG",
        PaddingControl = "OMNEON"|"NONE",
        Reference = "SELF_CONTAINED"|"EXTERNAL"
      ),
      Mp4Settings = list(
        AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
        CslgAtom = "INCLUDE"|"EXCLUDE",
        CttsVersion = 123,
        FreeSpaceBox = "INCLUDE"|"EXCLUDE",
        MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL",
        Mp4MajorBrand = "string"
      ),
      MpdSettings = list(
        AccessibilityCaptionHints = "INCLUDE"|"EXCLUDE",
        AudioDuration = "DEFAULT_CODEC_DURATION"|"MATCH_VIDEO_DURATION",
        CaptionContainerType = "RAW"|"FRAGMENTED_MP4",
        Scte35Esam = "INSERT"|"NONE",
        Scte35Source = "PASSTHROUGH"|"NONE"
      ),
      MxfSettings = list(
        AfdSignaling = "NO_COPY"|"COPY_FROM_VIDEO",
        Profile = "D_10"|"XDCAM"|"OP1A"
      )
    ),
    VideoDescription = list(
      AfdSignaling = "NONE"|"AUTO"|"FIXED",
      AntiAlias = "DISABLED"|"ENABLED",
      CodecSettings = list(
        Av1Settings = list(
          AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopSize = 123.0,
          MaxBitrate = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          QvbrSettings = list(
            QvbrQualityLevel = 123,
            QvbrQualityLevelFineTune = 123.0
          ),
          RateControlMode = "QVBR",
          Slices = 123,
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED"
        ),
        AvcIntraSettings = list(
          AvcIntraClass = "CLASS_50"|"CLASS_100"|"CLASS_200"|"CLASS_4K_2K",
          AvcIntraUhdSettings = list(
            QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS"
          ),
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
          SlowPal = "DISABLED"|"ENABLED",
          Telecine = "NONE"|"HARD"
        ),
        Codec = "AV1"|"AVC_INTRA"|"FRAME_CAPTURE"|"H_264"|"H_265"|"MPEG2"|"PRORES"|"VC3"|"VP8"|"VP9",
        FrameCaptureSettings = list(
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          MaxCaptures = 123,
          Quality = 123
        ),
        H264Settings = list(
          AdaptiveQuantization = "OFF"|"AUTO"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
          Bitrate = 123,
          CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_1_1"|"LEVEL_1_2"|"LEVEL_1_3"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_2_2"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_3_2"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_4_2"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2",
          CodecProfile = "BASELINE"|"HIGH"|"HIGH_10BIT"|"HIGH_422"|"HIGH_422_10BIT"|"MAIN",
          DynamicSubGop = "ADAPTIVE"|"STATIC",
          EntropyEncoding = "CABAC"|"CAVLC",
          FieldEncoding = "PAFF"|"FORCE_FIELD",
          FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopBReference = "DISABLED"|"ENABLED",
          GopClosedCadence = 123,
          GopSize = 123.0,
          GopSizeUnits = "FRAMES"|"SECONDS",
          HrdBufferInitialFillPercentage = 123,
          HrdBufferSize = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          MaxBitrate = 123,
          MinIInterval = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          NumberReferenceFrames = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
          QvbrSettings = list(
            MaxAverageBitrate = 123,
            QvbrQualityLevel = 123,
            QvbrQualityLevelFineTune = 123.0
          ),
          RateControlMode = "VBR"|"CBR"|"QVBR",
          RepeatPps = "DISABLED"|"ENABLED",
          ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
          SceneChangeDetect = "DISABLED"|"ENABLED"|"TRANSITION_DETECTION",
          Slices = 123,
          SlowPal = "DISABLED"|"ENABLED",
          Softness = 123,
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
          Syntax = "DEFAULT"|"RP2027",
          Telecine = "NONE"|"SOFT"|"HARD",
          TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
          UnregisteredSeiTimecode = "DISABLED"|"ENABLED"
        ),
        H265Settings = list(
          AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
          AlternateTransferFunctionSei = "DISABLED"|"ENABLED",
          Bitrate = 123,
          CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2"|"LEVEL_6"|"LEVEL_6_1"|"LEVEL_6_2",
          CodecProfile = "MAIN_MAIN"|"MAIN_HIGH"|"MAIN10_MAIN"|"MAIN10_HIGH"|"MAIN_422_8BIT_MAIN"|"MAIN_422_8BIT_HIGH"|"MAIN_422_10BIT_MAIN"|"MAIN_422_10BIT_HIGH",
          DynamicSubGop = "ADAPTIVE"|"STATIC",
          FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopBReference = "DISABLED"|"ENABLED",
          GopClosedCadence = 123,
          GopSize = 123.0,
          GopSizeUnits = "FRAMES"|"SECONDS",
          HrdBufferInitialFillPercentage = 123,
          HrdBufferSize = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          MaxBitrate = 123,
          MinIInterval = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          NumberReferenceFrames = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
          QvbrSettings = list(
            MaxAverageBitrate = 123,
            QvbrQualityLevel = 123,
            QvbrQualityLevelFineTune = 123.0
          ),
          RateControlMode = "VBR"|"CBR"|"QVBR",
          SampleAdaptiveOffsetFilterMode = "DEFAULT"|"ADAPTIVE"|"OFF",
          ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
          SceneChangeDetect = "DISABLED"|"ENABLED"|"TRANSITION_DETECTION",
          Slices = 123,
          SlowPal = "DISABLED"|"ENABLED",
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
          Telecine = "NONE"|"SOFT"|"HARD",
          TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
          TemporalIds = "DISABLED"|"ENABLED",
          Tiles = "DISABLED"|"ENABLED",
          UnregisteredSeiTimecode = "DISABLED"|"ENABLED",
          WriteMp4PackagingType = "HVC1"|"HEV1"
        ),
        Mpeg2Settings = list(
          AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH",
          Bitrate = 123,
          CodecLevel = "AUTO"|"LOW"|"MAIN"|"HIGH1440"|"HIGH",
          CodecProfile = "MAIN"|"PROFILE_422",
          DynamicSubGop = "ADAPTIVE"|"STATIC",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopClosedCadence = 123,
          GopSize = 123.0,
          GopSizeUnits = "FRAMES"|"SECONDS",
          HrdBufferInitialFillPercentage = 123,
          HrdBufferSize = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          IntraDcPrecision = "AUTO"|"INTRA_DC_PRECISION_8"|"INTRA_DC_PRECISION_9"|"INTRA_DC_PRECISION_10"|"INTRA_DC_PRECISION_11",
          MaxBitrate = 123,
          MinIInterval = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS",
          RateControlMode = "VBR"|"CBR",
          ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
          SceneChangeDetect = "DISABLED"|"ENABLED",
          SlowPal = "DISABLED"|"ENABLED",
          Softness = 123,
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
          Syntax = "DEFAULT"|"D_10",
          Telecine = "NONE"|"SOFT"|"HARD",
          TemporalAdaptiveQuantization = "DISABLED"|"ENABLED"
        ),
        ProresSettings = list(
          CodecProfile = "APPLE_PRORES_422"|"APPLE_PRORES_422_HQ"|"APPLE_PRORES_422_LT"|"APPLE_PRORES_422_PROXY",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
          SlowPal = "DISABLED"|"ENABLED",
          Telecine = "NONE"|"HARD"
        ),
        Vc3Settings = list(
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          InterlaceMode = "INTERLACED"|"PROGRESSIVE",
          ScanTypeConversionMode = "INTERLACED"|"INTERLACED_OPTIMIZE",
          SlowPal = "DISABLED"|"ENABLED",
          Telecine = "NONE"|"HARD",
          Vc3Class = "CLASS_145_8BIT"|"CLASS_220_8BIT"|"CLASS_220_10BIT"
        ),
        Vp8Settings = list(
          Bitrate = 123,
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopSize = 123.0,
          HrdBufferSize = 123,
          MaxBitrate = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "MULTI_PASS"|"MULTI_PASS_HQ",
          RateControlMode = "VBR"
        ),
        Vp9Settings = list(
          Bitrate = 123,
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE"|"FRAMEFORMER",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopSize = 123.0,
          HrdBufferSize = 123,
          MaxBitrate = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "MULTI_PASS"|"MULTI_PASS_HQ",
          RateControlMode = "VBR"
        )
      ),
      ColorMetadata = "IGNORE"|"INSERT",
      Crop = list(
        Height = 123,
        Width = 123,
        X = 123,
        Y = 123
      ),
      DropFrameTimecode = "DISABLED"|"ENABLED",
      FixedAfd = 123,
      Height = 123,
      Position = list(
        Height = 123,
        Width = 123,
        X = 123,
        Y = 123
      ),
      RespondToAfd = "NONE"|"RESPOND"|"PASSTHROUGH",
      ScalingBehavior = "DEFAULT"|"STRETCH_TO_OUTPUT",
      Sharpness = 123,
      TimecodeInsertion = "DISABLED"|"PIC_TIMING_SEI",
      VideoPreprocessors = list(
        ColorCorrector = list(
          Brightness = 123,
          ColorSpaceConversion = "NONE"|"FORCE_601"|"FORCE_709"|"FORCE_HDR10"|"FORCE_HLG_2020",
          Contrast = 123,
          Hdr10Metadata = list(
            BluePrimaryX = 123,
            BluePrimaryY = 123,
            GreenPrimaryX = 123,
            GreenPrimaryY = 123,
            MaxContentLightLevel = 123,
            MaxFrameAverageLightLevel = 123,
            MaxLuminance = 123,
            MinLuminance = 123,
            RedPrimaryX = 123,
            RedPrimaryY = 123,
            WhitePointX = 123,
            WhitePointY = 123
          ),
          Hue = 123,
          Saturation = 123
        ),
        Deinterlacer = list(
          Algorithm = "INTERPOLATE"|"INTERPOLATE_TICKER"|"BLEND"|"BLEND_TICKER",
          Control = "FORCE_ALL_FRAMES"|"NORMAL",
          Mode = "DEINTERLACE"|"INVERSE_TELECINE"|"ADAPTIVE"
        ),
        DolbyVision = list(
          L6Metadata = list(
            MaxCll = 123,
            MaxFall = 123
          ),
          L6Mode = "PASSTHROUGH"|"RECALCULATE"|"SPECIFY",
          Profile = "PROFILE_5"
        ),
        ImageInserter = list(
          InsertableImages = list(
            list(
              Duration = 123,
              FadeIn = 123,
              FadeOut = 123,
              Height = 123,
              ImageInserterInput = "string",
              ImageX = 123,
              ImageY = 123,
              Layer = 123,
              Opacity = 123,
              StartTime = "string",
              Width = 123
            )
          )
        ),
        NoiseReducer = list(
          Filter = "BILATERAL"|"MEAN"|"GAUSSIAN"|"LANCZOS"|"SHARPEN"|"CONSERVE"|"SPATIAL"|"TEMPORAL",
          FilterSettings = list(
            Strength = 123
          ),
          SpatialFilterSettings = list(
            PostFilterSharpenStrength = 123,
            Speed = 123,
            Strength = 123
          ),
          TemporalFilterSettings = list(
            AggressiveMode = 123,
            PostTemporalSharpening = "DISABLED"|"ENABLED"|"AUTO",
            Speed = 123,
            Strength = 123
          )
        ),
        PartnerWatermarking = list(
          NexguardFileMarkerSettings = list(
            License = "string",
            Payload = 123,
            Preset = "string",
            Strength = "LIGHTEST"|"LIGHTER"|"DEFAULT"|"STRONGER"|"STRONGEST"
          )
        ),
        TimecodeBurnin = list(
          FontSize = 123,
          Position = "TOP_CENTER"|"TOP_LEFT"|"TOP_RIGHT"|"MIDDLE_LEFT"|"MIDDLE_CENTER"|"MIDDLE_RIGHT"|"BOTTOM_LEFT"|"BOTTOM_CENTER"|"BOTTOM_RIGHT",
          Prefix = "string"
        )
      ),
      Width = 123
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
