% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_service.R
\name{medialive}
\alias{medialive}
\title{AWS Elemental MediaLive}
\usage{
medialive(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
API for AWS Elemental MediaLive
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- medialive(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=medialive_accept_input_device_transfer]{accept_input_device_transfer} \tab Accept an incoming input device transfer\cr
\link[=medialive_batch_delete]{batch_delete} \tab Starts delete of resources\cr
\link[=medialive_batch_start]{batch_start} \tab Starts existing resources\cr
\link[=medialive_batch_stop]{batch_stop} \tab Stops running resources\cr
\link[=medialive_batch_update_schedule]{batch_update_schedule} \tab Update a channel schedule\cr
\link[=medialive_cancel_input_device_transfer]{cancel_input_device_transfer} \tab Cancel an input device transfer that you have requested\cr
\link[=medialive_create_channel]{create_channel} \tab Creates a new channel\cr
\link[=medialive_create_input]{create_input} \tab Create an input\cr
\link[=medialive_create_input_security_group]{create_input_security_group} \tab Creates a Input Security Group\cr
\link[=medialive_create_multiplex]{create_multiplex} \tab Create a new multiplex\cr
\link[=medialive_create_multiplex_program]{create_multiplex_program} \tab Create a new program in the multiplex\cr
\link[=medialive_create_tags]{create_tags} \tab Create tags for a resource\cr
\link[=medialive_delete_channel]{delete_channel} \tab Starts deletion of channel\cr
\link[=medialive_delete_input]{delete_input} \tab Deletes the input end point\cr
\link[=medialive_delete_input_security_group]{delete_input_security_group} \tab Deletes an Input Security Group\cr
\link[=medialive_delete_multiplex]{delete_multiplex} \tab Delete a multiplex\cr
\link[=medialive_delete_multiplex_program]{delete_multiplex_program} \tab Delete a program from a multiplex\cr
\link[=medialive_delete_reservation]{delete_reservation} \tab Delete an expired reservation\cr
\link[=medialive_delete_schedule]{delete_schedule} \tab Delete all schedule actions on a channel\cr
\link[=medialive_delete_tags]{delete_tags} \tab Removes tags for a resource\cr
\link[=medialive_describe_channel]{describe_channel} \tab Gets details about a channel\cr
\link[=medialive_describe_input]{describe_input} \tab Produces details about an input\cr
\link[=medialive_describe_input_device]{describe_input_device} \tab Gets the details for the input device\cr
\link[=medialive_describe_input_device_thumbnail]{describe_input_device_thumbnail} \tab Get the latest thumbnail data for the input device\cr
\link[=medialive_describe_input_security_group]{describe_input_security_group} \tab Produces a summary of an Input Security Group\cr
\link[=medialive_describe_multiplex]{describe_multiplex} \tab Gets details about a multiplex\cr
\link[=medialive_describe_multiplex_program]{describe_multiplex_program} \tab Get the details for a program in a multiplex\cr
\link[=medialive_describe_offering]{describe_offering} \tab Get details for an offering\cr
\link[=medialive_describe_reservation]{describe_reservation} \tab Get details for a reservation\cr
\link[=medialive_describe_schedule]{describe_schedule} \tab Get a channel schedule\cr
\link[=medialive_list_channels]{list_channels} \tab Produces list of channels that have been created\cr
\link[=medialive_list_input_devices]{list_input_devices} \tab List input devices\cr
\link[=medialive_list_input_device_transfers]{list_input_device_transfers} \tab List input devices that are currently being transferred\cr
\link[=medialive_list_inputs]{list_inputs} \tab Produces list of inputs that have been created\cr
\link[=medialive_list_input_security_groups]{list_input_security_groups} \tab Produces a list of Input Security Groups for an account\cr
\link[=medialive_list_multiplexes]{list_multiplexes} \tab Retrieve a list of the existing multiplexes\cr
\link[=medialive_list_multiplex_programs]{list_multiplex_programs} \tab List the programs that currently exist for a specific multiplex\cr
\link[=medialive_list_offerings]{list_offerings} \tab List offerings available for purchase\cr
\link[=medialive_list_reservations]{list_reservations} \tab List purchased reservations\cr
\link[=medialive_list_tags_for_resource]{list_tags_for_resource} \tab Produces list of tags that have been created for a resource\cr
\link[=medialive_purchase_offering]{purchase_offering} \tab Purchase an offering and create a reservation\cr
\link[=medialive_reject_input_device_transfer]{reject_input_device_transfer} \tab Reject the transfer of the specified input device to your AWS account\cr
\link[=medialive_start_channel]{start_channel} \tab Starts an existing channel\cr
\link[=medialive_start_multiplex]{start_multiplex} \tab Start (run) the multiplex\cr
\link[=medialive_stop_channel]{stop_channel} \tab Stops a running channel\cr
\link[=medialive_stop_multiplex]{stop_multiplex} \tab Stops a running multiplex\cr
\link[=medialive_transfer_input_device]{transfer_input_device} \tab Start an input device transfer to another AWS account\cr
\link[=medialive_update_channel]{update_channel} \tab Updates a channel\cr
\link[=medialive_update_channel_class]{update_channel_class} \tab Changes the class of the channel\cr
\link[=medialive_update_input]{update_input} \tab Updates an input\cr
\link[=medialive_update_input_device]{update_input_device} \tab Updates the parameters for the input device\cr
\link[=medialive_update_input_security_group]{update_input_security_group} \tab Update an Input Security Group's Whilelists\cr
\link[=medialive_update_multiplex]{update_multiplex} \tab Updates a multiplex\cr
\link[=medialive_update_multiplex_program]{update_multiplex_program} \tab Update a program in a multiplex\cr
\link[=medialive_update_reservation]{update_reservation} \tab Update reservation
}
}

\examples{
\dontrun{
svc <- medialive()
svc$accept_input_device_transfer(
  Foo = 123
)
}

}
