% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_batch_start}
\alias{medialive_batch_start}
\title{Starts existing resources}
\usage{
medialive_batch_start(ChannelIds, MultiplexIds)
}
\arguments{
\item{ChannelIds}{List of channel IDs}

\item{MultiplexIds}{List of multiplex IDs}
}
\value{
A list with the following syntax:\preformatted{list(
  Failed = list(
    list(
      Arn = "string",
      Code = "string",
      Id = "string",
      Message = "string"
    )
  ),
  Successful = list(
    list(
      Arn = "string",
      Id = "string",
      State = "string"
    )
  )
)
}
}
\description{
Starts existing resources
}
\section{Request syntax}{
\preformatted{svc$batch_start(
  ChannelIds = list(
    "string"
  ),
  MultiplexIds = list(
    "string"
  )
)
}
}

\keyword{internal}
