% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_batch_stop}
\alias{medialive_batch_stop}
\title{Stops running resources}
\usage{
medialive_batch_stop(ChannelIds, MultiplexIds)
}
\arguments{
\item{ChannelIds}{List of channel IDs}

\item{MultiplexIds}{List of multiplex IDs}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Failed = list(
    list(
      Arn = "string",
      Code = "string",
      Id = "string",
      Message = "string"
    )
  ),
  Successful = list(
    list(
      Arn = "string",
      Id = "string",
      State = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Stops running resources
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$batch_stop(
  ChannelIds = list(
    "string"
  ),
  MultiplexIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
