% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_multiplex}
\alias{medialive_create_multiplex}
\title{Create a new multiplex}
\usage{
medialive_create_multiplex(AvailabilityZones, MultiplexSettings, Name,
  RequestId, Tags)
}
\arguments{
\item{AvailabilityZones}{[required] A list of availability zones for the multiplex. You must specify exactly
two.}

\item{MultiplexSettings}{[required] Configuration for a multiplex event.}

\item{Name}{[required] Name of multiplex.}

\item{RequestId}{[required] Unique request ID. This prevents retries from creating multiple
resources.}

\item{Tags}{A collection of key-value pairs.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Multiplex = list(
    Arn = "string",
    AvailabilityZones = list(
      "string"
    ),
    Destinations = list(
      list(
        MediaConnectSettings = list(
          EntitlementArn = "string"
        )
      )
    ),
    Id = "string",
    MultiplexSettings = list(
      MaximumVideoBufferDelayMilliseconds = 123,
      TransportStreamBitrate = 123,
      TransportStreamId = 123,
      TransportStreamReservedBitrate = 123
    ),
    Name = "string",
    PipelinesRunningCount = 123,
    ProgramCount = 123,
    State = "CREATING"|"CREATE_FAILED"|"IDLE"|"STARTING"|"RUNNING"|"RECOVERING"|"STOPPING"|"DELETING"|"DELETED",
    Tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Create a new multiplex.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_multiplex(
  AvailabilityZones = list(
    "string"
  ),
  MultiplexSettings = list(
    MaximumVideoBufferDelayMilliseconds = 123,
    TransportStreamBitrate = 123,
    TransportStreamId = 123,
    TransportStreamReservedBitrate = 123
  ),
  Name = "string",
  RequestId = "string",
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
