% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_multiplex_program}
\alias{medialive_create_multiplex_program}
\title{Create a new program in the multiplex}
\usage{
medialive_create_multiplex_program(MultiplexId,
  MultiplexProgramSettings, ProgramName, RequestId)
}
\arguments{
\item{MultiplexId}{[required] ID of the multiplex where the program is to be created.}

\item{MultiplexProgramSettings}{[required] The settings for this multiplex program.}

\item{ProgramName}{[required] Name of multiplex program.}

\item{RequestId}{[required] Unique request ID. This prevents retries from creating multiple
resources.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  MultiplexProgram = list(
    ChannelId = "string",
    MultiplexProgramSettings = list(
      PreferredChannelPipeline = "CURRENTLY_ACTIVE"|"PIPELINE_0"|"PIPELINE_1",
      ProgramNumber = 123,
      ServiceDescriptor = list(
        ProviderName = "string",
        ServiceName = "string"
      ),
      VideoSettings = list(
        ConstantBitrate = 123,
        StatmuxSettings = list(
          MaximumBitrate = 123,
          MinimumBitrate = 123,
          Priority = 123
        )
      )
    ),
    PacketIdentifiersMap = list(
      AudioPids = list(
        123
      ),
      DvbSubPids = list(
        123
      ),
      DvbTeletextPid = 123,
      EtvPlatformPid = 123,
      EtvSignalPid = 123,
      KlvDataPids = list(
        123
      ),
      PcrPid = 123,
      PmtPid = 123,
      PrivateMetadataPid = 123,
      Scte27Pids = list(
        123
      ),
      Scte35Pid = 123,
      TimedMetadataPid = 123,
      VideoPid = 123
    ),
    PipelineDetails = list(
      list(
        ActiveChannelPipeline = "string",
        PipelineId = "string"
      )
    ),
    ProgramName = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Create a new program in the multiplex.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_multiplex_program(
  MultiplexId = "string",
  MultiplexProgramSettings = list(
    PreferredChannelPipeline = "CURRENTLY_ACTIVE"|"PIPELINE_0"|"PIPELINE_1",
    ProgramNumber = 123,
    ServiceDescriptor = list(
      ProviderName = "string",
      ServiceName = "string"
    ),
    VideoSettings = list(
      ConstantBitrate = 123,
      StatmuxSettings = list(
        MaximumBitrate = 123,
        MinimumBitrate = 123,
        Priority = 123
      )
    )
  ),
  ProgramName = "string",
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
