% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_partner_input}
\alias{medialive_create_partner_input}
\title{Create a partner input}
\usage{
medialive_create_partner_input(InputId, RequestId, Tags)
}
\arguments{
\item{InputId}{[required] Unique ID of the input.}

\item{RequestId}{Unique identifier of the request to ensure the request is handled
exactly once in case of retries.}

\item{Tags}{A collection of key-value pairs.}
}
\value{
A list with the following syntax:\preformatted{list(
  Input = list(
    Arn = "string",
    AttachedChannels = list(
      "string"
    ),
    Destinations = list(
      list(
        Ip = "string",
        Port = "string",
        Url = "string",
        Vpc = list(
          AvailabilityZone = "string",
          NetworkInterfaceId = "string"
        )
      )
    ),
    Id = "string",
    InputClass = "STANDARD"|"SINGLE_PIPELINE",
    InputDevices = list(
      list(
        Id = "string"
      )
    ),
    InputPartnerIds = list(
      "string"
    ),
    InputSourceType = "STATIC"|"DYNAMIC",
    MediaConnectFlows = list(
      list(
        FlowArn = "string"
      )
    ),
    Name = "string",
    RoleArn = "string",
    SecurityGroups = list(
      "string"
    ),
    Sources = list(
      list(
        PasswordParam = "string",
        Url = "string",
        Username = "string"
      )
    ),
    State = "CREATING"|"DETACHED"|"ATTACHED"|"DELETING"|"DELETED",
    Tags = list(
      "string"
    ),
    Type = "UDP_PUSH"|"RTP_PUSH"|"RTMP_PUSH"|"RTMP_PULL"|"URL_PULL"|"MP4_FILE"|"MEDIACONNECT"|"INPUT_DEVICE"|"AWS_CDI"
  )
)
}
}
\description{
Create a partner input
}
\section{Request syntax}{
\preformatted{svc$create_partner_input(
  InputId = "string",
  RequestId = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
