% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_delete_input}
\alias{medialive_delete_input}
\title{Deletes the input end point}
\usage{
medialive_delete_input(InputId)
}
\arguments{
\item{InputId}{[required] Unique ID of the input}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the input end point
}
\section{Request syntax}{
\preformatted{svc$delete_input(
  InputId = "string"
)
}
}

\keyword{internal}
