% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_delete_input_security_group}
\alias{medialive_delete_input_security_group}
\title{Deletes an Input Security Group}
\usage{
medialive_delete_input_security_group(InputSecurityGroupId)
}
\arguments{
\item{InputSecurityGroupId}{[required] The Input Security Group to delete}
}
\value{
An empty list.
}
\description{
Deletes an Input Security Group
}
\section{Request syntax}{
\preformatted{svc$delete_input_security_group(
  InputSecurityGroupId = "string"
)
}
}

\keyword{internal}
