% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_delete_schedule}
\alias{medialive_delete_schedule}
\title{Delete all schedule actions on a channel}
\usage{
medialive_delete_schedule(ChannelId)
}
\arguments{
\item{ChannelId}{[required] Id of the channel whose schedule is being deleted.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Delete all schedule actions on a channel.
}
\section{Request syntax}{
\preformatted{svc$delete_schedule(
  ChannelId = "string"
)
}
}

\keyword{internal}
