% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_channel}
\alias{medialive_describe_channel}
\title{Gets details about a channel}
\usage{
medialive_describe_channel(ChannelId)
}
\arguments{
\item{ChannelId}{[required] channel ID}
}
\description{
Gets details about a channel
}
\section{Request syntax}{
\preformatted{svc$describe_channel(
  ChannelId = "string"
)
}
}

\keyword{internal}
