% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_input_device_thumbnail}
\alias{medialive_describe_input_device_thumbnail}
\title{Get the latest thumbnail data for the input device}
\usage{
medialive_describe_input_device_thumbnail(InputDeviceId, Accept)
}
\arguments{
\item{InputDeviceId}{[required] The unique ID of this input device. For example, hd-123456789abcdef.}

\item{Accept}{[required] The HTTP Accept header. Indicates the requested type for the thumbnail.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Body = raw,
  ContentType = "image/jpeg",
  ContentLength = 123,
  ETag = "string",
  LastModified = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Get the latest thumbnail data for the input device.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_input_device_thumbnail(
  InputDeviceId = "string",
  Accept = "image/jpeg"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
