% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_input_security_group}
\alias{medialive_describe_input_security_group}
\title{Produces a summary of an Input Security Group}
\usage{
medialive_describe_input_security_group(InputSecurityGroupId)
}
\arguments{
\item{InputSecurityGroupId}{[required] The id of the Input Security Group to describe}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Id = "string",
  Inputs = list(
    "string"
  ),
  State = "IDLE"|"IN_USE"|"UPDATING"|"DELETED",
  Tags = list(
    "string"
  ),
  WhitelistRules = list(
    list(
      Cidr = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Produces a summary of an Input Security Group
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_input_security_group(
  InputSecurityGroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
