% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_input_device_transfers}
\alias{medialive_list_input_device_transfers}
\title{List input devices that are currently being transferred}
\usage{
medialive_list_input_device_transfers(MaxResults, NextToken,
  TransferType)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}

\item{TransferType}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InputDeviceTransfers = list(
    list(
      Id = "string",
      Message = "string",
      TargetCustomerId = "string",
      TransferType = "OUTGOING"|"INCOMING"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List input devices that are currently being transferred. List input
devices that you are transferring from your AWS account or input devices
that another AWS account is transferring to you.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_input_device_transfers(
  MaxResults = 123,
  NextToken = "string",
  TransferType = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
