% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_input_devices}
\alias{medialive_list_input_devices}
\title{List input devices}
\usage{
medialive_list_input_devices(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\value{
A list with the following syntax:\preformatted{list(
  InputDevices = list(
    list(
      Arn = "string",
      ConnectionState = "DISCONNECTED"|"CONNECTED",
      DeviceSettingsSyncState = "SYNCED"|"SYNCING",
      DeviceUpdateStatus = "UP_TO_DATE"|"NOT_UP_TO_DATE",
      HdDeviceSettings = list(
        ActiveInput = "HDMI"|"SDI",
        ConfiguredInput = "AUTO"|"HDMI"|"SDI",
        DeviceState = "IDLE"|"STREAMING",
        Framerate = 123.0,
        Height = 123,
        MaxBitrate = 123,
        ScanType = "INTERLACED"|"PROGRESSIVE",
        Width = 123
      ),
      Id = "string",
      MacAddress = "string",
      Name = "string",
      NetworkSettings = list(
        DnsAddresses = list(
          "string"
        ),
        Gateway = "string",
        IpAddress = "string",
        IpScheme = "STATIC"|"DHCP",
        SubnetMask = "string"
      ),
      SerialNumber = "string",
      Type = "HD",
      UhdDeviceSettings = list(
        ActiveInput = "HDMI"|"SDI",
        ConfiguredInput = "AUTO"|"HDMI"|"SDI",
        DeviceState = "IDLE"|"STREAMING",
        Framerate = 123.0,
        Height = 123,
        MaxBitrate = 123,
        ScanType = "INTERLACED"|"PROGRESSIVE",
        Width = 123
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
List input devices
}
\section{Request syntax}{
\preformatted{svc$list_input_devices(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
