% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_inputs}
\alias{medialive_list_inputs}
\title{Produces list of inputs that have been created}
\usage{
medialive_list_inputs(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Inputs = list(
    list(
      Arn = "string",
      AttachedChannels = list(
        "string"
      ),
      Destinations = list(
        list(
          Ip = "string",
          Port = "string",
          Url = "string",
          Vpc = list(
            AvailabilityZone = "string",
            NetworkInterfaceId = "string"
          )
        )
      ),
      Id = "string",
      InputClass = "STANDARD"|"SINGLE_PIPELINE",
      InputDevices = list(
        list(
          Id = "string"
        )
      ),
      InputSourceType = "STATIC"|"DYNAMIC",
      MediaConnectFlows = list(
        list(
          FlowArn = "string"
        )
      ),
      Name = "string",
      RoleArn = "string",
      SecurityGroups = list(
        "string"
      ),
      Sources = list(
        list(
          PasswordParam = "string",
          Url = "string",
          Username = "string"
        )
      ),
      State = "CREATING"|"DETACHED"|"ATTACHED"|"DELETING"|"DELETED",
      Tags = list(
        "string"
      ),
      Type = "UDP_PUSH"|"RTP_PUSH"|"RTMP_PUSH"|"RTMP_PULL"|"URL_PULL"|"MP4_FILE"|"MEDIACONNECT"|"INPUT_DEVICE"|"AWS_CDI"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Produces list of inputs that have been created
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_inputs(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
