% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_offerings}
\alias{medialive_list_offerings}
\title{List offerings available for purchase}
\usage{
medialive_list_offerings(ChannelClass, ChannelConfiguration, Codec,
  Duration, MaxResults, MaximumBitrate, MaximumFramerate, NextToken,
  Resolution, ResourceType, SpecialFeature, VideoQuality)
}
\arguments{
\item{ChannelClass}{Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'}

\item{ChannelConfiguration}{Filter to offerings that match the configuration of an existing channel,
e.g. '2345678' (a channel ID)}

\item{Codec}{Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'}

\item{Duration}{Filter by offering duration, e.g. '12'}

\item{MaxResults}{}

\item{MaximumBitrate}{Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'}

\item{MaximumFramerate}{Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'}

\item{NextToken}{}

\item{Resolution}{Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'}

\item{ResourceType}{Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'}

\item{SpecialFeature}{Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'}

\item{VideoQuality}{Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextToken = "string",
  Offerings = list(
    list(
      Arn = "string",
      CurrencyCode = "string",
      Duration = 123,
      DurationUnits = "MONTHS",
      FixedPrice = 123.0,
      OfferingDescription = "string",
      OfferingId = "string",
      OfferingType = "NO_UPFRONT",
      Region = "string",
      ResourceSpecification = list(
        ChannelClass = "STANDARD"|"SINGLE_PIPELINE",
        Codec = "MPEG2"|"AVC"|"HEVC"|"AUDIO"|"LINK",
        MaximumBitrate = "MAX_10_MBPS"|"MAX_20_MBPS"|"MAX_50_MBPS",
        MaximumFramerate = "MAX_30_FPS"|"MAX_60_FPS",
        Resolution = "SD"|"HD"|"FHD"|"UHD",
        ResourceType = "INPUT"|"OUTPUT"|"MULTIPLEX"|"CHANNEL",
        SpecialFeature = "ADVANCED_AUDIO"|"AUDIO_NORMALIZATION",
        VideoQuality = "STANDARD"|"ENHANCED"|"PREMIUM"
      ),
      UsagePrice = 123.0
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
List offerings available for purchase.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_offerings(
  ChannelClass = "string",
  ChannelConfiguration = "string",
  Codec = "string",
  Duration = "string",
  MaxResults = 123,
  MaximumBitrate = "string",
  MaximumFramerate = "string",
  NextToken = "string",
  Resolution = "string",
  ResourceType = "string",
  SpecialFeature = "string",
  VideoQuality = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
