% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_reservations}
\alias{medialive_list_reservations}
\title{List purchased reservations}
\usage{
medialive_list_reservations(ChannelClass, Codec, MaxResults,
  MaximumBitrate, MaximumFramerate, NextToken, Resolution, ResourceType,
  SpecialFeature, VideoQuality)
}
\arguments{
\item{ChannelClass}{Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'}

\item{Codec}{Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'}

\item{MaxResults}{}

\item{MaximumBitrate}{Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'}

\item{MaximumFramerate}{Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'}

\item{NextToken}{}

\item{Resolution}{Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'}

\item{ResourceType}{Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'}

\item{SpecialFeature}{Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'}

\item{VideoQuality}{Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  Reservations = list(
    list(
      Arn = "string",
      Count = 123,
      CurrencyCode = "string",
      Duration = 123,
      DurationUnits = "MONTHS",
      End = "string",
      FixedPrice = 123.0,
      Name = "string",
      OfferingDescription = "string",
      OfferingId = "string",
      OfferingType = "NO_UPFRONT",
      Region = "string",
      ReservationId = "string",
      ResourceSpecification = list(
        ChannelClass = "STANDARD"|"SINGLE_PIPELINE",
        Codec = "MPEG2"|"AVC"|"HEVC"|"AUDIO"|"LINK",
        MaximumBitrate = "MAX_10_MBPS"|"MAX_20_MBPS"|"MAX_50_MBPS",
        MaximumFramerate = "MAX_30_FPS"|"MAX_60_FPS",
        Resolution = "SD"|"HD"|"FHD"|"UHD",
        ResourceType = "INPUT"|"OUTPUT"|"MULTIPLEX"|"CHANNEL",
        SpecialFeature = "ADVANCED_AUDIO"|"AUDIO_NORMALIZATION",
        VideoQuality = "STANDARD"|"ENHANCED"|"PREMIUM"
      ),
      Start = "string",
      State = "ACTIVE"|"EXPIRED"|"CANCELED"|"DELETED",
      Tags = list(
        "string"
      ),
      UsagePrice = 123.0
    )
  )
)
}
}
\description{
List purchased reservations.
}
\section{Request syntax}{
\preformatted{svc$list_reservations(
  ChannelClass = "string",
  Codec = "string",
  MaxResults = 123,
  MaximumBitrate = "string",
  MaximumFramerate = "string",
  NextToken = "string",
  Resolution = "string",
  ResourceType = "string",
  SpecialFeature = "string",
  VideoQuality = "string"
)
}
}

\keyword{internal}
