% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_tags_for_resource}
\alias{medialive_list_tags_for_resource}
\title{Produces list of tags that have been created for a resource}
\usage{
medialive_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Produces list of tags that have been created for a resource
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
