% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_purchase_offering}
\alias{medialive_purchase_offering}
\title{Purchase an offering and create a reservation}
\usage{
medialive_purchase_offering(Count, Name, OfferingId, RequestId, Start,
  Tags)
}
\arguments{
\item{Count}{[required] Number of resources}

\item{Name}{Name for the new reservation}

\item{OfferingId}{[required] Offering to purchase, e.g. '87654321'}

\item{RequestId}{Unique request ID to be specified. This is needed to prevent retries
from creating multiple resources.}

\item{Start}{Requested reservation start time (UTC) in ISO-8601 format. The specified
time must be between the first day of the current month and one year
from now. If no value is given, the default is now.}

\item{Tags}{A collection of key-value pairs}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Reservation = list(
    Arn = "string",
    Count = 123,
    CurrencyCode = "string",
    Duration = 123,
    DurationUnits = "MONTHS",
    End = "string",
    FixedPrice = 123.0,
    Name = "string",
    OfferingDescription = "string",
    OfferingId = "string",
    OfferingType = "NO_UPFRONT",
    Region = "string",
    ReservationId = "string",
    ResourceSpecification = list(
      ChannelClass = "STANDARD"|"SINGLE_PIPELINE",
      Codec = "MPEG2"|"AVC"|"HEVC"|"AUDIO"|"LINK",
      MaximumBitrate = "MAX_10_MBPS"|"MAX_20_MBPS"|"MAX_50_MBPS",
      MaximumFramerate = "MAX_30_FPS"|"MAX_60_FPS",
      Resolution = "SD"|"HD"|"FHD"|"UHD",
      ResourceType = "INPUT"|"OUTPUT"|"MULTIPLEX"|"CHANNEL",
      SpecialFeature = "ADVANCED_AUDIO"|"AUDIO_NORMALIZATION",
      VideoQuality = "STANDARD"|"ENHANCED"|"PREMIUM"
    ),
    Start = "string",
    State = "ACTIVE"|"EXPIRED"|"CANCELED"|"DELETED",
    Tags = list(
      "string"
    ),
    UsagePrice = 123.0
  )
)
}\if{html}{\out{</div>}}
}
\description{
Purchase an offering and create a reservation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$purchase_offering(
  Count = 123,
  Name = "string",
  OfferingId = "string",
  RequestId = "string",
  Start = "string",
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
