% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_reject_input_device_transfer}
\alias{medialive_reject_input_device_transfer}
\title{Reject the transfer of the specified input device to your AWS account}
\usage{
medialive_reject_input_device_transfer(InputDeviceId)
}
\arguments{
\item{InputDeviceId}{[required] The unique ID of the input device to reject. For example,
hd-123456789abcdef.}
}
\value{
An empty list.
}
\description{
Reject the transfer of the specified input device to your AWS account.
}
\section{Request syntax}{
\preformatted{svc$reject_input_device_transfer(
  InputDeviceId = "string"
)
}
}

\keyword{internal}
