% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_start_multiplex}
\alias{medialive_start_multiplex}
\title{Start (run) the multiplex}
\usage{
medialive_start_multiplex(MultiplexId)
}
\arguments{
\item{MultiplexId}{[required] The ID of the multiplex.}
}
\description{
Start (run) the multiplex. Starting the multiplex does not start the channels. You must explicitly start each channel.
}
\section{Request syntax}{
\preformatted{svc$start_multiplex(
  MultiplexId = "string"
)
}
}

\keyword{internal}
