% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_transfer_input_device}
\alias{medialive_transfer_input_device}
\title{Start an input device transfer to another AWS account}
\usage{
medialive_transfer_input_device(InputDeviceId, TargetCustomerId,
  TargetRegion, TransferMessage)
}
\arguments{
\item{InputDeviceId}{[required] The unique ID of this input device. For example, hd-123456789abcdef.}

\item{TargetCustomerId}{The AWS account ID (12 digits) for the recipient of the device transfer.}

\item{TargetRegion}{The target AWS region to transfer the device.}

\item{TransferMessage}{An optional message for the recipient. Maximum 280 characters.}
}
\value{
An empty list.
}
\description{
Start an input device transfer to another AWS account. After you make
the request, the other account must accept or reject the transfer.
}
\section{Request syntax}{
\preformatted{svc$transfer_input_device(
  InputDeviceId = "string",
  TargetCustomerId = "string",
  TargetRegion = "string",
  TransferMessage = "string"
)
}
}

\keyword{internal}
