% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_input_security_group}
\alias{medialive_update_input_security_group}
\title{Update an Input Security Group's Whilelists}
\usage{
medialive_update_input_security_group(InputSecurityGroupId, Tags,
  WhitelistRules)
}
\arguments{
\item{InputSecurityGroupId}{[required] The id of the Input Security Group to update.}

\item{Tags}{A collection of key-value pairs.}

\item{WhitelistRules}{List of IPv4 CIDR addresses to whitelist}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SecurityGroup = list(
    Arn = "string",
    Id = "string",
    Inputs = list(
      "string"
    ),
    State = "IDLE"|"IN_USE"|"UPDATING"|"DELETED",
    Tags = list(
      "string"
    ),
    WhitelistRules = list(
      list(
        Cidr = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Update an Input Security Group's Whilelists.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_input_security_group(
  InputSecurityGroupId = "string",
  Tags = list(
    "string"
  ),
  WhitelistRules = list(
    list(
      Cidr = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
