% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_multiplex}
\alias{medialive_update_multiplex}
\title{Updates a multiplex}
\usage{
medialive_update_multiplex(MultiplexId, MultiplexSettings, Name)
}
\arguments{
\item{MultiplexId}{[required] ID of the multiplex to update.}

\item{MultiplexSettings}{The new settings for a multiplex.}

\item{Name}{Name of the multiplex.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Multiplex = list(
    Arn = "string",
    AvailabilityZones = list(
      "string"
    ),
    Destinations = list(
      list(
        MediaConnectSettings = list(
          EntitlementArn = "string"
        )
      )
    ),
    Id = "string",
    MultiplexSettings = list(
      MaximumVideoBufferDelayMilliseconds = 123,
      TransportStreamBitrate = 123,
      TransportStreamId = 123,
      TransportStreamReservedBitrate = 123
    ),
    Name = "string",
    PipelinesRunningCount = 123,
    ProgramCount = 123,
    State = "CREATING"|"CREATE_FAILED"|"IDLE"|"STARTING"|"RUNNING"|"RECOVERING"|"STOPPING"|"DELETING"|"DELETED",
    Tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates a multiplex.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_multiplex(
  MultiplexId = "string",
  MultiplexSettings = list(
    MaximumVideoBufferDelayMilliseconds = 123,
    TransportStreamBitrate = 123,
    TransportStreamId = 123,
    TransportStreamReservedBitrate = 123
  ),
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
