% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_create_channel}
\alias{mediapackage_create_channel}
\title{Creates a new Channel}
\usage{
mediapackage_create_channel(Description, Id, Tags)
}
\arguments{
\item{Description}{A short text description of the Channel.}

\item{Id}{[required] The ID of the Channel. The ID must be unique within the region and it
cannot be changed after a Channel is created.}

\item{Tags}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Description = "string",
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  HlsIngest = list(
    IngestEndpoints = list(
      list(
        Id = "string",
        Password = "string",
        Url = "string",
        Username = "string"
      )
    )
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new Channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_channel(
  Description = "string",
  Id = "string",
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
