% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_create_origin_endpoint}
\alias{mediapackage_create_origin_endpoint}
\title{Creates a new OriginEndpoint record}
\usage{
mediapackage_create_origin_endpoint(ChannelId, CmafPackage, DashPackage,
  Description, HlsPackage, Id, ManifestName, MssPackage,
  StartoverWindowSeconds, Tags, TimeDelaySeconds, Whitelist)
}
\arguments{
\item{ChannelId}{[required] The ID of the Channel that the OriginEndpoint will be associated with.
This cannot be changed after the OriginEndpoint is created.}

\item{CmafPackage}{}

\item{DashPackage}{}

\item{Description}{A short text description of the OriginEndpoint.}

\item{HlsPackage}{}

\item{Id}{[required] The ID of the OriginEndpoint.  The ID must be unique within the region
and it cannot be changed after the OriginEndpoint is created.}

\item{ManifestName}{A short string that will be used as the filename of the OriginEndpoint URL (defaults to "index").}

\item{MssPackage}{}

\item{StartoverWindowSeconds}{Maximum duration (seconds) of content to retain for startover playback.
If not specified, startover playback will be disabled for the OriginEndpoint.}

\item{Tags}{}

\item{TimeDelaySeconds}{Amount of delay (seconds) to enforce on the playback of live content.
If not specified, there will be no time delay in effect for the OriginEndpoint.}

\item{Whitelist}{A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.}
}
\description{
Creates a new OriginEndpoint record.
}
\section{Request syntax}{
\preformatted{svc$create_origin_endpoint(
  ChannelId = "string",
  CmafPackage = list(
    Encryption = list(
      KeyRotationIntervalSeconds = 123,
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    HlsManifests = list(
      list(
        AdMarkers = "NONE"|"SCTE35_ENHANCED"|"PASSTHROUGH",
        Id = "string",
        IncludeIframeOnlyStream = TRUE|FALSE,
        ManifestName = "string",
        PlaylistType = "NONE"|"EVENT"|"VOD",
        PlaylistWindowSeconds = 123,
        ProgramDateTimeIntervalSeconds = 123
      )
    ),
    SegmentDurationSeconds = 123,
    SegmentPrefix = "string",
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    )
  ),
  DashPackage = list(
    Encryption = list(
      KeyRotationIntervalSeconds = 123,
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    ManifestLayout = "FULL"|"COMPACT",
    ManifestWindowSeconds = 123,
    MinBufferTimeSeconds = 123,
    MinUpdatePeriodSeconds = 123,
    PeriodTriggers = list(
      "ADS"
    ),
    Profile = "NONE"|"HBBTV_1_5",
    SegmentDurationSeconds = 123,
    SegmentTemplateFormat = "NUMBER_WITH_TIMELINE"|"TIME_WITH_TIMELINE"|"NUMBER_WITH_DURATION",
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    ),
    SuggestedPresentationDelaySeconds = 123
  ),
  Description = "string",
  HlsPackage = list(
    AdMarkers = "NONE"|"SCTE35_ENHANCED"|"PASSTHROUGH",
    Encryption = list(
      ConstantInitializationVector = "string",
      EncryptionMethod = "AES_128"|"SAMPLE_AES",
      KeyRotationIntervalSeconds = 123,
      RepeatExtXKey = TRUE|FALSE,
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    IncludeIframeOnlyStream = TRUE|FALSE,
    PlaylistType = "NONE"|"EVENT"|"VOD",
    PlaylistWindowSeconds = 123,
    ProgramDateTimeIntervalSeconds = 123,
    SegmentDurationSeconds = 123,
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    ),
    UseAudioRenditionGroup = TRUE|FALSE
  ),
  Id = "string",
  ManifestName = "string",
  MssPackage = list(
    Encryption = list(
      SpekeKeyProvider = list(
        CertificateArn = "string",
        ResourceId = "string",
        RoleArn = "string",
        SystemIds = list(
          "string"
        ),
        Url = "string"
      )
    ),
    ManifestWindowSeconds = 123,
    SegmentDurationSeconds = 123,
    StreamSelection = list(
      MaxVideoBitsPerSecond = 123,
      MinVideoBitsPerSecond = 123,
      StreamOrder = "ORIGINAL"|"VIDEO_BITRATE_ASCENDING"|"VIDEO_BITRATE_DESCENDING"
    )
  ),
  StartoverWindowSeconds = 123,
  Tags = list(
    "string"
  ),
  TimeDelaySeconds = 123,
  Whitelist = list(
    "string"
  )
)
}
}

\keyword{internal}
