% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_delete_channel}
\alias{mediapackage_delete_channel}
\title{Deletes an existing Channel}
\usage{
mediapackage_delete_channel(Id)
}
\arguments{
\item{Id}{[required] The ID of the Channel to delete.}
}
\value{
An empty list.
}
\description{
Deletes an existing Channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_channel(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
