% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_delete_origin_endpoint}
\alias{mediapackage_delete_origin_endpoint}
\title{Deletes an existing OriginEndpoint}
\usage{
mediapackage_delete_origin_endpoint(Id)
}
\arguments{
\item{Id}{[required] The ID of the OriginEndpoint to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an existing OriginEndpoint.
}
\section{Request syntax}{
\preformatted{svc$delete_origin_endpoint(
  Id = "string"
)
}
}

\keyword{internal}
