% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_describe_channel}
\alias{mediapackage_describe_channel}
\title{Gets details about a Channel}
\usage{
mediapackage_describe_channel(Id)
}
\arguments{
\item{Id}{[required] The ID of a Channel.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Description = "string",
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  HlsIngest = list(
    IngestEndpoints = list(
      list(
        Id = "string",
        Password = "string",
        Url = "string",
        Username = "string"
      )
    )
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets details about a Channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_channel(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
