% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_describe_harvest_job}
\alias{mediapackage_describe_harvest_job}
\title{Gets details about an existing HarvestJob}
\usage{
mediapackage_describe_harvest_job(Id)
}
\arguments{
\item{Id}{[required] The ID of the HarvestJob.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  ChannelId = "string",
  CreatedAt = "string",
  EndTime = "string",
  Id = "string",
  OriginEndpointId = "string",
  S3Destination = list(
    BucketName = "string",
    ManifestKey = "string",
    RoleArn = "string"
  ),
  StartTime = "string",
  Status = "IN_PROGRESS"|"SUCCEEDED"|"FAILED"
)
}
}
\description{
Gets details about an existing HarvestJob.
}
\section{Request syntax}{
\preformatted{svc$describe_harvest_job(
  Id = "string"
)
}
}

\keyword{internal}
