% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_list_harvest_jobs}
\alias{mediapackage_list_harvest_jobs}
\title{Returns a collection of HarvestJob records}
\usage{
mediapackage_list_harvest_jobs(IncludeChannelId, IncludeStatus,
  MaxResults, NextToken)
}
\arguments{
\item{IncludeChannelId}{When specified, the request will return only HarvestJobs associated with
the given Channel ID.}

\item{IncludeStatus}{When specified, the request will return only HarvestJobs in the given
status.}

\item{MaxResults}{The upper bound on the number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  HarvestJobs = list(
    list(
      Arn = "string",
      ChannelId = "string",
      CreatedAt = "string",
      EndTime = "string",
      Id = "string",
      OriginEndpointId = "string",
      S3Destination = list(
        BucketName = "string",
        ManifestKey = "string",
        RoleArn = "string"
      ),
      StartTime = "string",
      Status = "IN_PROGRESS"|"SUCCEEDED"|"FAILED"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a collection of HarvestJob records.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_harvest_jobs(
  IncludeChannelId = "string",
  IncludeStatus = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
