% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_rotate_channel_credentials}
\alias{mediapackage_rotate_channel_credentials}
\title{Changes the Channel's first IngestEndpoint's username and password}
\usage{
mediapackage_rotate_channel_credentials(Id)
}
\arguments{
\item{Id}{[required] The ID of the channel to update.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Description = "string",
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  HlsIngest = list(
    IngestEndpoints = list(
      list(
        Id = "string",
        Password = "string",
        Url = "string",
        Username = "string"
      )
    )
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Changes the Channel's first IngestEndpoint's username and password.
WARNING - This API is deprecated. Please use
RotateIngestEndpointCredentials instead
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$rotate_channel_credentials(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
