% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_rotate_ingest_endpoint_credentials}
\alias{mediapackage_rotate_ingest_endpoint_credentials}
\title{Rotate the IngestEndpoint's username and password, as specified by the
IngestEndpoint's id}
\usage{
mediapackage_rotate_ingest_endpoint_credentials(Id, IngestEndpointId)
}
\arguments{
\item{Id}{[required] The ID of the channel the IngestEndpoint is on.}

\item{IngestEndpointId}{[required] The id of the IngestEndpoint whose credentials should be rotated}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Description = "string",
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  HlsIngest = list(
    IngestEndpoints = list(
      list(
        Id = "string",
        Password = "string",
        Url = "string",
        Username = "string"
      )
    )
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Rotate the IngestEndpoint's username and password, as specified by the
IngestEndpoint's id.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$rotate_ingest_endpoint_credentials(
  Id = "string",
  IngestEndpointId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
