% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_untag_resource}
\alias{mediapackage_untag_resource}
\title{Untag resource}
\usage{
mediapackage_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required]}

\item{TagKeys}{[required] The key(s) of tag to be deleted}
}
\value{
An empty list.
}
\description{
Untag resource
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
