% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_service.R
\name{mediastore}
\alias{mediastore}
\title{AWS Elemental MediaStore}
\usage{
mediastore(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
An AWS Elemental MediaStore container is a namespace that holds folders
and objects. You use a container endpoint to create, read, and delete
objects.
}
\section{Service syntax}{
\preformatted{svc <- mediastore(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=mediastore_create_container]{create_container} \tab Creates a storage container to hold objects \cr
\link[=mediastore_delete_container]{delete_container} \tab Deletes the specified container \cr
\link[=mediastore_delete_container_policy]{delete_container_policy} \tab Deletes the access policy that is associated with the specified container \cr
\link[=mediastore_delete_cors_policy]{delete_cors_policy} \tab Deletes the cross-origin resource sharing (CORS) configuration information that is set for the container \cr
\link[=mediastore_delete_lifecycle_policy]{delete_lifecycle_policy} \tab Removes an object lifecycle policy from a container \cr
\link[=mediastore_delete_metric_policy]{delete_metric_policy} \tab Deletes the metric policy that is associated with the specified container \cr
\link[=mediastore_describe_container]{describe_container} \tab Retrieves the properties of the requested container \cr
\link[=mediastore_get_container_policy]{get_container_policy} \tab Retrieves the access policy for the specified container \cr
\link[=mediastore_get_cors_policy]{get_cors_policy} \tab Returns the cross-origin resource sharing (CORS) configuration information that is set for the container \cr
\link[=mediastore_get_lifecycle_policy]{get_lifecycle_policy} \tab Retrieves the object lifecycle policy that is assigned to a container \cr
\link[=mediastore_get_metric_policy]{get_metric_policy} \tab Returns the metric policy for the specified container \cr
\link[=mediastore_list_containers]{list_containers} \tab Lists the properties of all containers in AWS Elemental MediaStore \cr
\link[=mediastore_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of the tags assigned to the specified container \cr
\link[=mediastore_put_container_policy]{put_container_policy} \tab Creates an access policy for the specified container to restrict the users and clients that can access it \cr
\link[=mediastore_put_cors_policy]{put_cors_policy} \tab Sets the cross-origin resource sharing (CORS) configuration on a container so that the container can service cross-origin requests\cr
\link[=mediastore_put_lifecycle_policy]{put_lifecycle_policy} \tab Writes an object lifecycle policy to a container \cr
\link[=mediastore_put_metric_policy]{put_metric_policy} \tab The metric policy that you want to add to the container \cr
\link[=mediastore_start_access_logging]{start_access_logging} \tab Starts access logging on the specified container \cr
\link[=mediastore_stop_access_logging]{stop_access_logging} \tab Stops access logging on the specified container \cr
\link[=mediastore_tag_resource]{tag_resource} \tab Adds tags to the specified AWS Elemental MediaStore container \cr
\link[=mediastore_untag_resource]{untag_resource} \tab Removes tags from the specified container
}
}

\examples{
\dontrun{
svc <- mediastore()
svc$create_container(
  Foo = 123
)
}

}
