% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_create_container}
\alias{mediastore_create_container}
\title{Creates a storage container to hold objects}
\usage{
mediastore_create_container(ContainerName, Tags)
}
\arguments{
\item{ContainerName}{[required] The name for the container. The name must be from 1 to 255 characters.
Container names must be unique to your AWS account within a specific
region. As an example, you could create a container named \code{movies} in
every region, as long as you don’t have an existing container with that
name.}

\item{Tags}{An array of key:value pairs that you define. These values can be
anything that you want. Typically, the tag key represents a category
(such as "environment") and the tag value represents a specific value
within that category (such as "test," "development," or "production").
You can add up to 50 tags to each container. For more information about
tagging, including naming and usage conventions, see \href{https://docs.aws.amazon.com/mediastore/latest/ug/tagging.html}{Tagging Resources in MediaStore}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Container = list(
    Endpoint = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    ARN = "string",
    Name = "string",
    Status = "ACTIVE"|"CREATING"|"DELETING",
    AccessLoggingEnabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a storage container to hold objects. A container is similar to a
bucket in the Amazon S3 service.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_container(
  ContainerName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
