% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_delete_container}
\alias{mediastore_delete_container}
\title{Deletes the specified container}
\usage{
mediastore_delete_container(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container to delete.}
}
\description{
Deletes the specified container. Before you make a \code{DeleteContainer}
request, delete any objects in the container or in any folders in the
container. You can delete only empty containers.
}
\section{Request syntax}{
\preformatted{svc$delete_container(
  ContainerName = "string"
)
}
}

\keyword{internal}
