% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_delete_container_policy}
\alias{mediastore_delete_container_policy}
\title{Deletes the access policy that is associated with the specified
container}
\usage{
mediastore_delete_container_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that holds the policy.}
}
\value{
An empty list.
}
\description{
Deletes the access policy that is associated with the specified
container.
}
\section{Request syntax}{
\preformatted{svc$delete_container_policy(
  ContainerName = "string"
)
}
}

\keyword{internal}
