% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_delete_cors_policy}
\alias{mediastore_delete_cors_policy}
\title{Deletes the cross-origin resource sharing (CORS) configuration
information that is set for the container}
\usage{
mediastore_delete_cors_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container to remove the policy from.}
}
\value{
An empty list.
}
\description{
Deletes the cross-origin resource sharing (CORS) configuration
information that is set for the container.

To use this operation, you must have permission to perform the
\code{MediaStore:DeleteCorsPolicy} action. The container owner has this
permission by default and can grant this permission to others.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_cors_policy(
  ContainerName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
