% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_delete_lifecycle_policy}
\alias{mediastore_delete_lifecycle_policy}
\title{Removes an object lifecycle policy from a container}
\usage{
mediastore_delete_lifecycle_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that holds the object lifecycle policy.}
}
\value{
An empty list.
}
\description{
Removes an object lifecycle policy from a container. It takes up to 20
minutes for the change to take effect.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_lifecycle_policy(
  ContainerName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
