% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_delete_metric_policy}
\alias{mediastore_delete_metric_policy}
\title{Deletes the metric policy that is associated with the specified
container}
\usage{
mediastore_delete_metric_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that is associated with the metric policy that
you want to delete.}
}
\description{
Deletes the metric policy that is associated with the specified
container. If there is no metric policy associated with the container,
MediaStore doesn\'t send metrics to CloudWatch.
}
\section{Request syntax}{
\preformatted{svc$delete_metric_policy(
  ContainerName = "string"
)
}
}

\keyword{internal}
