% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_get_lifecycle_policy}
\alias{mediastore_get_lifecycle_policy}
\title{Retrieves the object lifecycle policy that is assigned to a container}
\usage{
mediastore_get_lifecycle_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that the object lifecycle policy is assigned
to.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LifecyclePolicy = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the object lifecycle policy that is assigned to a container.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_lifecycle_policy(
  ContainerName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
