% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_list_containers}
\alias{mediastore_list_containers}
\title{Lists the properties of all containers in AWS Elemental MediaStore}
\usage{
mediastore_list_containers(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{Only if you used \code{MaxResults} in the first command, enter the token
(which was included in the previous response) to obtain the next set of
containers. This token is included in a response only if there actually
are more containers to list.}

\item{MaxResults}{Enter the maximum number of containers in the response. Use from 1 to
255 characters.}
}
\value{
A list with the following syntax:\preformatted{list(
  Containers = list(
    list(
      Endpoint = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      ARN = "string",
      Name = "string",
      Status = "ACTIVE"|"CREATING"|"DELETING",
      AccessLoggingEnabled = TRUE|FALSE
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the properties of all containers in AWS Elemental MediaStore.

You can query to receive all the containers in one response. Or you can
include the \code{MaxResults} parameter to receive a limited number of
containers in each response. In this case, the response includes a
token. To get the next set of containers, send the command again, this
time with the \code{NextToken} parameter (with the returned token as its
value). The next set of responses appears, with a token if there are
still more containers to receive.

See also \code{\link[=mediastore_describe_container]{describe_container}}, which
gets the properties of one container.
}
\section{Request syntax}{
\preformatted{svc$list_containers(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
