% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_list_tags_for_resource}
\alias{mediastore_list_tags_for_resource}
\title{Returns a list of the tags assigned to the specified container}
\usage{
mediastore_list_tags_for_resource(Resource)
}
\arguments{
\item{Resource}{[required] The Amazon Resource Name (ARN) for the container.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of the tags assigned to the specified container.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  Resource = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
