% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_put_lifecycle_policy}
\alias{mediastore_put_lifecycle_policy}
\title{Writes an object lifecycle policy to a container}
\usage{
mediastore_put_lifecycle_policy(ContainerName, LifecyclePolicy)
}
\arguments{
\item{ContainerName}{[required] The name of the container that you want to assign the object lifecycle
policy to.}

\item{LifecyclePolicy}{[required] The object lifecycle policy to apply to the container.}
}
\value{
An empty list.
}
\description{
Writes an object lifecycle policy to a container. If the container
already has an object lifecycle policy, the service replaces the
existing policy with the new policy. It takes up to 20 minutes for the
change to take effect.

For information about how to construct an object lifecycle policy, see
\href{https://docs.aws.amazon.com/mediastore/latest/ug/policies-object-lifecycle-components.html}{Components of an Object Lifecycle Policy}.
}
\section{Request syntax}{
\preformatted{svc$put_lifecycle_policy(
  ContainerName = "string",
  LifecyclePolicy = "string"
)
}
}

\keyword{internal}
