% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_tag_resource}
\alias{mediastore_tag_resource}
\title{Adds tags to the specified AWS Elemental MediaStore container}
\usage{
mediastore_tag_resource(Resource, Tags)
}
\arguments{
\item{Resource}{[required] The Amazon Resource Name (ARN) for the container.}

\item{Tags}{[required] An array of key:value pairs that you want to add to the container. You
need to specify only the tags that you want to add or update. For
example, suppose a container already has two tags (customer:CompanyA and
priority:High). You want to change the priority tag and also add a third
tag (type:Contract). For TagResource, you specify the following tags:
priority:Medium, type:Contract. The result is that your container has
three tags: customer:CompanyA, priority:Medium, and type:Contract.}
}
\description{
Adds tags to the specified AWS Elemental MediaStore container. Tags are
key:value pairs that you can associate with AWS resources. For example,
the tag key might be "customer" and the tag value might be
"companyA." You can specify one or more tags to add to each container.
You can add up to 50 tags to each container. For more information about
tagging, including naming and usage conventions, see Tagging Resources
in MediaStore.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  Resource = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
