% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastoredata_operations.R
\name{mediastoredata_delete_object}
\alias{mediastoredata_delete_object}
\title{Deletes an object at the specified path}
\usage{
mediastoredata_delete_object(Path)
}
\arguments{
\item{Path}{[required] The path (including the file name) where the object is stored in the
container. Format: \<folder name\>/\<folder name\>/\<file name\>}
}
\value{
An empty list.
}
\description{
Deletes an object at the specified path.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_object(
  Path = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
