% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastoredata_operations.R
\name{mediastoredata_describe_object}
\alias{mediastoredata_describe_object}
\title{Gets the headers for an object at the specified path}
\usage{
mediastoredata_describe_object(Path)
}
\arguments{
\item{Path}{[required] The path (including the file name) where the object is stored in the
container. Format: \<folder name\>/\<folder name\>/\<file name\>}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ETag = "string",
  ContentType = "string",
  ContentLength = 123,
  CacheControl = "string",
  LastModified = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the headers for an object at the specified path.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_object(
  Path = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
