% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_create_channel}
\alias{mediatailor_create_channel}
\title{Creates a channel}
\usage{
mediatailor_create_channel(ChannelName, Outputs, PlaybackMode, Tags)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{Outputs}{[required] The channel's output properties.}

\item{PlaybackMode}{[required] The type of playback mode for this channel. The only supported value is
LOOP.}

\item{Tags}{The tags to assign to the channel.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  ChannelName = "string",
  ChannelState = "RUNNING"|"STOPPED",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  Outputs = list(
    list(
      DashPlaylistSettings = list(
        ManifestWindowSeconds = 123,
        MinBufferTimeSeconds = 123,
        MinUpdatePeriodSeconds = 123,
        SuggestedPresentationDelaySeconds = 123
      ),
      HlsPlaylistSettings = list(
        ManifestWindowSeconds = 123
      ),
      ManifestName = "string",
      PlaybackUrl = "string",
      SourceGroup = "string"
    )
  ),
  PlaybackMode = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Creates a channel.
}
\section{Request syntax}{
\preformatted{svc$create_channel(
  ChannelName = "string",
  Outputs = list(
    list(
      DashPlaylistSettings = list(
        ManifestWindowSeconds = 123,
        MinBufferTimeSeconds = 123,
        MinUpdatePeriodSeconds = 123,
        SuggestedPresentationDelaySeconds = 123
      ),
      HlsPlaylistSettings = list(
        ManifestWindowSeconds = 123
      ),
      ManifestName = "string",
      SourceGroup = "string"
    )
  ),
  PlaybackMode = "LOOP",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
