% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_create_program}
\alias{mediatailor_create_program}
\title{Creates a program}
\usage{
mediatailor_create_program(AdBreaks, ChannelName, ProgramName,
  ScheduleConfiguration, SourceLocationName, VodSourceName)
}
\arguments{
\item{AdBreaks}{The ad break configuration settings.}

\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{ProgramName}{[required] The identifier for the program you are working on.}

\item{ScheduleConfiguration}{[required] The schedule configuration settings.}

\item{SourceLocationName}{[required] The name of the source location.}

\item{VodSourceName}{[required] The name that's used to refer to a VOD source.}
}
\value{
A list with the following syntax:\preformatted{list(
  AdBreaks = list(
    list(
      MessageType = "SPLICE_INSERT",
      OffsetMillis = 123,
      Slate = list(
        SourceLocationName = "string",
        VodSourceName = "string"
      ),
      SpliceInsertMessage = list(
        AvailNum = 123,
        AvailsExpected = 123,
        SpliceEventId = 123,
        UniqueProgramId = 123
      )
    )
  ),
  Arn = "string",
  ChannelName = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  ProgramName = "string",
  SourceLocationName = "string",
  VodSourceName = "string"
)
}
}
\description{
Creates a program.
}
\section{Request syntax}{
\preformatted{svc$create_program(
  AdBreaks = list(
    list(
      MessageType = "SPLICE_INSERT",
      OffsetMillis = 123,
      Slate = list(
        SourceLocationName = "string",
        VodSourceName = "string"
      ),
      SpliceInsertMessage = list(
        AvailNum = 123,
        AvailsExpected = 123,
        SpliceEventId = 123,
        UniqueProgramId = 123
      )
    )
  ),
  ChannelName = "string",
  ProgramName = "string",
  ScheduleConfiguration = list(
    Transition = list(
      RelativePosition = "BEFORE_PROGRAM"|"AFTER_PROGRAM",
      RelativeProgram = "string",
      Type = "string"
    )
  ),
  SourceLocationName = "string",
  VodSourceName = "string"
)
}
}

\keyword{internal}
