% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_create_vod_source}
\alias{mediatailor_create_vod_source}
\title{Creates name for a specific VOD source in a source location}
\usage{
mediatailor_create_vod_source(HttpPackageConfigurations,
  SourceLocationName, Tags, VodSourceName)
}
\arguments{
\item{HttpPackageConfigurations}{[required] An array of HTTP package configuration parameters for this VOD source.}

\item{SourceLocationName}{[required] The identifier for the source location you are working on.}

\item{Tags}{The tags to assign to the VOD source.}

\item{VodSourceName}{[required] The identifier for the VOD source you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  HttpPackageConfigurations = list(
    list(
      Path = "string",
      SourceGroup = "string",
      Type = "DASH"|"HLS"
    )
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  SourceLocationName = "string",
  Tags = list(
    "string"
  ),
  VodSourceName = "string"
)
}
}
\description{
Creates name for a specific VOD source in a source location.
}
\section{Request syntax}{
\preformatted{svc$create_vod_source(
  HttpPackageConfigurations = list(
    list(
      Path = "string",
      SourceGroup = "string",
      Type = "DASH"|"HLS"
    )
  ),
  SourceLocationName = "string",
  Tags = list(
    "string"
  ),
  VodSourceName = "string"
)
}
}

\keyword{internal}
