% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_delete_channel}
\alias{mediatailor_delete_channel}
\title{Deletes a channel}
\usage{
mediatailor_delete_channel(ChannelName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}
}
\value{
An empty list.
}
\description{
Deletes a channel. You must stop the channel before it can be deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_channel(
  ChannelName = "string"
)
}
}

\keyword{internal}
