% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_delete_channel_policy}
\alias{mediatailor_delete_channel_policy}
\title{Deletes a channel's IAM policy}
\usage{
mediatailor_delete_channel_policy(ChannelName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}
}
\value{
An empty list.
}
\description{
Deletes a channel's IAM policy.
}
\section{Request syntax}{
\preformatted{svc$delete_channel_policy(
  ChannelName = "string"
)
}
}

\keyword{internal}
