% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_delete_program}
\alias{mediatailor_delete_program}
\title{Deletes a specific program on a specific channel}
\usage{
mediatailor_delete_program(ChannelName, ProgramName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{ProgramName}{[required] The identifier for the program you are working on.}
}
\value{
An empty list.
}
\description{
Deletes a specific program on a specific channel.
}
\section{Request syntax}{
\preformatted{svc$delete_program(
  ChannelName = "string",
  ProgramName = "string"
)
}
}

\keyword{internal}
