% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_delete_source_location}
\alias{mediatailor_delete_source_location}
\title{Deletes a source location on a specific channel}
\usage{
mediatailor_delete_source_location(SourceLocationName)
}
\arguments{
\item{SourceLocationName}{[required] The identifier for the source location you are working on.}
}
\value{
An empty list.
}
\description{
Deletes a source location on a specific channel.
}
\section{Request syntax}{
\preformatted{svc$delete_source_location(
  SourceLocationName = "string"
)
}
}

\keyword{internal}
