% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_delete_vod_source}
\alias{mediatailor_delete_vod_source}
\title{Deletes a specific VOD source in a specific source location}
\usage{
mediatailor_delete_vod_source(SourceLocationName, VodSourceName)
}
\arguments{
\item{SourceLocationName}{[required] The identifier for the source location you are working on.}

\item{VodSourceName}{[required] The identifier for the VOD source you are working on.}
}
\value{
An empty list.
}
\description{
Deletes a specific VOD source in a specific source location.
}
\section{Request syntax}{
\preformatted{svc$delete_vod_source(
  SourceLocationName = "string",
  VodSourceName = "string"
)
}
}

\keyword{internal}
