% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_describe_channel}
\alias{mediatailor_describe_channel}
\title{Describes the properties of a specific channel}
\usage{
mediatailor_describe_channel(ChannelName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  ChannelName = "string",
  ChannelState = "RUNNING"|"STOPPED",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  Outputs = list(
    list(
      DashPlaylistSettings = list(
        ManifestWindowSeconds = 123,
        MinBufferTimeSeconds = 123,
        MinUpdatePeriodSeconds = 123,
        SuggestedPresentationDelaySeconds = 123
      ),
      HlsPlaylistSettings = list(
        ManifestWindowSeconds = 123
      ),
      ManifestName = "string",
      PlaybackUrl = "string",
      SourceGroup = "string"
    )
  ),
  PlaybackMode = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Describes the properties of a specific channel.
}
\section{Request syntax}{
\preformatted{svc$describe_channel(
  ChannelName = "string"
)
}
}

\keyword{internal}
