% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_describe_program}
\alias{mediatailor_describe_program}
\title{Retrieves the properties of the requested program}
\usage{
mediatailor_describe_program(ChannelName, ProgramName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{ProgramName}{[required] The identifier for the program you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  AdBreaks = list(
    list(
      MessageType = "SPLICE_INSERT",
      OffsetMillis = 123,
      Slate = list(
        SourceLocationName = "string",
        VodSourceName = "string"
      ),
      SpliceInsertMessage = list(
        AvailNum = 123,
        AvailsExpected = 123,
        SpliceEventId = 123,
        UniqueProgramId = 123
      )
    )
  ),
  Arn = "string",
  ChannelName = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  ProgramName = "string",
  SourceLocationName = "string",
  VodSourceName = "string"
)
}
}
\description{
Retrieves the properties of the requested program.
}
\section{Request syntax}{
\preformatted{svc$describe_program(
  ChannelName = "string",
  ProgramName = "string"
)
}
}

\keyword{internal}
