% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_describe_source_location}
\alias{mediatailor_describe_source_location}
\title{Retrieves the properties of the requested source location}
\usage{
mediatailor_describe_source_location(SourceLocationName)
}
\arguments{
\item{SourceLocationName}{[required] The identifier for the source location you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  AccessConfiguration = list(
    AccessType = "S3_SIGV4"
  ),
  Arn = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  DefaultSegmentDeliveryConfiguration = list(
    BaseUrl = "string"
  ),
  HttpConfiguration = list(
    BaseUrl = "string"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  SourceLocationName = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Retrieves the properties of the requested source location.
}
\section{Request syntax}{
\preformatted{svc$describe_source_location(
  SourceLocationName = "string"
)
}
}

\keyword{internal}
