% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_describe_vod_source}
\alias{mediatailor_describe_vod_source}
\title{Provides details about a specific VOD source in a specific source
location}
\usage{
mediatailor_describe_vod_source(SourceLocationName, VodSourceName)
}
\arguments{
\item{SourceLocationName}{[required] The identifier for the source location you are working on.}

\item{VodSourceName}{[required] The identifier for the VOD source you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  HttpPackageConfigurations = list(
    list(
      Path = "string",
      SourceGroup = "string",
      Type = "DASH"|"HLS"
    )
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  SourceLocationName = "string",
  Tags = list(
    "string"
  ),
  VodSourceName = "string"
)
}
}
\description{
Provides details about a specific VOD source in a specific source
location.
}
\section{Request syntax}{
\preformatted{svc$describe_vod_source(
  SourceLocationName = "string",
  VodSourceName = "string"
)
}
}

\keyword{internal}
