% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_get_channel_policy}
\alias{mediatailor_get_channel_policy}
\title{Retrieves information about a channel's IAM policy}
\usage{
mediatailor_get_channel_policy(ChannelName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  Policy = "string"
)
}
}
\description{
Retrieves information about a channel's IAM policy.
}
\section{Request syntax}{
\preformatted{svc$get_channel_policy(
  ChannelName = "string"
)
}
}

\keyword{internal}
