% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_get_channel_schedule}
\alias{mediatailor_get_channel_schedule}
\title{Retrieves information about your channel's schedule}
\usage{
mediatailor_get_channel_schedule(ChannelName, DurationMinutes,
  MaxResults, NextToken)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{DurationMinutes}{The schedule duration in minutes. The maximum duration is 4320 minutes
(three days).}

\item{MaxResults}{Upper bound on number of records to return. The maximum number of
results is 100.}

\item{NextToken}{Pagination token from the GET list request. Use the token to fetch the
next page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Items = list(
    list(
      ApproximateDurationSeconds = 123,
      ApproximateStartTime = as.POSIXct(
        "2015-01-01"
      ),
      Arn = "string",
      ChannelName = "string",
      ProgramName = "string",
      SourceLocationName = "string",
      VodSourceName = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves information about your channel's schedule.
}
\section{Request syntax}{
\preformatted{svc$get_channel_schedule(
  ChannelName = "string",
  DurationMinutes = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
