% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_list_channels}
\alias{mediatailor_list_channels}
\title{Retrieves a list of channels that are associated with this account}
\usage{
mediatailor_list_channels(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return. The maximum number of
results is 100.}

\item{NextToken}{Pagination token from the GET list request. Use the token to fetch the
next page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Items = list(
    list(
      Arn = "string",
      ChannelName = "string",
      ChannelState = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      Outputs = list(
        list(
          DashPlaylistSettings = list(
            ManifestWindowSeconds = 123,
            MinBufferTimeSeconds = 123,
            MinUpdatePeriodSeconds = 123,
            SuggestedPresentationDelaySeconds = 123
          ),
          HlsPlaylistSettings = list(
            ManifestWindowSeconds = 123
          ),
          ManifestName = "string",
          PlaybackUrl = "string",
          SourceGroup = "string"
        )
      ),
      PlaybackMode = "string",
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves a list of channels that are associated with this account.
}
\section{Request syntax}{
\preformatted{svc$list_channels(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
