% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_list_source_locations}
\alias{mediatailor_list_source_locations}
\title{Retrieves a list of source locations}
\usage{
mediatailor_list_source_locations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return. The maximum number of
results is 100.}

\item{NextToken}{Pagination token from the GET list request. Use the token to fetch the
next page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Items = list(
    list(
      AccessConfiguration = list(
        AccessType = "S3_SIGV4"
      ),
      Arn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      DefaultSegmentDeliveryConfiguration = list(
        BaseUrl = "string"
      ),
      HttpConfiguration = list(
        BaseUrl = "string"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      SourceLocationName = "string",
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves a list of source locations.
}
\section{Request syntax}{
\preformatted{svc$list_source_locations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
