% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_list_tags_for_resource}
\alias{mediatailor_list_tags_for_resource}
\title{Returns a list of the tags assigned to the specified playback
configuration resource}
\usage{
mediatailor_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the playback configuration. You can
get this from the response to any playback configuration request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of the tags assigned to the specified playback
configuration resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
