% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_list_vod_sources}
\alias{mediatailor_list_vod_sources}
\title{Lists all the VOD sources in a source location}
\usage{
mediatailor_list_vod_sources(MaxResults, NextToken, SourceLocationName)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return. The maximum number of
results is 100.}

\item{NextToken}{Pagination token from the GET list request. Use the token to fetch the
next page of results.}

\item{SourceLocationName}{[required] The identifier for the source location you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  Items = list(
    list(
      Arn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      HttpPackageConfigurations = list(
        list(
          Path = "string",
          SourceGroup = "string",
          Type = "DASH"|"HLS"
        )
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      SourceLocationName = "string",
      Tags = list(
        "string"
      ),
      VodSourceName = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all the VOD sources in a source location.
}
\section{Request syntax}{
\preformatted{svc$list_vod_sources(
  MaxResults = 123,
  NextToken = "string",
  SourceLocationName = "string"
)
}
}

\keyword{internal}
