% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_put_channel_policy}
\alias{mediatailor_put_channel_policy}
\title{Creates an IAM policy for the channel}
\usage{
mediatailor_put_channel_policy(ChannelName, Policy)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{Policy}{[required] Adds an IAM role that determines the permissions of your channel.}
}
\value{
An empty list.
}
\description{
Creates an IAM policy for the channel.
}
\section{Request syntax}{
\preformatted{svc$put_channel_policy(
  ChannelName = "string",
  Policy = "string"
)
}
}

\keyword{internal}
