% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_put_playback_configuration}
\alias{mediatailor_put_playback_configuration}
\title{Adds a new playback configuration to AWS Elemental MediaTailor}
\usage{
mediatailor_put_playback_configuration(AdDecisionServerUrl,
  CdnConfiguration, DashConfiguration, LivePreRollConfiguration, Name,
  SlateAdUrl, Tags, TranscodeProfileName, VideoContentSourceUrl)
}
\arguments{
\item{AdDecisionServerUrl}{The URL for the ad decision server (ADS). This includes the
specification of static parameters and placeholders for dynamic
parameters. AWS Elemental MediaTailor substitutes player-specific and
session-specific parameters as needed when calling the ADS. Alternately,
for testing you can provide a static VAST URL. The maximum length is
25,000 characters.}

\item{CdnConfiguration}{The configuration for using a content delivery network (CDN), like
Amazon CloudFront, for content and ad segment management.}

\item{DashConfiguration}{The configuration for DASH content.}

\item{LivePreRollConfiguration}{The configuration for pre-roll ad insertion.}

\item{Name}{The identifier for the playback configuration.}

\item{SlateAdUrl}{The URL for a high-quality video asset to transcode and use to fill in
time that\'s not used by ads. AWS Elemental MediaTailor shows the slate
to fill in gaps in media content. Configuring the slate is optional for
non-VPAID configurations. For VPAID, the slate is required because
MediaTailor provides it in the slots that are designated for dynamic ad
content. The slate must be a high-quality asset that contains both audio
and video.}

\item{Tags}{The tags to assign to the playback configuration.}

\item{TranscodeProfileName}{The name that is used to associate this playback configuration with a
custom transcode profile. This overrides the dynamic transcoding
defaults of MediaTailor. Use this only if you have already set up custom
profiles with the help of AWS Support.}

\item{VideoContentSourceUrl}{The URL prefix for the master playlist for the stream, minus the asset
ID. The maximum length is 512 characters.}
}
\description{
Adds a new playback configuration to AWS Elemental MediaTailor.
}
\section{Request syntax}{
\preformatted{svc$put_playback_configuration(
  AdDecisionServerUrl = "string",
  CdnConfiguration = list(
    AdSegmentUrlPrefix = "string",
    ContentSegmentUrlPrefix = "string"
  ),
  DashConfiguration = list(
    MpdLocation = "string",
    OriginManifestType = "SINGLE_PERIOD"|"MULTI_PERIOD"
  ),
  LivePreRollConfiguration = list(
    AdDecisionServerUrl = "string",
    MaxDurationSeconds = 123
  ),
  Name = "string",
  SlateAdUrl = "string",
  Tags = list(
    "string"
  ),
  TranscodeProfileName = "string",
  VideoContentSourceUrl = "string"
)
}
}

\keyword{internal}
