% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_start_channel}
\alias{mediatailor_start_channel}
\title{Starts a specific channel}
\usage{
mediatailor_start_channel(ChannelName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}
}
\value{
An empty list.
}
\description{
Starts a specific channel.
}
\section{Request syntax}{
\preformatted{svc$start_channel(
  ChannelName = "string"
)
}
}

\keyword{internal}
