% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_stop_channel}
\alias{mediatailor_stop_channel}
\title{Stops a specific channel}
\usage{
mediatailor_stop_channel(ChannelName)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}
}
\value{
An empty list.
}
\description{
Stops a specific channel.
}
\section{Request syntax}{
\preformatted{svc$stop_channel(
  ChannelName = "string"
)
}
}

\keyword{internal}
