% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_tag_resource}
\alias{mediatailor_tag_resource}
\title{Adds tags to the specified playback configuration resource}
\usage{
mediatailor_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the playback configuration. You can
get this from the response to any playback configuration request.}

\item{Tags}{[required] A comma-separated list of tag key:value pairs.}
}
\value{
An empty list.
}
\description{
Adds tags to the specified playback configuration resource. You can
specify one or more tags to add.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
