% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_untag_resource}
\alias{mediatailor_untag_resource}
\title{Removes tags from the specified playback configuration resource}
\usage{
mediatailor_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the playback configuration. You can
get this from the response to any playback configuration request.}

\item{TagKeys}{[required] A comma-separated list of the tag keys to remove from the playback
configuration.}
}
\description{
Removes tags from the specified playback configuration resource. You can
specify one or more tags to remove.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
