% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_update_channel}
\alias{mediatailor_update_channel}
\title{Updates an existing channel}
\usage{
mediatailor_update_channel(ChannelName, Outputs)
}
\arguments{
\item{ChannelName}{[required] The identifier for the channel you are working on.}

\item{Outputs}{[required] The channel's output properties.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  ChannelName = "string",
  ChannelState = "RUNNING"|"STOPPED",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  Outputs = list(
    list(
      DashPlaylistSettings = list(
        ManifestWindowSeconds = 123,
        MinBufferTimeSeconds = 123,
        MinUpdatePeriodSeconds = 123,
        SuggestedPresentationDelaySeconds = 123
      ),
      HlsPlaylistSettings = list(
        ManifestWindowSeconds = 123
      ),
      ManifestName = "string",
      PlaybackUrl = "string",
      SourceGroup = "string"
    )
  ),
  PlaybackMode = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Updates an existing channel.
}
\section{Request syntax}{
\preformatted{svc$update_channel(
  ChannelName = "string",
  Outputs = list(
    list(
      DashPlaylistSettings = list(
        ManifestWindowSeconds = 123,
        MinBufferTimeSeconds = 123,
        MinUpdatePeriodSeconds = 123,
        SuggestedPresentationDelaySeconds = 123
      ),
      HlsPlaylistSettings = list(
        ManifestWindowSeconds = 123
      ),
      ManifestName = "string",
      SourceGroup = "string"
    )
  )
)
}
}

\keyword{internal}
