% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_update_source_location}
\alias{mediatailor_update_source_location}
\title{Updates a source location on a specific channel}
\usage{
mediatailor_update_source_location(AccessConfiguration,
  DefaultSegmentDeliveryConfiguration, HttpConfiguration,
  SourceLocationName)
}
\arguments{
\item{AccessConfiguration}{Access configuration parameters. Configures the type of authentication
used to access content from your source location.}

\item{DefaultSegmentDeliveryConfiguration}{The optional configuration for the host server that serves segments.}

\item{HttpConfiguration}{[required] The HTTP configuration for the source location.}

\item{SourceLocationName}{[required] The identifier for the source location you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  AccessConfiguration = list(
    AccessType = "S3_SIGV4"
  ),
  Arn = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  DefaultSegmentDeliveryConfiguration = list(
    BaseUrl = "string"
  ),
  HttpConfiguration = list(
    BaseUrl = "string"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  SourceLocationName = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Updates a source location on a specific channel.
}
\section{Request syntax}{
\preformatted{svc$update_source_location(
  AccessConfiguration = list(
    AccessType = "S3_SIGV4"
  ),
  DefaultSegmentDeliveryConfiguration = list(
    BaseUrl = "string"
  ),
  HttpConfiguration = list(
    BaseUrl = "string"
  ),
  SourceLocationName = "string"
)
}
}

\keyword{internal}
