% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_update_vod_source}
\alias{mediatailor_update_vod_source}
\title{Updates a specific VOD source in a specific source location}
\usage{
mediatailor_update_vod_source(HttpPackageConfigurations,
  SourceLocationName, VodSourceName)
}
\arguments{
\item{HttpPackageConfigurations}{[required] An array of HTTP package configurations for the VOD source on this
account.}

\item{SourceLocationName}{[required] The identifier for the source location you are working on.}

\item{VodSourceName}{[required] The identifier for the VOD source you are working on.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  HttpPackageConfigurations = list(
    list(
      Path = "string",
      SourceGroup = "string",
      Type = "DASH"|"HLS"
    )
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  SourceLocationName = "string",
  Tags = list(
    "string"
  ),
  VodSourceName = "string"
)
}
}
\description{
Updates a specific VOD source in a specific source location.
}
\section{Request syntax}{
\preformatted{svc$update_vod_source(
  HttpPackageConfigurations = list(
    list(
      Path = "string",
      SourceGroup = "string",
      Type = "DASH"|"HLS"
    )
  ),
  SourceLocationName = "string",
  VodSourceName = "string"
)
}
}

\keyword{internal}
