% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_describe_signaling_channel}
\alias{kinesisvideo_describe_signaling_channel}
\title{Returns the most current information about the signaling channel}
\usage{
kinesisvideo_describe_signaling_channel(ChannelName, ChannelARN)
}
\arguments{
\item{ChannelName}{The name of the signaling channel that you want to describe.}

\item{ChannelARN}{The ARN of the signaling channel that you want to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelInfo = list(
    ChannelName = "string",
    ChannelARN = "string",
    ChannelType = "SINGLE_MASTER",
    ChannelStatus = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    SingleMasterConfiguration = list(
      MessageTtlSeconds = 123
    ),
    Version = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the most current information about the signaling channel. You
must specify either the name or the Amazon Resource Name (ARN) of the
channel that you want to describe.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_signaling_channel(
  ChannelName = "string",
  ChannelARN = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
