% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideo_operations.R
\name{kinesisvideo_list_signaling_channels}
\alias{kinesisvideo_list_signaling_channels}
\title{Returns an array of ChannelInfo objects}
\usage{
kinesisvideo_list_signaling_channels(MaxResults, NextToken,
  ChannelNameCondition)
}
\arguments{
\item{MaxResults}{The maximum number of channels to return in the response. The default is
500.}

\item{NextToken}{If you specify this parameter, when the result of a
\code{\link[=kinesisvideo_list_signaling_channels]{list_signaling_channels}}
operation is truncated, the call returns the \code{NextToken} in the
response. To get another batch of channels, provide this token in your
next request.}

\item{ChannelNameCondition}{Optional: Returns only the channels that satisfy a specific condition.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelInfoList = list(
    list(
      ChannelName = "string",
      ChannelARN = "string",
      ChannelType = "SINGLE_MASTER",
      ChannelStatus = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      SingleMasterConfiguration = list(
        MessageTtlSeconds = 123
      ),
      Version = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns an array of \code{ChannelInfo} objects. Each object describes a
signaling channel. To retrieve only those channels that satisfy a
specific condition, you can specify a \code{ChannelNameCondition}.
}
\section{Request syntax}{
\preformatted{svc$list_signaling_channels(
  MaxResults = 123,
  NextToken = "string",
  ChannelNameCondition = list(
    ComparisonOperator = "BEGINS_WITH",
    ComparisonValue = "string"
  )
)
}
}

\keyword{internal}
