% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideoarchivedmedia_service.R
\name{kinesisvideoarchivedmedia}
\alias{kinesisvideoarchivedmedia}
\title{Amazon Kinesis Video Streams Archived Media}
\usage{
kinesisvideoarchivedmedia(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Kinesis Video Streams Archived Media
}
\section{Service syntax}{
\preformatted{svc <- kinesisvideoarchivedmedia(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=kinesisvideoarchivedmedia_get_dash_streaming_session_url]{get_dash_streaming_session_url} \tab Retrieves an MPEG Dynamic Adaptive Streaming over HTTP (DASH) URL for the stream \cr
\link[=kinesisvideoarchivedmedia_get_hls_streaming_session_url]{get_hls_streaming_session_url} \tab Retrieves an HTTP Live Streaming (HLS) URL for the stream \cr
\link[=kinesisvideoarchivedmedia_get_media_for_fragment_list]{get_media_for_fragment_list} \tab Gets media for a list of fragments (specified by fragment number) from the archived data in an Amazon Kinesis video stream\cr
\link[=kinesisvideoarchivedmedia_list_fragments]{list_fragments} \tab Returns a list of Fragment objects from the specified stream and timestamp range within the archived data
}
}

\examples{
\donttest{svc <- kinesisvideoarchivedmedia()
svc$get_dash_streaming_session_url(
  Foo = 123
)}

}
