% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_add_flow_outputs}
\alias{mediaconnect_add_flow_outputs}
\title{Adds outputs to an existing flow}
\usage{
mediaconnect_add_flow_outputs(FlowArn, Outputs)
}
\arguments{
\item{FlowArn}{[required] The flow that you want to add outputs to.}

\item{Outputs}{[required] A list of outputs that you want to add.}
}
\description{
Adds outputs to an existing flow. You can create up to 20 outputs per flow.
}
\section{Request syntax}{
\preformatted{svc$add_flow_outputs(
  FlowArn = "string",
  Outputs = list(
    list(
      Description = "string",
      Destination = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        ConstantInitializationVector = "string",
        DeviceId = "string",
        KeyType = "speke"|"static-key",
        Region = "string",
        ResourceId = "string",
        RoleArn = "string",
        SecretArn = "string",
        Url = "string"
      ),
      MaxLatency = 123,
      Name = "string",
      Port = 123,
      Protocol = "zixi-push"|"rtp-fec"|"rtp",
      SmoothingLatency = 123,
      StreamId = "string"
    )
  )
)
}
}

\keyword{internal}
