% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_describe_offering}
\alias{mediaconnect_describe_offering}
\title{Displays the details of an offering}
\usage{
mediaconnect_describe_offering(OfferingArn)
}
\arguments{
\item{OfferingArn}{[required] The Amazon Resource Name (ARN) of the offering.}
}
\description{
Displays the details of an offering. The response includes the offering description, duration, outbound bandwidth, price, and Amazon Resource Name (ARN).
}
\section{Request syntax}{
\preformatted{svc$describe_offering(
  OfferingArn = "string"
)
}
}

\keyword{internal}
