% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_describe_reservation}
\alias{mediaconnect_describe_reservation}
\title{Displays the details of a reservation}
\usage{
mediaconnect_describe_reservation(ReservationArn)
}
\arguments{
\item{ReservationArn}{[required] The Amazon Resource Name (ARN) of the reservation.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Reservation = list(
    CurrencyCode = "string",
    Duration = 123,
    DurationUnits = "MONTHS",
    End = "string",
    OfferingArn = "string",
    OfferingDescription = "string",
    PricePerUnit = "string",
    PriceUnits = "HOURLY",
    ReservationArn = "string",
    ReservationName = "string",
    ReservationState = "ACTIVE"|"EXPIRED"|"PROCESSING"|"CANCELED",
    ResourceSpecification = list(
      ReservedBitrate = 123,
      ResourceType = "Mbps_Outbound_Bandwidth"
    ),
    Start = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Displays the details of a reservation. The response includes the
reservation name, state, start date and time, and the details of the
offering that make up the rest of the reservation (such as price,
duration, and outbound bandwidth).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_reservation(
  ReservationArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
