% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_stop_flow}
\alias{mediaconnect_stop_flow}
\title{Stops a flow}
\usage{
mediaconnect_stop_flow(FlowArn)
}
\arguments{
\item{FlowArn}{[required] The ARN of the flow that you want to stop.}
}
\description{
Stops a flow.
}
\section{Request syntax}{
\preformatted{svc$stop_flow(
  FlowArn = "string"
)
}
}

\keyword{internal}
