% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_update_flow_output}
\alias{mediaconnect_update_flow_output}
\title{Updates an existing flow output}
\usage{
mediaconnect_update_flow_output(Description, Destination, Encryption,
  FlowArn, MaxLatency, OutputArn, Port, Protocol, SmoothingLatency,
  StreamId)
}
\arguments{
\item{Description}{A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.}

\item{Destination}{The IP address where you want to send the output.}

\item{Encryption}{The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).}

\item{FlowArn}{[required] The flow that is associated with the output that you want to update.}

\item{MaxLatency}{The maximum latency in milliseconds for Zixi-based streams.}

\item{OutputArn}{[required] The ARN of the output that you want to update.}

\item{Port}{The port to use when content is distributed to this output.}

\item{Protocol}{The protocol to use for the output.}

\item{SmoothingLatency}{The smoothing latency in milliseconds for RTP and RTP-FEC streams.}

\item{StreamId}{The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.}
}
\description{
Updates an existing flow output.
}
\section{Request syntax}{
\preformatted{svc$update_flow_output(
  Description = "string",
  Destination = "string",
  Encryption = list(
    Algorithm = "aes128"|"aes192"|"aes256",
    KeyType = "static-key",
    RoleArn = "string",
    SecretArn = "string"
  ),
  FlowArn = "string",
  MaxLatency = 123,
  OutputArn = "string",
  Port = 123,
  Protocol = "zixi-push"|"rtp-fec"|"rtp",
  SmoothingLatency = 123,
  StreamId = "string"
)
}
}

\keyword{internal}
