% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_delete_multiplex}
\alias{medialive_delete_multiplex}
\title{Delete a multiplex}
\usage{
medialive_delete_multiplex(MultiplexId)
}
\arguments{
\item{MultiplexId}{[required] The ID of the multiplex.}
}
\description{
Delete a multiplex. The multiplex must be idle.
}
\section{Request syntax}{
\preformatted{svc$delete_multiplex(
  MultiplexId = "string"
)
}
}

\keyword{internal}
