% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_input}
\alias{medialive_describe_input}
\title{Produces details about an input}
\usage{
medialive_describe_input(InputId)
}
\arguments{
\item{InputId}{[required] Unique ID of the input}
}
\description{
Produces details about an input
}
\section{Request syntax}{
\preformatted{svc$describe_input(
  InputId = "string"
)
}
}

\keyword{internal}
