% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_multiplex}
\alias{medialive_describe_multiplex}
\title{Gets details about a multiplex}
\usage{
medialive_describe_multiplex(MultiplexId)
}
\arguments{
\item{MultiplexId}{[required] The ID of the multiplex.}
}
\description{
Gets details about a multiplex.
}
\section{Request syntax}{
\preformatted{svc$describe_multiplex(
  MultiplexId = "string"
)
}
}

\keyword{internal}
