% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_multiplex_program}
\alias{medialive_describe_multiplex_program}
\title{Get the details for a program in a multiplex}
\usage{
medialive_describe_multiplex_program(MultiplexId, ProgramName)
}
\arguments{
\item{MultiplexId}{[required] The ID of the multiplex that the program belongs to.}

\item{ProgramName}{[required] The name of the program.}
}
\description{
Get the details for a program in a multiplex.
}
\section{Request syntax}{
\preformatted{svc$describe_multiplex_program(
  MultiplexId = "string",
  ProgramName = "string"
)
}
}

\keyword{internal}
