% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_input_device_transfers}
\alias{medialive_list_input_device_transfers}
\title{List input devices that are currently being transferred}
\usage{
medialive_list_input_device_transfers(MaxResults, NextToken,
  TransferType)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}

\item{TransferType}{[required]}
}
\description{
List input devices that are currently being transferred. List input devices that you are transferring from your AWS account or input devices that another AWS account is transferring to you.
}
\section{Request syntax}{
\preformatted{svc$list_input_device_transfers(
  MaxResults = 123,
  NextToken = "string",
  TransferType = "string"
)
}
}

\keyword{internal}
