% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_input_security_groups}
\alias{medialive_list_input_security_groups}
\title{Produces a list of Input Security Groups for an account}
\usage{
medialive_list_input_security_groups(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InputSecurityGroups = list(
    list(
      Arn = "string",
      Id = "string",
      Inputs = list(
        "string"
      ),
      State = "IDLE"|"IN_USE"|"UPDATING"|"DELETED",
      Tags = list(
        "string"
      ),
      WhitelistRules = list(
        list(
          Cidr = "string"
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Produces a list of Input Security Groups for an account
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_input_security_groups(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
