% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_multiplexes}
\alias{medialive_list_multiplexes}
\title{Retrieve a list of the existing multiplexes}
\usage{
medialive_list_multiplexes(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items to return.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Multiplexes = list(
    list(
      Arn = "string",
      AvailabilityZones = list(
        "string"
      ),
      Id = "string",
      MultiplexSettings = list(
        TransportStreamBitrate = 123
      ),
      Name = "string",
      PipelinesRunningCount = 123,
      ProgramCount = 123,
      State = "CREATING"|"CREATE_FAILED"|"IDLE"|"STARTING"|"RUNNING"|"RECOVERING"|"STOPPING"|"DELETING"|"DELETED",
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieve a list of the existing multiplexes.
}
\section{Request syntax}{
\preformatted{svc$list_multiplexes(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
