% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_multiplexes}
\alias{medialive_list_multiplexes}
\title{Retrieve a list of the existing multiplexes}
\usage{
medialive_list_multiplexes(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items to return.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\description{
Retrieve a list of the existing multiplexes.
}
\section{Request syntax}{
\preformatted{svc$list_multiplexes(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
