% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_input_device}
\alias{medialive_update_input_device}
\title{Updates the parameters for the input device}
\usage{
medialive_update_input_device(HdDeviceSettings, InputDeviceId, Name)
}
\arguments{
\item{HdDeviceSettings}{The settings that you want to apply to the input device.}

\item{InputDeviceId}{[required] The unique ID of the input device. For example, hd-123456789abcdef.}

\item{Name}{The name that you assigned to this input device (not the unique ID).}
}
\description{
Updates the parameters for the input device.
}
\section{Request syntax}{
\preformatted{svc$update_input_device(
  HdDeviceSettings = list(
    ConfiguredInput = "AUTO"|"HDMI"|"SDI",
    MaxBitrate = 123
  ),
  InputDeviceId = "string",
  Name = "string"
)
}
}

\keyword{internal}
