% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_update_multiplex_program}
\alias{medialive_update_multiplex_program}
\title{Update a program in a multiplex}
\usage{
medialive_update_multiplex_program(MultiplexId,
  MultiplexProgramSettings, ProgramName)
}
\arguments{
\item{MultiplexId}{[required] The ID of the multiplex of the program to update.}

\item{MultiplexProgramSettings}{The new settings for a multiplex program.}

\item{ProgramName}{[required] The name of the program to update.}
}
\description{
Update a program in a multiplex.
}
\section{Request syntax}{
\preformatted{svc$update_multiplex_program(
  MultiplexId = "string",
  MultiplexProgramSettings = list(
    ProgramNumber = 123,
    ServiceDescriptor = list(
      ProviderName = "string",
      ServiceName = "string"
    ),
    VideoSettings = list(
      ConstantBitrate = 123,
      StatmuxSettings = list(
        MaximumBitrate = 123,
        MinimumBitrate = 123
      )
    )
  ),
  ProgramName = "string"
)
}
}

\keyword{internal}
