% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_service.R
\name{mediapackage}
\alias{mediapackage}
\title{AWS Elemental MediaPackage}
\usage{
mediapackage(config = NULL)
}
\arguments{
\item{config}{An optional list of custom configurations for the service. Currently
supports adding custom credentials, endpoint, and region.}
}
\description{
AWS Elemental MediaPackage
}
\section{Service syntax}{
\preformatted{svc <- mediapackage(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string",
        provider_name = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=mediapackage_create_channel]{create_channel} \tab Creates a new Channel \cr
\link[=mediapackage_create_origin_endpoint]{create_origin_endpoint} \tab Creates a new OriginEndpoint record \cr
\link[=mediapackage_delete_channel]{delete_channel} \tab Deletes an existing Channel \cr
\link[=mediapackage_delete_origin_endpoint]{delete_origin_endpoint} \tab Deletes an existing OriginEndpoint \cr
\link[=mediapackage_describe_channel]{describe_channel} \tab Gets details about a Channel \cr
\link[=mediapackage_describe_origin_endpoint]{describe_origin_endpoint} \tab Gets details about an existing OriginEndpoint \cr
\link[=mediapackage_list_channels]{list_channels} \tab Returns a collection of Channels \cr
\link[=mediapackage_list_origin_endpoints]{list_origin_endpoints} \tab Returns a collection of OriginEndpoint records \cr
\link[=mediapackage_list_tags_for_resource]{list_tags_for_resource} \tab List tags for resource \cr
\link[=mediapackage_rotate_channel_credentials]{rotate_channel_credentials} \tab Changes the Channel's first IngestEndpoint's username and password \cr
\link[=mediapackage_rotate_ingest_endpoint_credentials]{rotate_ingest_endpoint_credentials} \tab Rotate the IngestEndpoint's username and password, as specified by the IngestEndpoint's id\cr
\link[=mediapackage_tag_resource]{tag_resource} \tab Tag resource \cr
\link[=mediapackage_untag_resource]{untag_resource} \tab Untag resource \cr
\link[=mediapackage_update_channel]{update_channel} \tab Updates an existing Channel \cr
\link[=mediapackage_update_origin_endpoint]{update_origin_endpoint} \tab Updates an existing OriginEndpoint
}
}

\examples{
\donttest{svc <- mediapackage()
svc$create_channel(
  Foo = 123
)}

}
