% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_configure_logs}
\alias{mediapackage_configure_logs}
\title{Changes the Channel's properities to configure log subscription}
\usage{
mediapackage_configure_logs(EgressAccessLogs, Id, IngressAccessLogs)
}
\arguments{
\item{EgressAccessLogs}{}

\item{Id}{[required] The ID of the channel to log subscription.}

\item{IngressAccessLogs}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Description = "string",
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  HlsIngest = list(
    IngestEndpoints = list(
      list(
        Id = "string",
        Password = "string",
        Url = "string",
        Username = "string"
      )
    )
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Changes the Channel's properities to configure log subscription
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$configure_logs(
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
