% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_configure_logs}
\alias{mediapackage_configure_logs}
\title{Changes the Channel's properities to configure log subscription}
\usage{
mediapackage_configure_logs(EgressAccessLogs, Id, IngressAccessLogs)
}
\arguments{
\item{EgressAccessLogs}{}

\item{Id}{[required] The ID of the channel to log subscription.}

\item{IngressAccessLogs}{}
}
\description{
Changes the Channel's properities to configure log subscription
}
\section{Request syntax}{
\preformatted{svc$configure_logs(
  EgressAccessLogs = list(
    LogGroupName = "string"
  ),
  Id = "string",
  IngressAccessLogs = list(
    LogGroupName = "string"
  )
)
}
}

\keyword{internal}
