% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_list_channels}
\alias{mediapackage_list_channels}
\title{Returns a collection of Channels}
\usage{
mediapackage_list_channels(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}
}
\value{
A list with the following syntax:\preformatted{list(
  Channels = list(
    list(
      Arn = "string",
      Description = "string",
      EgressAccessLogs = list(
        LogGroupName = "string"
      ),
      HlsIngest = list(
        IngestEndpoints = list(
          list(
            Id = "string",
            Password = "string",
            Url = "string",
            Username = "string"
          )
        )
      ),
      Id = "string",
      IngressAccessLogs = list(
        LogGroupName = "string"
      ),
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a collection of Channels.
}
\section{Request syntax}{
\preformatted{svc$list_channels(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
