% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_tag_resource}
\alias{mediapackage_tag_resource}
\title{Tag resource}
\usage{
mediapackage_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required]}

\item{Tags}{[required]}
}
\description{

}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
