% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_describe_container}
\alias{mediastore_describe_container}
\title{Retrieves the properties of the requested container}
\usage{
mediastore_describe_container(ContainerName)
}
\arguments{
\item{ContainerName}{The name of the container to query.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Container = list(
    Endpoint = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    ARN = "string",
    Name = "string",
    Status = "ACTIVE"|"CREATING"|"DELETING",
    AccessLoggingEnabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the properties of the requested container. This request is
commonly used to retrieve the endpoint of a container. An endpoint is a
value assigned by the service when a new container is created. A
container's endpoint does not change after it has been assigned. The
\code{\link[=mediastore_describe_container]{describe_container}} request returns a
single \code{Container} object based on \code{ContainerName}. To return all
\code{Container} objects that are associated with a specified AWS account,
use \code{\link[=mediastore_list_containers]{list_containers}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_container(
  ContainerName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
