% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastore_operations.R
\name{mediastore_get_container_policy}
\alias{mediastore_get_container_policy}
\title{Retrieves the access policy for the specified container}
\usage{
mediastore_get_container_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container.}
}
\description{
Retrieves the access policy for the specified container. For information
about the data that is included in an access policy, see the \href{https://aws.amazon.com/documentation/iam/}{AWS Identity and Access Management User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_container_policy(
  ContainerName = "string"
)
}
}

\keyword{internal}
