% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediastoredata_operations.R
\name{mediastoredata_delete_object}
\alias{mediastoredata_delete_object}
\title{Deletes an object at the specified path}
\usage{
mediastoredata_delete_object(Path)
}
\arguments{
\item{Path}{[required] The path (including the file name) where the object is stored in the
container. Format: <folder name>/<folder name>/<file
name>}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an object at the specified path.
}
\section{Request syntax}{
\preformatted{svc$delete_object(
  Path = "string"
)
}
}

\keyword{internal}
