% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_delete_playback_configuration}
\alias{mediatailor_delete_playback_configuration}
\title{Deletes the playback configuration for the specified name}
\usage{
mediatailor_delete_playback_configuration(Name)
}
\arguments{
\item{Name}{[required] The identifier for the playback configuration.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the playback configuration for the specified name.
}
\section{Request syntax}{
\preformatted{svc$delete_playback_configuration(
  Name = "string"
)
}
}

\keyword{internal}
