% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_operations.R
\name{mediatailor_put_playback_configuration}
\alias{mediatailor_put_playback_configuration}
\title{Adds a new playback configuration to AWS Elemental MediaTailor}
\usage{
mediatailor_put_playback_configuration(AdDecisionServerUrl,
  AvailSuppression, Bumper, CdnConfiguration, ConfigurationAliases,
  DashConfiguration, LivePreRollConfiguration, ManifestProcessingRules,
  Name, PersonalizationThresholdSeconds, SlateAdUrl, Tags,
  TranscodeProfileName, VideoContentSourceUrl)
}
\arguments{
\item{AdDecisionServerUrl}{The URL for the ad decision server (ADS). This includes the
specification of static parameters and placeholders for dynamic
parameters. AWS Elemental MediaTailor substitutes player-specific and
session-specific parameters as needed when calling the ADS. Alternately,
for testing you can provide a static VAST URL. The maximum length is
25,000 characters.}

\item{AvailSuppression}{The configuration for avail suppression, also known as ad suppression.
For more information about ad suppression, see \href{https://docs.aws.amazon.com/mediatailor/latest/ug/ad-behavior.html}{Ad Suppression}.}

\item{Bumper}{The configuration for bumpers. Bumpers are short audio or video clips
that play at the start or before the end of an ad break. To learn more
about bumpers, see
\href{https://docs.aws.amazon.com/mediatailor/latest/ug/bumpers.html}{Bumpers}.}

\item{CdnConfiguration}{The configuration for using a content delivery network (CDN), like
Amazon CloudFront, for content and ad segment management.}

\item{ConfigurationAliases}{The player parameters and aliases used as dynamic variables during
session initialization. For more information, see \href{https://docs.aws.amazon.com/mediatailor/latest/ug/}{Domain Variables}.}

\item{DashConfiguration}{The configuration for DASH content.}

\item{LivePreRollConfiguration}{The configuration for pre-roll ad insertion.}

\item{ManifestProcessingRules}{The configuration for manifest processing rules. Manifest processing
rules enable customization of the personalized manifests created by
MediaTailor.}

\item{Name}{The identifier for the playback configuration.}

\item{PersonalizationThresholdSeconds}{Defines the maximum duration of underfilled ad time (in seconds) allowed
in an ad break. If the duration of underfilled ad time exceeds the
personalization threshold, then the personalization of the ad break is
abandoned and the underlying content is shown. This feature applies to
\emph{ad replacement} in live and VOD streams, rather than ad insertion,
because it relies on an underlying content stream. For more information
about ad break behavior, including ad replacement and insertion, see \href{https://docs.aws.amazon.com/mediatailor/latest/ug/ad-behavior.html}{Ad Behavior in AWS Elemental MediaTailor}.}

\item{SlateAdUrl}{The URL for a high-quality video asset to transcode and use to fill in
time that's not used by ads. AWS Elemental MediaTailor shows the slate
to fill in gaps in media content. Configuring the slate is optional for
non-VPAID configurations. For VPAID, the slate is required because
MediaTailor provides it in the slots that are designated for dynamic ad
content. The slate must be a high-quality asset that contains both audio
and video.}

\item{Tags}{The tags to assign to the playback configuration.}

\item{TranscodeProfileName}{The name that is used to associate this playback configuration with a
custom transcode profile. This overrides the dynamic transcoding
defaults of MediaTailor. Use this only if you have already set up custom
profiles with the help of AWS Support.}

\item{VideoContentSourceUrl}{The URL prefix for the parent manifest for the stream, minus the asset
ID. The maximum length is 512 characters.}
}
\value{
A list with the following syntax:\preformatted{list(
  AdDecisionServerUrl = "string",
  AvailSuppression = list(
    Mode = "OFF"|"BEHIND_LIVE_EDGE",
    Value = "string"
  ),
  Bumper = list(
    EndUrl = "string",
    StartUrl = "string"
  ),
  CdnConfiguration = list(
    AdSegmentUrlPrefix = "string",
    ContentSegmentUrlPrefix = "string"
  ),
  ConfigurationAliases = list(
    list(
      "string"
    )
  ),
  DashConfiguration = list(
    ManifestEndpointPrefix = "string",
    MpdLocation = "string",
    OriginManifestType = "SINGLE_PERIOD"|"MULTI_PERIOD"
  ),
  HlsConfiguration = list(
    ManifestEndpointPrefix = "string"
  ),
  LivePreRollConfiguration = list(
    AdDecisionServerUrl = "string",
    MaxDurationSeconds = 123
  ),
  ManifestProcessingRules = list(
    AdMarkerPassthrough = list(
      Enabled = TRUE|FALSE
    )
  ),
  Name = "string",
  PersonalizationThresholdSeconds = 123,
  PlaybackConfigurationArn = "string",
  PlaybackEndpointPrefix = "string",
  SessionInitializationEndpointPrefix = "string",
  SlateAdUrl = "string",
  Tags = list(
    "string"
  ),
  TranscodeProfileName = "string",
  VideoContentSourceUrl = "string"
)
}
}
\description{
Adds a new playback configuration to AWS Elemental MediaTailor.
}
\section{Request syntax}{
\preformatted{svc$put_playback_configuration(
  AdDecisionServerUrl = "string",
  AvailSuppression = list(
    Mode = "OFF"|"BEHIND_LIVE_EDGE",
    Value = "string"
  ),
  Bumper = list(
    EndUrl = "string",
    StartUrl = "string"
  ),
  CdnConfiguration = list(
    AdSegmentUrlPrefix = "string",
    ContentSegmentUrlPrefix = "string"
  ),
  ConfigurationAliases = list(
    list(
      "string"
    )
  ),
  DashConfiguration = list(
    MpdLocation = "string",
    OriginManifestType = "SINGLE_PERIOD"|"MULTI_PERIOD"
  ),
  LivePreRollConfiguration = list(
    AdDecisionServerUrl = "string",
    MaxDurationSeconds = 123
  ),
  ManifestProcessingRules = list(
    AdMarkerPassthrough = list(
      Enabled = TRUE|FALSE
    )
  ),
  Name = "string",
  PersonalizationThresholdSeconds = 123,
  SlateAdUrl = "string",
  Tags = list(
    "string"
  ),
  TranscodeProfileName = "string",
  VideoContentSourceUrl = "string"
)
}
}

\keyword{internal}
